### 2024-07-03 f1dc743
Fix examples.jl after Case move
