+++ test/examples/examples.jl
@@ -90,7 +90,7 @@
         @trixi_testset "fluid/dam_break_oil_film_2d.jl" begin
             @test_nowarn_mod trixi_include(@__MODULE__,
                                            joinpath(examples_dir(), "fluid",
-                                                    "dam_break_two_fluids_2d.jl"),
+                                                    "dam_break_oil_film_2d.jl"),
                                            tspan=(0.0, 0.1)) [
                 r"┌ Info: The desired tank length in y-direction .*\n",
                 r"└ New tank length in y-direction.*\n",
