### 2024-07-03 #6961 @jbytecode
@danielskatz - Thank you for inviting me. I have been investigating the repository and the manuscript, indeed it seems to be good submission. I am always happy editing Julia submissions, however, the subject is a little bit out of my experiences. It would be better to invite someone works on Physics, I think. By the way, if you fail to find someone available and you believe I can edit this, please ping me again, I will try to help and do my best anyway.
