+++ src/schemes/boundary/open_boundary/system.jl
@@ -33,30 +33,31 @@ about the method see [description below](@ref method_of_characteristics).
 !!! warning "Experimental Implementation"
 	This is an experimental feature and may change in any future releases.
 """
-struct OpenBoundarySPHSystem{BZ, NDIMS, ELTYPE <: Real, IC, FS, ARRAY1D, ARRAY2D, RV, RP,
-                             RD, B} <: System{NDIMS, IC}
-    initial_condition        :: IC
-    fluid_system             :: FS
-    mass                     :: ARRAY1D # Array{ELTYPE, 1}: [particle]
-    density                  :: ARRAY1D # Array{ELTYPE, 1}: [particle]
-    volume                   :: ARRAY1D # Array{ELTYPE, 1}: [particle]
-    pressure                 :: ARRAY1D # Array{ELTYPE, 1}: [particle]
-    characteristics          :: ARRAY2D # Array{ELTYPE, 2}: [characteristic, particle]
-    previous_characteristics :: ARRAY2D # Array{ELTYPE, 2}: [characteristic, particle]
-    sound_speed              :: ELTYPE
-    boundary_zone            :: BZ
-    flow_direction           :: SVector{NDIMS, ELTYPE}
-    reference_velocity       :: RV
-    reference_pressure       :: RP
-    reference_density        :: RD
-    buffer                   :: B
-    update_callback_used     :: Ref{Bool}
+struct OpenBoundarySPHSystem{BM, BZ, NDIMS, ELTYPE <: Real, IC, FS, ARRAY1D, RV,
+                             RP, RD, B, C} <: System{NDIMS, IC}
+    initial_condition    :: IC
+    fluid_system         :: FS
+    boundary_model       :: BM
+    mass                 :: ARRAY1D # Array{ELTYPE, 1}: [particle]
+    density              :: ARRAY1D # Array{ELTYPE, 1}: [particle]
+    volume               :: ARRAY1D # Array{ELTYPE, 1}: [particle]
+    pressure             :: ARRAY1D # Array{ELTYPE, 1}: [particle]
+    sound_speed          :: ELTYPE
+    boundary_zone        :: BZ
+    flow_direction       :: SVector{NDIMS, ELTYPE}
+    reference_velocity   :: RV
+    reference_pressure   :: RP
+    reference_density    :: RD
+    buffer               :: B
+    update_callback_used :: Ref{Bool}
+    cache                :: C
 
-    function OpenBoundarySPHSystem(boundary_zone::Union{InFlow, OutFlow}; sound_speed,
-                                   fluid_system::FluidSystem, buffer_size::Integer,
-                                   reference_velocity=zeros(ndims(boundary_zone)),
-                                   reference_pressure=0.0,
-                                   reference_density=first(boundary_zone.initial_condition.density))
+    function OpenBoundarySPHSystem(boundary_zone::Union{InFlow, OutFlow};
+                                   sound_speed, fluid_system::FluidSystem,
+                                   buffer_size::Integer, boundary_model,
+                                   reference_velocity=nothing,
+                                   reference_pressure=nothing,
+                                   reference_density=nothing)
         (; initial_condition) = boundary_zone
 
         buffer = SystemBuffer(nparticles(initial_condition), buffer_size)
@@ -66,7 +67,7 @@ struct OpenBoundarySPHSystem{BZ, NDIMS, ELTYPE <: Real, IC, FS, ARRAY1D, ARRAY2D
         NDIMS = ndims(initial_condition)
         ELTYPE = eltype(initial_condition)
 
-        if !(reference_velocity isa Function ||
+        if !(reference_velocity isa Function || isnothing(reference_velocity) ||
[…]
