+++ examples/fluid/pipe_flow_2d.jl
@@ -88,6 +88,7 @@ inflow = InFlow(; plane=([0.0, 0.0], [0.0, domain_size[2]]), flow_direction,
                 open_boundary_layers, density=fluid_density, particle_spacing)
 
 open_boundary_in = OpenBoundarySPHSystem(inflow; sound_speed, fluid_system,
+                                         boundary_model=BoundaryModelLastiwka(),
                                          buffer_size=n_buffer_particles,
                                          reference_pressure=pressure,
                                          reference_velocity=velocity_function)
@@ -97,6 +98,7 @@ outflow = OutFlow(; plane=([domain_size[1], 0.0], [domain_size[1], domain_size[2
                   particle_spacing)
 
 open_boundary_out = OpenBoundarySPHSystem(outflow; sound_speed, fluid_system,
+                                          boundary_model=BoundaryModelLastiwka(),
                                           buffer_size=n_buffer_particles,
                                           reference_pressure=pressure,
                                           reference_velocity=velocity_function)
