+++ src/schemes/boundary/open_boundary/tafuni.jl
@@ -1,0 +1,171 @@
+# Two ways of providing the information to an open boundary are considered:
+# - physical quantities are either assigned a priori
+# - or extrapolated from the fluid domain to the buffer zones (inflow and outflow) using ghost nodes
+
+# The position of the ghost nodes is obtained by mirroring the boundary particles
+# into the fluid along a direction that is normal to the open boundary.
+
+# In order to calculate fluid quantities at the ghost nodes, a standard particle interpolation
+# would not be consistent due to the proximity of these points to an open boundary,
+# which translates into the lack of a full kernel support.
+struct BoundaryModelTafuni end
+
+function update_quantities!(system, ::BoundaryModelTafuni, v, u, v_ode, u_ode, semi, t)
+    @trixi_timeit timer() "interpolate and correct values" interpolate_values!(system,
+                                                                               v_ode, u_ode,
+                                                                               semi, t;
+                                                                               system.cache...)
+end
+
+update_final!(system, ::BoundaryModelTafuni, v, u, v_ode, u_ode, semi, t) = system
+
+function interpolate_values!(system, v_ode, u_ode, semi, t;
+                             prescribed_pressure=false, prescribed_velocity=false)
+    (; pressure, density, fluid_system, boundary_zone,
+    reference_velocity, reference_pressure) = system
+
+    svector_(f, system) = SVector(ntuple(@inline(dim->f[dim + 1]), ndims(system)))
+
+    zero_svector(system) = SVector(ntuple(@inline(dim->zero(eltype(system))),
+                                          ndims(system) + 1))
+    zero_smatrix(system) = SMatrix{ndims(system) + 1,
+                                   ndims(system) + 1}(ntuple(@inline(i->zero(eltype(system))),
+                                                             Val((ndims(system) + 1)^2)))
+
+    v_open_boundary = wrap_v(v_ode, system, semi)
+    v_fluid = wrap_v(v_ode, fluid_system, semi)
+    u_open_boundary = wrap_u(u_ode, system, semi)
+    u_fluid = wrap_u(u_ode, fluid_system, semi)
+
+    neighborhood_search = get_neighborhood_search(system, fluid_system, semi)
+
+    interpolated_velocity = [zero_svector(system) for _ in 1:ndims(system)]
+
+    for particle in each_moving_particle(system)
+        particle_coords = current_coords(u_open_boundary, system, particle)
+        ghost_node_position = mirror_position(particle_coords, boundary_zone)
+
+        # Set zero
+        correction_matrix = zero_smatrix(system)
+        interpolated_density = zero_svector(system)
+        interpolated_pressure = zero_svector(system)
+        @inbounds for dim in 1:ndims(system)
+            interpolated_velocity[dim] = zero_svector(system)
+        end
+
+        for neighbor in PointNeighbors.eachneighbor(ghost_node_position,
+                                                    neighborhood_search)
+            neighbor_coords = current_coords(u_fluid, fluid_system, neighbor)
+            pos_diff = ghost_node_position - neighbor_coords
+            distance2 = dot(pos_diff, pos_diff)
+
+            if distance2 <= neighborhood_search.search_radius^2
+                distance = sqrt(distance2)
+
[…]
