### 2024-07-04 25140d7
extrapolate rev-values for Lastiwka model
