+++ src/schemes/boundary/open_boundary/lastiwka.jl
@@ -1,9 +1,23 @@
-struct BoundaryModelLastiwka end
+struct BoundaryModelLastiwka
+    extra_polate_reference_values::Bool
+    function BoundaryModelLastiwka(; extra_polate_reference_values::Bool=true)
+        return new{}(extra_polate_reference_values)
+    end
+end
 
-@inline function update_quantities!(system, ::BoundaryModelLastiwka,
+@inline function update_quantities!(system, boundary_model::BoundaryModelLastiwka,
                                     v, u, v_ode, u_ode, semi, t)
     (; density, pressure, cache, flow_direction, sound_speed,
     reference_velocity, reference_pressure, reference_density) = system
+
+    if boundary_model.extra_polate_reference_values
+        (; prescribed_pressure, prescribed_velocity) = cache
+
+        @trixi_timeit timer() "interpolate and correct values" begin
+            interpolate_values!(system, v_ode, u_ode, semi, t; prescribed_pressure,
+                                prescribed_velocity)
+        end
+    end
 
     # Update quantities based on the characteristic variables
     @threaded system for particle in each_moving_particle(system)
@@ -13,13 +27,16 @@ struct BoundaryModelLastiwka end
         J2 = cache.characteristics[2, particle]
         J3 = cache.characteristics[3, particle]
 
-        rho_ref = reference_density(particle_position, t)
+        rho_ref = reference_value(reference_density, density, system, particle,
+                                  particle_position, t)
         density[particle] = rho_ref + ((-J1 + 0.5 * (J2 + J3)) / sound_speed^2)
 
-        p_ref = reference_pressure(particle_position, t)
+        p_ref = reference_value(reference_pressure, pressure, system, particle,
+                                particle_position, t)
         pressure[particle] = p_ref + 0.5 * (J2 + J3)
 
-        v_ref = reference_velocity(particle_position, t)
+        v_ref = reference_value(reference_velocity, v, system, particle,
+                                particle_position, t)
         rho = density[particle]
         v_ = v_ref + ((J2 - J3) / (2 * sound_speed * rho)) * flow_direction
 
@@ -100,7 +117,7 @@ end
 
 function evaluate_characteristics!(system, neighbor_system::FluidSystem,
                                    v, u, v_ode, u_ode, semi, t)
-    (; volume, sound_speed, cache, flow_direction,
+    (; volume, sound_speed, cache, flow_direction, density, pressure,
     reference_velocity, reference_pressure, reference_density) = system
     (; characteristics) = cache
 
@@ -119,13 +136,16 @@ function evaluate_characteristics!(system, neighbor_system::FluidSystem,
 
         # Determine current and prescribed quantities
         rho_b = particle_density(v_neighbor_system, neighbor_system, neighbor)
-        rho_ref = reference_density(neighbor_position, t)
+        rho_ref = reference_value(reference_density, density, system, particle,
+                                  neighbor_position, t)
 
         p_b = particle_pressure(v_neighbor_system, neighbor_system, neighbor)
-        p_ref = reference_pressure(neighbor_position, t)
+        p_ref = reference_value(reference_pressure, pressure, system, particle,
+                                neighbor_position, t)
 
         v_b = current_velocity(v_neighbor_system, neighbor_system, neighbor)
-        v_neighbor_ref = reference_velocity(neighbor_position, t)
+        v_neighbor_ref = reference_value(reference_velocity, v, system, particle,
+                                         neighbor_position, t)
 
         # Determine characteristic variables
         density_term = -sound_speed^2 * (rho_b - rho_ref)
