### 2024-07-04 d2a5fca
density with inverse state equation
