+++ examples/fluid/pipe_flow_2d.jl
@@ -8,13 +8,13 @@ using OrdinaryDiffEq
 particle_spacing = 0.02
 
 # Make sure that the kernel support of fluid particles at a boundary is always fully sampled
-boundary_layers = 3
+boundary_layers = 5
 
 # Make sure that the kernel support of fluid particles at an open boundary is always
 # fully sampled.
 # Note: Due to the dynamics at the inlets and outlets of open boundaries,
 # it is recommended to use `open_boundary_layers > boundary_layers`
-open_boundary_layers = 6
+open_boundary_layers = 8
 
 # ==========================================================================================
 # ==== Experiment Setup
@@ -34,9 +34,9 @@ fluid_density = 1000.0
 
 # For this particular example, it is necessary to have a background pressure.
 # Otherwise the suction at the outflow is to big and the simulation becomes unstable.
-pressure = 1000.0
+pressure = 0.0
 
-sound_speed = 10 * prescribed_velocity
+sound_speed = 20 * prescribed_velocity
 
 state_equation = nothing
 
@@ -66,6 +66,8 @@ fluid_system = EntropicallyDampedSPHSystem(pipe.fluid, smoothing_kernel, smoothi
                                            buffer_size=n_buffer_particles)
 
 # Alternatively the WCSPH scheme can be used
+# state_equation = StateEquationCole(; sound_speed, reference_density=fluid_density,
+#                                    exponent=1, background_pressure=pressure)
 # alpha = 8 * kinematic_viscosity / (smoothing_length * sound_speed)
 # viscosity = ArtificialViscosityMonaghan(; alpha, beta=0.0)
 
@@ -87,10 +89,8 @@ inflow = InFlow(; plane=([0.0, 0.0], [0.0, domain_size[2]]), flow_direction,
                 open_boundary_layers, density=fluid_density, particle_spacing)
 
 open_boundary_in = OpenBoundarySPHSystem(inflow; sound_speed, fluid_system,
-                                         boundary_model=BoundaryModelLastiwka(),
+                                         boundary_model=BoundaryModelTafuni(),
                                          buffer_size=n_buffer_particles,
-                                         reference_pressure=pressure,
-                                         #   reference_density=fluid_density,
                                          reference_velocity=velocity_function)
 
 outflow = OutFlow(; plane=([domain_size[1], 0.0], [domain_size[1], domain_size[2]]),
@@ -98,9 +98,8 @@ outflow = OutFlow(; plane=([domain_size[1], 0.0], [domain_size[1], domain_size[2
                   particle_spacing)
 
 open_boundary_out = OpenBoundarySPHSystem(outflow; sound_speed, fluid_system,
-                                          boundary_model=BoundaryModelLastiwka(),
+                                          boundary_model=BoundaryModelTafuni(),
                                           buffer_size=n_buffer_particles,
-                                          # reference_density=fluid_density,
                                           # reference_velocity=velocity_function,
                                           reference_pressure=pressure)
 
@@ -110,7 +109,7 @@ open_boundary_out = OpenBoundarySPHSystem(outflow; sound_speed, fluid_system,
 boundary_model = BoundaryModelDummyParticles(pipe.boundary.density, pipe.boundary.mass,
                                              AdamiPressureExtrapolation(),
                                              state_equation=state_equation,
-                                             #viscosity=ViscosityAdami(nu=1e-4),
+                                             viscosity=ViscosityAdami(nu=1e-4),
                                              smoothing_kernel, smoothing_length)
 
 boundary_system = BoundarySPHSystem(pipe.boundary, boundary_model)
