+++ src/schemes/fluid/fluid.jl
@@ -39,6 +39,8 @@ write_v0!(v0, system, density_calculator) = v0
 
 @inline viscosity_model(system::FluidSystem) = system.viscosity
 
+@inline system_state_equation(system::FluidSystem) = system.state_equation
+
 function compute_density!(system, u, u_ode, semi, ::ContinuityDensity)
     # No density update with `ContinuityDensity`
     return system
