+++ Project.toml
@@ -39,7 +39,7 @@ GPUArraysCore = "0.1"
 JSON = "0.21"
 KernelAbstractions = "0.9"
 MuladdMacro = "0.2"
-PointNeighbors = "0.4.1"
+PointNeighbors = "0.4.2"
 Polyester = "0.7.5"
 RecipesBase = "1"
 Reexport = "1"
