### 2024-07-09 964c2e9
Use `@threaded` macro from PointNeighbors.jl (#573)
* Use `@threaded` from PointNeighbors.jl

* Remove `adapt_structure` definitions that are also in PointNeighbors.jl

* Use PointNeighbors.jl v0.4.1

* Use v0.4.2 of PointNeighbors.jl
