### 2024-07-09 #6961 @danielskatz
@HaoZeke - Here is a second possible submission for you to consider editing. If you can take one of the two, I would appreciate it. (Both would also be great, of course 🙂)
