### 2024-07-09 #6961 @HaoZeke
Thanks for the invite @danielskatz, I'm happy to accept

@LasNikas, do you have any suggestions from the [reviewer database](https://reviewers.joss.theoj.org/lookup)? I will start sending out invitations by the end of this week.
