+++ examples/fluid/dam_break_oil_film_2d.jl
@@ -94,8 +94,7 @@ boundary_system = BoundarySPHSystem(tank.boundary, boundary_model, adhesion_coef
 # ==========================================================================================
 # ==== Simulation
 semi = Semidiscretization(fluid_system, oil_system, boundary_system,
-                          neighborhood_search=GridNeighborhoodSearch,
-                          threaded_nhs_update=true)
+                          neighborhood_search=GridNeighborhoodSearch{2}(update_strategy=nothing))
 ode = semidiscretize(semi, tspan, data_type=nothing)
 
 info_callback = InfoCallback(interval=100)
