+++ src/schemes/fluid/viscosity.jl
@@ -141,8 +141,10 @@ end
     v_b = viscous_velocity(v_neighbor_system, neighbor_system, neighbor)
     v_diff = v_a - v_b
 
-    nu_a = kinematic_viscosity(particle_system, viscosity_model(particle_system, neighbor_system))
-    nu_b = kinematic_viscosity(neighbor_system, viscosity_model(neighbor_system, particle_system))
+    nu_a = kinematic_viscosity(particle_system,
+                               viscosity_model(particle_system, neighbor_system))
+    nu_b = kinematic_viscosity(neighbor_system,
+                               viscosity_model(neighbor_system, particle_system))
 
     pi_ab = viscosity(sound_speed, v_diff, pos_diff, distance, rho_mean, rho_a, rho_b,
                       smoothing_length, grad_kernel, nu_a, nu_b)
@@ -240,8 +242,10 @@ end
     (; smoothing_length) = particle_system
 
     epsilon = viscosity.epsilon
-    nu_a = kinematic_viscosity(particle_system, viscosity_model(particle_system, neighbor_system))
-    nu_b = kinematic_viscosity(neighbor_system, viscosity_model(neighbor_system, particle_system))
+    nu_a = kinematic_viscosity(particle_system,
+                               viscosity_model(particle_system, neighbor_system))
+    nu_b = kinematic_viscosity(neighbor_system,
+                               viscosity_model(neighbor_system, particle_system))
 
     v_a = viscous_velocity(v_particle_system, particle_system, particle)
     v_b = viscous_velocity(v_neighbor_system, neighbor_system, neighbor)
