+++ examples/fluid/dam_break_oil_film_2d.jl
@@ -7,37 +7,30 @@ using OrdinaryDiffEq
 H = 0.6
 W = 2 * H
 
-# ==========================================================================================
-# ==== Resolution
-fluid_particle_spacing = H / 60
-
-# Change spacing ratio to 3 and boundary layers to 1 when using Monaghan-Kajtar boundary model
-boundary_layers = 4
-spacing_ratio = 1
-
-boundary_particle_spacing = fluid_particle_spacing / spacing_ratio
-
-# ==========================================================================================
-# ==== Experiment Setup
 gravity = 9.81
 
-tspan = (0.0, 5.7 / sqrt(gravity))
+# Resolution
+fluid_particle_spacing = H / 60
 
-# Boundary geometry and initial fluid particle positions
-initial_fluid_size = (W, H)
-oil_size = (W, 0.1 * H)
-tank_size = (floor(5.366 * H / boundary_particle_spacing) * boundary_particle_spacing, 4.0)
-
-fluid_density = 1000.0
-oil_density = 700.0
-
+# Numerical settings
+smoothing_length = 3.5 * fluid_particle_spacing
 sound_speed = 20 * sqrt(gravity * H)
-state_equation = StateEquationCole(; sound_speed, reference_density=fluid_density,
-                                   exponent=1, clip_negative_pressure=false)
 
-tank = RectangularTank(fluid_particle_spacing, initial_fluid_size, tank_size, fluid_density,
-                       n_layers=boundary_layers, spacing_ratio=spacing_ratio,
-                       acceleration=(0.0, -gravity), state_equation=state_equation)
+nu = 0.02 * smoothing_length * sound_speed / 8
+oil_viscosity = ViscosityMorris(nu=10 * nu)
+
+trixi_include(@__MODULE__, joinpath(examples_dir(), "fluid", "dam_break_2d.jl"),
+              sol=nothing, fluid_particle_spacing=fluid_particle_spacing,
+              viscosity=ViscosityMorris(nu=nu), smoothing_length=smoothing_length,
+              gravity=gravity,
+              density_diffusion=DensityDiffusionMolteniColagrossi(delta=0.1),
+              sound_speed=sound_speed)
+
+# ==========================================================================================
+# ==== Setup oil layer
+
+oil_size = (W, 0.1 * H)
+oil_density = 700.0
 
 oil = RectangularShape(fluid_particle_spacing,
                        round.(Int, oil_size ./ fluid_particle_spacing),
@@ -48,30 +41,6 @@ for i in axes(oil.coordinates, 2)
     oil.coordinates[:, i] .+= [0.0, H]
 end
 
-# ==========================================================================================
-# ==== Fluid
-smoothing_length = 3.5 * fluid_particle_spacing
-smoothing_kernel = WendlandC2Kernel{2}()
-
-fluid_density_calculator = ContinuityDensity()
-
-nu = 0.02 * smoothing_length * sound_speed / 8
-viscosity = ViscosityMorris(nu=nu)
-oil_viscosity = ViscosityMorris(nu=10 * nu)
-
-# Alternatively the density diffusion model by Molteni & Colagrossi can be used,
-# which will run faster.
-density_diffusion = DensityDiffusionMolteniColagrossi(delta=0.1)
-# density_diffusion = DensityDiffusionAntuono(tank.fluid, delta=0.1)
-# oil_density_diffusion = DensityDiffusionAntuono(oil, delta=0.1)
-
-fluid_system = WeaklyCompressibleSPHSystem(tank.fluid, fluid_density_calculator,
-                                           state_equation, smoothing_kernel,
-                                           smoothing_length, viscosity=viscosity,
-                                           density_diffusion=density_diffusion,
[…]
