+++ examples/preprocessing/complex_shape_3d.jl
@@ -12,10 +12,12 @@ filename = joinpath("examples", "preprocessing", file * ".stl")
 # Returns `Shape`
 shape = load_shape(filename)
 
+point_in_shape_algorithm = WindingNumberJacobson(; shape,
+                                                 # winding_number_factor=0.4,
+                                                 hierarchical_winding=true)
+
 # Returns `InitialCondition`.
 shape_sampled = ComplexShape(shape; particle_spacing, density=1.0,
-                             point_in_shape_algorithm=WindingNumberJacobson(; shape,
-                                                                            # winding_number_factor=0.4,
-                                                                            hierarchical_winding=true))
+                             point_in_shape_algorithm)
 
 trixi2vtk(shape_sampled)
