+++ src/preprocessing/preprocessing.jl
@@ -1,2 +1,2 @@
 include("shapes/shapes.jl")
-include("point_in_poly/algorithm.jl")
+include("point_in_poly/point_in_poly.jl")
