+++ src/preprocessing/point_in_poly/point_in_poly.jl
@@ -1,3 +1,3 @@
+include("hierarchical_winding.jl")
 include("winding_number_horman.jl")
 include("winding_number_jacobson.jl")
-include("hierarchical_winding.jl")
