### 2024-07-11 f7d7eaf
for consitency lets change the nu assignment
