+++ src/schemes/boundary/system.jl
@@ -383,6 +383,8 @@ function restart_with!(system::BoundarySPHSystem{<:BoundaryModelDummyParticles{C
     return system
 end
 
+@inline viscosity_model(system::BoundarySPHSystem, neighbor_system::FluidSystem) = neighbor_system.viscosity
+
 function calculate_dt(v_ode, u_ode, cfl_number, system::BoundarySystem)
     return Inf
 end
