### 2024-07-12 #6961 @luraess
Hi @HaoZeke , I would be able to review this paper. However, I am on holiday the coming 3 weeks and will thus not be able to give any feedback before mid to end August. If that's OK, then count me in.
