### 2024-07-12 #6961 @williamfgc
@HaoZeke likewise, the rest of the month is going to be difficult for me. I can start next month. Also, I want to disclose a potential conflict since we collaborate actively with one of the authors (Michael) on Julia venues and a pub.
