### 2024-07-15 99665f1
Implement suggestions
