+++ src/general/neighborhood_search.jl
@@ -15,8 +15,8 @@ function PointNeighbors.foreach_point_neighbor(f, system::GPUSystem, neighbor_sy
                                                neighborhood_search;
                                                points=each_moving_particle(system),
                                                parallel=true)
-    # For `GPUSystem`s, explicitly pass the backend, so that with a `GPUSystem` with CPU
-    # backend, it will actually launch the KernelAbstractions.jl kernels on the CPU.
+    # For `GPUSystem`s, explicitly pass the backend, so a `GPUSystem` with a CPU
+    # backend will actually launch the KernelAbstractions.jl kernels on the CPU.
     foreach_point_neighbor(f, system_coords, neighbor_coords, neighborhood_search;
                            points, parallel=KernelAbstractions.get_backend(system_coords))
 end
