+++ src/schemes/boundary/open_boundary/system.jl
@@ -482,8 +482,9 @@ function wrap_reference_function(constant_vector_, ::Val{NDIMS}) where {NDIMS}
     return constant_vector(coords, t) = SVector{NDIMS}(constant_vector_)
 end
 
-# The open boundary system either uses the fluids viscosity model or at the boundary the model
-# of the boundary which is the normal behavior to incorporate the boundaries impact in the viscosity term of the rhs.
+# At the boundary, the viscosity model of the boundary has to be used to incorporate
+# the boundary's impact in the viscosity term of the rhs.
 @inline viscosity_model(system::OpenBoundarySPHSystem, neighbor_system::FluidSystem) = neighbor_system.viscosity
 @inline viscosity_model(system::OpenBoundarySPHSystem, neighbor_system::BoundarySystem) = neighbor_system.boundary_model.viscosity
+# When the neighbor is an open boundary system, just use the viscosity of the fluid `system` instead
 @inline viscosity_model(system, neighbor_system::OpenBoundarySPHSystem) = system.viscosity
