+++ src/schemes/fluid/fluid.jl
@@ -37,8 +37,8 @@ end
 
 write_v0!(v0, system, density_calculator) = v0
 
-# At the boundary, the viscosity model of the boundary has to be used to incorporate
-# the boundary's impact in the viscosity term of the rhs.
+# To account for boundary effects in the viscosity term of the RHS, use the viscosity model
+# of the neighboring particle systems.
 @inline viscosity_model(system::FluidSystem, neighbor_system::FluidSystem) = neighbor_system.viscosity
 @inline viscosity_model(system::FluidSystem, neighbor_system::BoundarySystem) = neighbor_system.boundary_model.viscosity
 
