### 2024-07-15 b82162f
Fix viscosity calculation when using two different fluids (#569)
* extract changes from #564

* fix

* fix nus

* fix

* add example

* fix comment

* format

* fix tests

* fix tlsph

* fix error

* format

* remove unecessary include

* implement review suggestions

* Fix examples.jl after Case move

* Forgot a part

* fix example

* review implementation

* format & example restruct

* fix test

* fix tests

* fix

* for consitency lets change the nu assignment

* fix tests

* implement suggestion

* change comment

* change comment again
