+++ test/examples/examples.jl
@@ -87,6 +87,18 @@
             @test count_rhs_allocations(sol, semi) == 0
         end
 
+        @trixi_testset "fluid/dam_break_oil_film_2d.jl" begin
+            @test_nowarn_mod trixi_include(@__MODULE__,
+                                           joinpath(examples_dir(), "fluid",
+                                                    "dam_break_oil_film_2d.jl"),
+                                           tspan=(0.0, 0.1)) [
+                r"┌ Info: The desired tank length in y-direction .*\n",
+                r"└ New tank length in y-direction.*\n",
+            ]
+            @test sol.retcode == ReturnCode.Success
+            @test count_rhs_allocations(sol, semi) == 0
+        end
+
         @trixi_testset "fluid/dam_break_2d.jl with KernelAbstractions.jl" begin
             # Emulate the GPU code on the CPU by passing `data_type = Array`
             @test_nowarn_mod trixi_include(@__MODULE__,
