### 2024-07-15 b3219cb
GPU-compatible neighborhood search (#563)
* Use PointNeighbors.jl#45

* Reformat

* Implement suggestions

---------

Co-authored-by: Niklas Neher <73897120+LasNikas@users.noreply.github.com>
