+++ src/setups/complex_shape.jl
@@ -4,7 +4,7 @@
                  point_in_shape_algorithm=WindingNumberJacobson(; shape,
                                                                 hierarchical_winding=false,
                                                                 winding_number_factor=sqrt(eps())),
-                 seed=shape.min_corner, max_nparticles=Int(1e7),
+                 grid_offset::AbstractFloat=0.0, max_nparticles=10^7,
                  pad_initial_particle_grid=2particle_spacing)
 
 Complex shape filled with particles. Returns an [`InitialCondition`](@ref).
@@ -32,8 +32,8 @@ For more information about the method see [`WindingNumberJacobson`](@ref) or [`W
 - `point_in_shape_algorithm`: Algorithm for sampling the complex shape with particles.
                               It basically checks whether a particle is inside an object or not.
                               For more information see [`WindingNumberJacobson`](@ref) or [`WindingNumberHorman`](@ref)
-- `seed`:               Coordinates of the minimum corner of the initial particle grid (excluding padding).
-- `max_nparticles`:     Maximum number of particles in the initial particle grid.
+- `grid_offset`: Offset of the initial particle grid to the minimum corner of the bounding box of the `shape`.
+- `max_nparticles`: Maximum number of particles in the initial particle grid.
 - `pad_initial_particle_grid`: Padding of the initial particle grid.
 
 
@@ -45,15 +45,16 @@ function ComplexShape(shape::Shapes; particle_spacing, density, pressure=0.0, ma
                       point_in_shape_algorithm=WindingNumberJacobson(; shape,
                                                                      hierarchical_winding=false,
                                                                      winding_number_factor=sqrt(eps())),
-                      store_winding_number=false, seed=shape.min_corner,
-                      max_nparticles=Int(1e7), pad_initial_particle_grid=2particle_spacing)
+                      store_winding_number=false, grid_offset::AbstractFloat=0.0,
+                      max_nparticles=10^7, pad_initial_particle_grid=2particle_spacing)
     if ndims(shape) == 3 && point_in_shape_algorithm isa WindingNumberHorman
         throw(ArgumentError("`WindingNumberHorman` only supports 2D shapes"))
     end
 
     return sample(shape; particle_spacing, density, pressure, mass, velocity,
-                  point_in_shape_algorithm, store_winding_number, seed, max_nparticles,
-                  pad=pad_initial_particle_grid)
+                  point_in_shape_algorithm, store_winding_number, grid_offset,
+                  max_nparticles,
+                  padding=pad_initial_particle_grid)
 end
 
 function sample(shape::Shapes; particle_spacing, density, pressure=0.0, mass=nothing,
@@ -61,9 +62,10 @@ function sample(shape::Shapes; particle_spacing, density, pressure=0.0, mass=not
                 point_in_shape_algorithm=WindingNumberJacobson(; shape,
                                                                hierarchical_winding=false,
                                                                winding_number_factor=sqrt(eps())),
-                store_winding_number=false, seed=shape.min_corner, max_nparticles=Int(1e7),
-                pad=2particle_spacing)
-    grid = particle_grid(shape, particle_spacing; pad, seed, max_nparticles)
+                store_winding_number=false, grid_offset=shape.min_corner,
+                max_nparticles=10^7,
+                padding=2particle_spacing)
+    grid = particle_grid(shape, particle_spacing; padding, grid_offset, max_nparticles)
 
[…]
