### 2024-07-15 0f3536c
add new bnd density calculator
