+++ src/schemes/boundary/dummy_particles/dummy_particles.jl
@@ -88,6 +88,24 @@ struct AdamiPressureExtrapolation{ELTYPE}
 end
 
 @doc raw"""
+    BernoulliPressureExtrapolation(; pressure_offset=0.0)
+
+`density_calculator` for `BoundaryModelDummyParticles`.
+
+# Keywords
+- `pressure_offset=0.0`: Sometimes it is necessary to artificially increase the boundary pressure
+                         to prevent penetration, which is possible by increasing this value.
+
+"""
+struct BernoulliPressureExtrapolation{ELTYPE}
+    pressure_offset::ELTYPE
+
+    function BernoulliPressureExtrapolation(; pressure_offset=0.0)
+        return new{eltype(pressure_offset)}(pressure_offset)
+    end
+end
+
+@doc raw"""
     PressureMirroring()
 
 `density_calculator` for `BoundaryModelDummyParticles`.
@@ -145,7 +163,9 @@ end
 
 @inline create_cache_model(initial_density, ::ContinuityDensity) = (; initial_density)
 
-function create_cache_model(initial_density, ::AdamiPressureExtrapolation)
+function create_cache_model(initial_density,
+                            ::Union{AdamiPressureExtrapolation,
+                                    BernoulliPressureExtrapolation})
     density = copy(initial_density)
     volume = similar(initial_density)
 
@@ -205,7 +225,7 @@ end
 # Note that the other density calculators are dispatched in `density_calculators.jl`
 @inline function particle_density(v,
                                   ::Union{AdamiPressureExtrapolation, PressureMirroring,
-                                          PressureZeroing},
+                                          PressureZeroing, BernoulliPressureExtrapolation},
                                   boundary_model, particle)
     (; cache) = boundary_model
 
@@ -227,10 +247,11 @@ end
 
 function compute_density!(boundary_model,
                           ::Union{ContinuityDensity, AdamiPressureExtrapolation,
+                                  BernoulliPressureExtrapolation,
                                   PressureMirroring, PressureZeroing},
                           system, v, u, v_ode, u_ode, semi)
     # No density update for `ContinuityDensity`, `PressureMirroring` and `PressureZeroing`.
-    # For `AdamiPressureExtrapolation`, the density is updated in `compute_pressure!`.
+    # For `AdamiPressureExtrapolation` and `BernoulliPressureExtrapolation`, the density is updated in `compute_pressure!`.
     return boundary_model
 end
 
@@ -310,7 +331,10 @@ end
     boundary_model.pressure[particle] = max(boundary_model.state_equation(density), 0.0)
 end
 
-function compute_pressure!(boundary_model, ::AdamiPressureExtrapolation, system, v, u,
+function compute_pressure!(boundary_model,
+                           bnd_density_calc::Union{AdamiPressureExtrapolation,
+                                                   BernoulliPressureExtrapolation}, system,
+                           v, u,
                            v_ode, u_ode, semi)
     (; pressure, cache, viscosity) = boundary_model
 
@@ -338,16 +362,18 @@ function compute_pressure!(boundary_model, ::AdamiPressureExtrapolation, system,
             nhs = get_neighborhood_search(neighbor_system, system, semi)
 
             # Loop over fluid particles and then the neighboring boundary particles to extrapolate fluid pressure to the boundaries
-            adami_pressure_extrapolation_neighbor!(boundary_model, system, neighbor_system,
-                                                   system_coords, neighbor_coords, v,
-                                                   v_neighbor_system, nhs)
+            bnd_pressure_extrapolation_neighbor!(boundary_model, bnd_density_calc, system,
+                                                 neighbor_system,
+                                                 system_coords, neighbor_coords, v,
+                                                 v_neighbor_system, nhs)
         else
[…]
