+++ src/general/system.jl
@@ -93,6 +93,11 @@ end
 
 @inline current_velocity(v, system, particle) = extract_svector(v, system, particle)
 
+@inline function current_acceleration(system, particle)
+    # TODO: Return `dv` of solid particles
+    return zero(SVector{ndims(system), eltype(system)})
+end
+
 @inline function smoothing_kernel(system, distance)
     (; smoothing_kernel, smoothing_length) = system
     return kernel(smoothing_kernel, distance, smoothing_length)
