+++ test/schemes/boundary/dummy_particles/dummy_particles.jl
@@ -54,12 +54,14 @@
 
             TrixiParticles.reset_cache!(boundary_system.boundary_model.cache,
                                         boundary_system.boundary_model.viscosity)
-            TrixiParticles.adami_pressure_extrapolation!(boundary_model, boundary_system,
-                                                         fluid_system, boundary.coordinates,
-                                                         fluid.coordinates, v_fluid,
-                                                         v_fluid .*
-                                                         ones(size(fluid.coordinates)),
-                                                         neighborhood_search)
+            TrixiParticles.bnd_pressure_extrapolation!(boundary_model,
+                                                       AdamiPressureExtrapolation(),
+                                                       boundary_system,
+                                                       fluid_system, boundary.coordinates,
+                                                       fluid.coordinates, v_fluid,
+                                                       v_fluid .*
+                                                       ones(size(fluid.coordinates)),
+                                                       neighborhood_search)
 
             for particle in TrixiParticles.eachparticle(boundary_system)
                 if volume[particle] > eps()
@@ -92,11 +94,13 @@
 
             TrixiParticles.reset_cache!(boundary_system.boundary_model.cache,
                                         boundary_system.boundary_model.viscosity)
-            TrixiParticles.adami_pressure_extrapolation!(boundary_model, boundary_system,
-                                                         fluid_system, boundary.coordinates,
-                                                         fluid.coordinates, v_fluid,
-                                                         v_fluid,
-                                                         neighborhood_search)
+            TrixiParticles.bnd_pressure_extrapolation!(boundary_model,
+                                                       AdamiPressureExtrapolation(),
+                                                       boundary_system,
+                                                       fluid_system, boundary.coordinates,
+                                                       fluid.coordinates, v_fluid,
+                                                       v_fluid,
+                                                       neighborhood_search)
 
             for particle in TrixiParticles.eachparticle(boundary_system)
                 if volume[particle] > eps()
