+++ test/schemes/boundary/dummy_particles/dummy_particles.jl
@@ -55,7 +55,6 @@
             TrixiParticles.reset_cache!(boundary_system.boundary_model.cache,
                                         boundary_system.boundary_model.viscosity)
             TrixiParticles.bnd_pressure_extrapolation!(boundary_model,
-                                                       AdamiPressureExtrapolation(),
                                                        boundary_system,
                                                        fluid_system, boundary.coordinates,
                                                        fluid.coordinates, v_fluid,
@@ -94,9 +93,7 @@
 
             TrixiParticles.reset_cache!(boundary_system.boundary_model.cache,
                                         boundary_system.boundary_model.viscosity)
-            TrixiParticles.bnd_pressure_extrapolation!(boundary_model,
-                                                       AdamiPressureExtrapolation(),
-                                                       boundary_system,
+            TrixiParticles.bnd_pressure_extrapolation!(boundary_model, boundary_system,
                                                        fluid_system, boundary.coordinates,
                                                        fluid.coordinates, v_fluid,
                                                        v_fluid,
