+++ docs/src/systems/boundary.md
@@ -60,7 +60,7 @@ We provide six options to compute the boundary density and pressure, determined 
    fluid according to (Adami et al., 2012), and the density is obtained by applying the inverse of the state equation.
    This option usually yields the best results of the options listed here.
 2. (Only relevant for FSI) With `BernoulliPressureExtrapolation`, the pressure is extrapolated from the 
-   pressure similar to the `AdamiPressureExtrapolation`, but a relative velocity dependent pressure part
+   pressure similar to the [`AdamiPressureExtrapolation`](@ref), but a relative velocity dependent pressure part
    is calculated between moving solids and fluids, which increases the boundary pressure in areas prone to 
    penetrations.
 3. With [`SummationDensity`](@ref), the density is calculated by summation over the neighboring particles,
@@ -97,7 +97,12 @@ where the sum is over all fluid particles, ``\rho_f`` and ``p_f`` denote the den
     AdamiPressureExtrapolation
 ```
 
-#### 4. [`PressureZeroing`](@ref)
+#### 2. [`BernoulliPressureExtrapolation`](@ref)
+```@docs
+    BernoulliPressureExtrapolation
+```
+
+#### 5. [`PressureZeroing`](@ref)
 
 This is the simplest way to implement dummy boundary particles.
 The density of each particle is set to the reference density and the pressure to the
@@ -106,7 +111,7 @@ reference pressure (the corresponding pressure to the reference density by the s
     PressureZeroing
 ```
 
-#### 5. [`PressureMirroring`](@ref)
+#### 6. [`PressureMirroring`](@ref)
 
 Instead of calculating density and pressure for each boundary particle, we modify the
 momentum equation,
