+++ src/schemes/boundary/open_boundary/system.jl
@@ -1,14 +1,12 @@
 @doc raw"""
     OpenBoundarySPHSystem(boundary_zone::Union{InFlow, OutFlow}; sound_speed,
                           fluid_system::FluidSystem, buffer_size::Integer,
-                          reference_velocity=zeros(ndims(boundary_zone)),
-                          reference_pressure=0.0,
-                          reference_density=first(boundary_zone.initial_condition.density))
+                          boundary_model,
+                          reference_velocity=nothing,
+                          reference_pressure=nothing,
+                          reference_density=nothing)
 
 Open boundary system for in- and outflow particles.
-These open boundaries use the characteristic variables to propagate the appropriate values
-to the outlet or inlet and have been proposed by Lastiwka et al. (2009). For more information
-about the method see [description below](@ref method_of_characteristics).
 
 # Arguments
 - `boundary_zone`: Use [`InFlow`](@ref) for an inflow and [`OutFlow`](@ref) for an outflow boundary.
@@ -16,48 +14,51 @@ about the method see [description below](@ref method_of_characteristics).
 # Keywords
 - `sound_speed`: Speed of sound.
 - `fluid_system`: The corresponding fluid system
+- `boundary_model`: Boundary model (see [BoundaryModelLastiwka](@ref))
 - `buffer_size`: Number of buffer particles.
 - `reference_velocity`: Reference velocity is either a function mapping each particle's coordinates
                         and time to its velocity, an array where the ``i``-th column holds
                         the velocity of particle ``i`` or, for a constant fluid velocity,
-                        a vector holding this velocity. Velocity is constant zero by default.
+                        a vector holding this velocity.
 - `reference_pressure`: Reference pressure is either a function mapping each particle's coordinates
                         and time to its pressure, a vector holding the pressure of each particle,
                         or a scalar for a constant pressure over all particles.
-                        Pressure is constant zero by default.
 - `reference_density`: Reference density is either a function mapping each particle's coordinates
                        and time to its density, a vector holding the density of each particle,
                        or a scalar for a constant density over all particles.
-                       Density is the density of the first particle in the initial condition by default.
 
 !!! warning "Experimental Implementation"
 	This is an experimental feature and may change in any future releases.
 """
-struct OpenBoundarySPHSystem{BZ, NDIMS, ELTYPE <: Real, IC, FS, ARRAY1D, ARRAY2D, RV, RP,
-                             RD, B} <: System{NDIMS, IC}
-    initial_condition        :: IC
-    fluid_system             :: FS
-    mass                     :: ARRAY1D # Array{ELTYPE, 1}: [particle]
-    density                  :: ARRAY1D # Array{ELTYPE, 1}: [particle]
-    volume                   :: ARRAY1D # Array{ELTYPE, 1}: [particle]
-    pressure                 :: ARRAY1D # Array{ELTYPE, 1}: [particle]
-    characteristics          :: ARRAY2D # Array{ELTYPE, 2}: [characteristic, particle]
-    previous_characteristics :: ARRAY2D # Array{ELTYPE, 2}: [characteristic, particle]
-    sound_speed              :: ELTYPE
-    boundary_zone            :: BZ
-    flow_direction           :: SVector{NDIMS, ELTYPE}
-    reference_velocity       :: RV
-    reference_pressure       :: RP
-    reference_density        :: RD
-    buffer                   :: B
-    update_callback_used     :: Ref{Bool}
+struct OpenBoundarySPHSystem{BM, BZ, NDIMS, ELTYPE <: Real, IC, FS, ARRAY1D, RV,
+                             RP, RD, B, C} <: System{NDIMS, IC}
[…]
