### 2024-07-16 dda2202
remove abstract type `Shapes`
