+++ src/preprocessing/shapes/shapes.jl
@@ -1,12 +1,6 @@
-abstract type Shapes{NDIMS} end
-
 include("polygon_shape.jl")
 include("triangle_mesh.jl")
 include("io.jl")
-
-@inline Base.ndims(::Shapes{NDIMS}) where {NDIMS} = NDIMS
-
-@inline Base.eltype(shape::Shapes) = eltype(first(first(shape.min_corner)))
 
 @inline eachface(mesh) = Base.OneTo(nfaces(mesh))
 
