+++ src/general/system.jl
@@ -98,6 +98,10 @@ end
     return zero(SVector{ndims(system), eltype(system)})
 end
 
+@inline set_particle_density!(v, system, particle, density) = v
+
+@inline set_particle_pressure!(v, system, particle, pressure) = v
+
 @inline function smoothing_kernel(system, distance)
     (; smoothing_kernel, smoothing_length) = system
     return kernel(smoothing_kernel, distance, smoothing_length)
