### 2024-07-16 46e4647
make type stable
