### 2024-07-16 eeec66a
implement suggestions
