+++ examples/fluid/falling_water_spheres_2d.jl
@@ -50,26 +50,23 @@ alpha = 8 * nu / (fluid_smoothing_length * sound_speed)
 viscosity = ArtificialViscosityMonaghan(alpha=alpha, beta=0.0)
 density_diffusion = DensityDiffusionAntuono(sphere2, delta=0.1)
 
-# sphere_surface_tension = WeaklyCompressibleSPHSystem(sphere1, fluid_density_calculator,
-#                                                      state_equation, fluid_smoothing_kernel,
-#                                                      fluid_smoothing_length,
-#                                                      viscosity=viscosity,
-#                                                      acceleration=(0.0, -gravity),
-#                                                      surface_tension=SurfaceTensionAkinci(surface_tension_coefficient=0.05),
-#                                                      correction=AkinciFreeSurfaceCorrection(fluid_density),
-#                                                      surface_normal_method=AkinciSurfaceNormal(smoothing_kernel=WendlandC6Kernel{3}(),
-#                                                                                                smoothing_length=4 *
-#                                                                                                                 fluid_particle_spacing))
-
-sphere_surface_tension = EntropicallyDampedSPHSystem(sphere1, fluid_smoothing_kernel,
+sphere_surface_tension = WeaklyCompressibleSPHSystem(sphere1, fluid_density_calculator,
+                                                     state_equation, fluid_smoothing_kernel,
                                                      fluid_smoothing_length,
-                                                     sound_speed, viscosity=viscosity,
-                                                     density_calculator=ContinuityDensity(),
+                                                     viscosity=viscosity,
                                                      acceleration=(0.0, -gravity),
                                                      surface_tension=SurfaceTensionAkinci(surface_tension_coefficient=0.05),
-                                                    surface_normal_method=AkinciSurfaceNormal(smoothing_kernel=WendlandC6Kernel{3}(),
-                                                                                            smoothing_length=4 *
-                                                                                                                fluid_particle_spacing))
+                                                     correction=AkinciFreeSurfaceCorrection(fluid_density))
+
+# sphere_surface_tension = EntropicallyDampedSPHSystem(sphere1, fluid_smoothing_kernel,
+#                                                      fluid_smoothing_length,
+#                                                      sound_speed, viscosity=viscosity,
+#                                                      density_calculator=ContinuityDensity(),
+#                                                      acceleration=(0.0, -gravity),
+#                                                      surface_tension=SurfaceTensionAkinci(surface_tension_coefficient=0.05),
+#                                                     surface_normal_method=AkinciSurfaceNormal(smoothing_kernel=WendlandC6Kernel{3}(),
+#                                                                                             smoothing_length=4 *
+#                                                                                                                 fluid_particle_spacing))
 
 # sphere = WeaklyCompressibleSPHSystem(sphere2, fluid_density_calculator,
 #                                      state_equation, fluid_smoothing_kernel,
@@ -97,7 +94,7 @@ ode = semidiscretize(semi, tspan)
 
 info_callback = InfoCallback(interval=50)
 saving_callback = SolutionSavingCallback(dt=0.01, output_directory="out",
-                                         prefix="higherWend", write_meta_data=true)
+                                         prefix="", write_meta_data=true)
 
 callbacks = CallbackSet(info_callback, saving_callback)
 
