### 2024-07-17 #6961 @giordano
I will also be busy until mid-September between holidays and conference-going (unless I carve some time out of my holidays, but don't count too much on this :slightly_smiling_face:), otherwise I'm happy to do this.

As William said, I also personally know some of the authors (Erik and Michael), but I'm not involved in the subject of the paper.  I know that knowing authors hasn't been a reason for rejecting JOSS reviewers in the past (I don't know if policies have changed since), but I'm pointing this out for transparency.
