### 2024-07-17 44ce444
add `unique_soted`
