+++ src/schemes/fluid/surface_normal_sph.jl
@@ -10,10 +10,8 @@ end
 # Section 2.2 in Akinci et al. 2013 "Versatile Surface Tension and Adhesion for SPH Fluids"
 # Note: This is the simplest form of normal approximation commonly used in SPH and comes
 # with serious deficits in accuracy especially at corners, small neighborhoods and boundaries
-function calc_normal_akinci!(system, neighbor_system::FluidSystem,
-                             surface_tension::SurfaceTensionAkinci, u_system,
-                             v_neighbor_system, u_neighbor_system,
-                             semi, surfn)
+function calc_normal_akinci!(system, neighbor_system::FluidSystem, u_system, v,
+                             v_neighbor_system, u_neighbor_system, semi, surfn)
     (; cache) = system
     (; smoothing_kernel, smoothing_length) = surfn
 
@@ -21,18 +19,18 @@ function calc_normal_akinci!(system, neighbor_system::FluidSystem,
     neighbor_system_coords = current_coordinates(u_neighbor_system, neighbor_system)
     nhs = get_neighborhood_search(system, neighbor_system, semi)
 
-    # println("sml ", smoothing_length != system.smoothing_length)
-    # println("smk ", smoothing_kernel !== system.smoothing_kernel)
     if smoothing_length != system.smoothing_length ||
        smoothing_kernel !== system.smoothing_kernel
+        # TODO: this is really slow but there is no way to easily implement multiple search radia
         search_radius = compact_support(smoothing_kernel, smoothing_length)
-        nhs = PointNeighbors.copy_neighborhood_search(nhs, search_radius, nparticles(system))
+        nhs = PointNeighbors.copy_neighborhood_search(nhs, search_radius,
+                                                      nparticles(system))
         PointNeighbors.initialize!(nhs, system_coords, neighbor_system_coords)
     end
 
     foreach_point_neighbor(system, neighbor_system,
-                          system_coords, neighbor_system_coords,
-                          nhs) do particle, neighbor, pos_diff, distance
+                           system_coords, neighbor_system_coords,
+                           nhs) do particle, neighbor, pos_diff, distance
         m_b = hydrodynamic_mass(neighbor_system, neighbor)
         density_neighbor = particle_density(v_neighbor_system,
                                             neighbor_system, neighbor)
@@ -50,9 +48,95 @@ function calc_normal_akinci!(system, neighbor_system::FluidSystem,
     return system
 end
 
-function calc_normal_akinci!(system, neighbor_system, surface_tension, u_system,
-                             v_neighbor_system, u_neighbor_system,
-                             semi, surfn)
+# Section 2.2 in Akinci et al. 2013 "Versatile Surface Tension and Adhesion for SPH Fluids"
+# Note: This is the simplest form of normal approximation commonly used in SPH and comes
+# with serious deficits in accuracy especially at corners, small neighborhoods and boundaries
+function calc_normal_akinci!(system, neighbor_system::BoundarySystem, u_system, v,
+                             v_neighbor_system, u_neighbor_system, semi, surfn)
+    (; cache) = system
+    (; smoothing_kernel, smoothing_length) = surfn
+
+    system_coords = current_coordinates(u_system, system)
+    neighbor_system_coords = current_coordinates(u_neighbor_system, neighbor_system)
+    nhs = get_neighborhood_search(system, neighbor_system, semi)
+    # nhs_bnd = get_neighborhood_search(neighbor_system, neighbor_system, semi)
+
+    # if smoothing_length != system.smoothing_length ||
+    #    smoothing_kernel !== system.smoothing_kernel
+    #     # TODO: this is really slow but there is no way to easily implement multiple search radia
[…]
