### 2024-07-18 09c6462
implement suggestions
