+++ README.md
@@ -21,6 +21,7 @@ Its features include:
 - Solid-body mechanics
   - Methods:  Total Lagrangian SPH (TLSPH), Discrete Element Method (DEM)
 - Fluid-Structure Interaction
+- Preprocess complex shapes
 - Output formats:
   - VTK
 
