### 2024-07-18 05e811a
implement suggestions
