+++ test/systems/open_boundary_system.jl
@@ -52,17 +52,13 @@
                                                                     reference_pressure=0)
     end
     @testset "`show`" begin
-        f1(pos, t) = SVector(0.0, 0.0)
-        f2(pos, t) = 0.0
-        f3(pos, t) = 0.0
-
         inflow = InFlow(; plane=([0.0, 0.0], [0.0, 1.0]), particle_spacing=0.05,
                         flow_direction=(1.0, 0.0), density=1.0, open_boundary_layers=4)
         system = OpenBoundarySPHSystem(inflow; sound_speed=1.0, buffer_size=0,
                                        boundary_model=BoundaryModelLastiwka(),
-                                       reference_density=f1,
-                                       reference_pressure=f2,
-                                       reference_velocity=f3,
+                                       reference_density=0.0,
+                                       reference_pressure=0.0,
+                                       reference_velocity=[0.0, 0.0],
                                        fluid_system=FluidSystemMock2())
 
         show_compact = "OpenBoundarySPHSystem{2}(InFlow) with 80 particles"
@@ -77,25 +73,21 @@
         │ boundary model: ……………………………………… BoundaryModelLastiwka                                            │
         │ boundary: ……………………………………………………… InFlow                                                           │
         │ flow direction: ……………………………………… [1.0, 0.0]                                                       │
-        │ prescribed velocity: ………………………… #f │
-        │ prescribed pressure: ………………………… #f │
-        │ prescribed density: …………………………… #f │
+        │ prescribed velocity: ………………………… constant_vector                                                  │
+        │ prescribed pressure: ………………………… constant_scalar                                                  │
+        │ prescribed density: …………………………… constant_scalar                                                  │
         │ width: ……………………………………………………………… 0.2                                                              │
         └──────────────────────────────────────────────────────────────────────────────────────────────────┘"""
 
-        result = repr("text/plain", system)
-        # Remove numbers like `#f1#83`
-        result = replace(result, r"#f\d#.*|" => "#f |")
-
-        @test result == show_box
+        @test repr("text/plain", system) == show_box
 
         outflow = OutFlow(; plane=([0.0, 0.0], [0.0, 1.0]), particle_spacing=0.05,
                           flow_direction=(1.0, 0.0), density=1.0, open_boundary_layers=4)
         system = OpenBoundarySPHSystem(outflow; sound_speed=1.0, buffer_size=0,
                                        boundary_model=BoundaryModelLastiwka(),
-                                       reference_density=f1,
-                                       reference_pressure=f2,
-                                       reference_velocity=f3,
+                                       reference_density=0.0,
+                                       reference_pressure=0.0,
+                                       reference_velocity=[0.0, 0.0],
                                        fluid_system=FluidSystemMock2())
 
         show_compact = "OpenBoundarySPHSystem{2}(OutFlow) with 80 particles"
@@ -110,16 +102,12 @@
         │ boundary model: ……………………………………… BoundaryModelLastiwka                                            │
         │ boundary: ……………………………………………………… OutFlow                                                          │
         │ flow direction: ……………………………………… [1.0, 0.0]                                                       │
-        │ prescribed velocity: ………………………… #f │
-        │ prescribed pressure: ………………………… #f │
-        │ prescribed density: …………………………… #f │
+        │ prescribed velocity: ………………………… constant_vector                                                  │
+        │ prescribed pressure: ………………………… constant_scalar                                                  │
+        │ prescribed density: …………………………… constant_scalar                                                  │
         │ width: ……………………………………………………………… 0.2                                                              │
         └──────────────────────────────────────────────────────────────────────────────────────────────────┘"""
 
-        result = repr("text/plain", system)
[…]
