+++ src/schemes/boundary/system.jl
@@ -30,6 +30,8 @@ struct BoundarySPHSystem{BM, NDIMS, ELTYPE <: Real, IC, CO, M, IM,
                                ismoving, adhesion_coefficient, cache, buffer)
         ELTYPE = eltype(coordinates)
 
+        initialize_boundary_model!(boundary_model, initial_condition.particle_spacing)
+
         new{typeof(boundary_model), size(coordinates, 1), ELTYPE, typeof(initial_condition),
             typeof(coordinates), typeof(movement), typeof(ismoving),
             typeof(cache)}(initial_condition, coordinates, boundary_model, movement,
@@ -397,10 +399,15 @@ function initialize!(system::BoundarySPHSystem, neighborhood_search)
     (; smoothing_kernel, smoothing_length) = system.boundary_model
 
     foreach_point_neighbor(system, system,
-                            system_coords, system_coords,
-                            neighborhood_search, points=eachparticle(system)) do particle, neighbor, pos_diff, distance
-        system.boundary_model.cache.colorfield[particle] += kernel(smoothing_kernel,
-                                                                   distance,
-                                                                   smoothing_length)
+                           system_coords, system_coords,
+                           neighborhood_search,
+                           points=eachparticle(system)) do particle, neighbor, pos_diff,
+                                                           distance
+        system.boundary_model.cache.colorfield_bnd[particle] += kernel(smoothing_kernel,
+                                                                       distance,
+                                                                       smoothing_length)
+        system.boundary_model.cache.neighbor_count[particle] += 1
     end
 end
+
+initialize_boundary_model!(boundary_model, initial_particle_spacing) = boundary_model
