+++ src/schemes/boundary/boundary.jl
@@ -4,3 +4,5 @@ include("open_boundary/boundary_zones.jl")
 include("open_boundary/system.jl")
 # Monaghan-Kajtar repulsive boundary particles require the `BoundarySPHSystem`
 # and the `TotalLagrangianSPHSystem` and are therefore included later.
+
+@inline Base.ndims(boundary_model::BoundaryModelDummyParticles) = ndims(boundary_model.smoothing_kernel)
