### 2024-07-19 a687574
Inline viscosity
