+++ examples/fluid/falling_water_spheres_3d.jl
@@ -14,19 +14,19 @@ sound_speed = 50
 
 sphere1_radius = 0.05
 
-sphere1_center = (0.5, 0.5, 0.2)
-sphere2_center = (1.5, 0.5, 0.2)
+sphere1_center = (0.5, 0.5, 0.075)
+sphere2_center = (1.5, 0.5, 0.075)
 sphere1 = SphereShape(fluid_particle_spacing, sphere1_radius, sphere1_center,
-                      fluid_density, sphere_type=VoxelSphere(), velocity=(0.0, 0.0, -2.0))
+                      fluid_density, sphere_type=VoxelSphere(), velocity=(0.0, 0.0, -1.0))
 sphere2 = SphereShape(fluid_particle_spacing, sphere1_radius, sphere2_center,
-                      fluid_density, sphere_type=VoxelSphere(), velocity=(0.0, 0.0, -2.0))
+                      fluid_density, sphere_type=VoxelSphere(), velocity=(0.0, 0.0, -1.0))
 
 # `compact_support` needs to be `2.0 * particle_spacing` to be correct
 fluid_smoothing_length = 1.0 * fluid_particle_spacing
 
 trixi_include(@__MODULE__,
               joinpath(examples_dir(), "fluid", "falling_water_spheres_2d.jl"),
-              fluid_particle_spacing=fluid_particle_spacing, tspan=(0.0, 0.2),
+              fluid_particle_spacing=fluid_particle_spacing, tspan=(0.0, 0.1),
               initial_fluid_size=(0.0, 0.0, 0.0),
               tank_size=(2.0, 1.0, 0.1), sound_speed=sound_speed,
               faces=(true, true, true, true, true, false),
@@ -34,4 +34,4 @@ trixi_include(@__MODULE__,
               fluid_smoothing_length=fluid_smoothing_length,
               fluid_smoothing_kernel=SchoenbergCubicSplineKernel{3}(),
               nu=nu, alpha=10 * nu / (fluid_smoothing_length * sound_speed),
-              surface_tension_coefficient=3, adhesion_coefficient=0.25)
+              surface_tension_coefficient=10, adhesion_coefficient=0.1)
