+++ examples/fluid/falling_water_spheres_3d.jl
@@ -3,7 +3,7 @@ using OrdinaryDiffEq
 
 # ==========================================================================================
 # ==== Resolution
-fluid_particle_spacing = 0.0025
+fluid_particle_spacing = 0.005
 
 # ==========================================================================================
 # ==== Experiment Setup
