+++ examples/fluid/sphere_surface_tension_2d.jl
@@ -36,10 +36,7 @@ fluid_system = WeaklyCompressibleSPHSystem(fluid, SummationDensity(),
                                                                                  beta=0.0),
                                            surface_tension=SurfaceTensionAkinci(surface_tension_coefficient=0.02),
                                            correction=AkinciFreeSurfaceCorrection(fluid_density),
-                                           source_terms=source_terms,
-                                           surface_normal_method=AkinciSurfaceNormal(smoothing_kernel=WendlandC6Kernel{3}(),
-                                                                                     smoothing_length=3.0 *
-                                                                                                      particle_spacing))
+                                           source_terms=source_terms)
 
 # ==========================================================================================
 # ==== Simulation
@@ -51,7 +48,7 @@ ode = semidiscretize(semi, tspan)
 info_callback = InfoCallback(interval=100)
 
 # For overwriting via `trixi_include`
-saving_callback = SolutionSavingCallback(dt=0.02, prefix="wendland")
+saving_callback = SolutionSavingCallback(dt=0.02)
 
 stepsize_callback = StepsizeCallback(cfl=1.0)
 
