### 2024-07-22 2a47d65
Make `SolutionSavingCallback` work on GPUs
