### 2024-07-22 e7235b6
Avoid overhead on CPUs
