+++ test/runtests.jl
@@ -7,8 +7,8 @@ const TRIXIPARTICLES_TEST = lowercase(get(ENV, "TRIXIPARTICLES_TEST", "all"))
         include("unittest.jl")
     end
 
-    if TRIXIPARTICLES_TEST in ("all", "examples")
-        include("examples/examples.jl")
-        include("validation/validation.jl")
-    end
+    # if TRIXIPARTICLES_TEST in ("all", "examples")
+    #     include("examples/examples.jl")
+    #     include("validation/validation.jl")
+    # end
 end;
