### 2024-07-23 86f65c5
rename `shape` to `geometry`
