+++ examples/preprocessing/complex_shape_3d.jl
@@ -10,14 +10,14 @@ filename = joinpath("examples", "preprocessing", file * ".stl")
 # filename = joinpath("examples", "preprocessing", "drive_gear.stl")
 
 # Returns `Shape`
-shape = load_shape(filename)
+geometry = load_geometry(filename)
 
-point_in_shape_algorithm = WindingNumberJacobson(; shape,
-                                                 # winding_number_factor=0.4,
-                                                 hierarchical_winding=true)
+point_in_geometry_algorithm = WindingNumberJacobson(; geometry,
+                                                    # winding_number_factor=0.4,
+                                                    hierarchical_winding=true)
 
 # Returns `InitialCondition`.
-shape_sampled = ComplexShape(shape; particle_spacing, density=1.0,
-                             point_in_shape_algorithm)
+shape_sampled = ComplexShape(geometry; particle_spacing, density=1.0,
+                             point_in_geometry_algorithm)
 
 trixi2vtk(shape_sampled)
