+++ src/preprocessing/point_in_poly/winding_number_jacobson.jl
@@ -9,8 +9,8 @@ end
 @inline function naive_winding(mesh, faces, query_point)
     winding_number = sum(faces, init=0.0) do face
 
-        # A. Van Oosterom 1983,
-        # The Solid Angle of a Plane Triangle (doi: 10.1109/TBME.1983.325207)
+        # Eq. 6 of Jacobsen et al. Based on A. Van Oosterom (1983),
+        # "The Solid Angle of a Plane Triangle" (doi: 10.1109/TBME.1983.325207)
         a = face_vertex(mesh, face, 1) - query_point
         b = face_vertex(mesh, face, 2) - query_point
         c = face_vertex(mesh, face, 3) - query_point
@@ -20,36 +20,39 @@ end
 
         divisor = a_ * b_ * c_ + dot(a, b) * c_ + dot(b, c) * a_ + dot(c, a) * b_
 
-        return 2atan(det([a b c]), divisor)
+        return 2 * atan(det([a b c]), divisor)
     end
 
     return winding_number
 end
 
 """
-    WindingNumberJacobson(; shape=nothing, winding_number_factor=sqrt(eps()),
+    WindingNumberJacobson(; geometry=nothing, winding_number_factor=sqrt(eps()),
                           hierarchical_winding=false)
-Algorithm for inside-outside segmentation of a complex shape proposed by Jacobson et al. (2013).
+Algorithm for inside-outside segmentation of a complex geometry proposed by Jacobson et al. (2013).
 
 # Keywords
-- `shape`: Complex shape returned by [`load_shape`](@ref) and is only required when using
-           `hierarchical_winding=true`.
-- `hierarchical_winding`: If set to `true`, an optimised hierarchical approach will be used
+- `geometry`: Complex geometry returned by [`load_geometry`](@ref) and is only required when using
+              `hierarchical_winding=true`.
+- `hierarchical_winding`: If set to `true`, an optimized hierarchical approach will be used
                           which gives a significant speedup.
-                          It is only supported for 3D shapes yet.
-- `winding_number_factor`: For leaky shapes, a factor of `0.4` will give a better inside-outside segmentation.
+                          It is only supported for 3D geometries yet.
+- `winding_number_factor`: For leaky geometries, a factor of `0.4` will give a better inside-outside segmentation.
+
+!!! warning "Experimental Implementation"
+    This is an experimental feature and may change in any future releases.
 """
 struct WindingNumberJacobson{ELTYPE}
     winding_number_factor :: ELTYPE
     winding               :: Union{NaiveWinding, HierarchicalWinding}
 
-    function WindingNumberJacobson(; shape=nothing, winding_number_factor=sqrt(eps()),
+    function WindingNumberJacobson(; geometry=nothing, winding_number_factor=sqrt(eps()),
                                    hierarchical_winding=false)
-        if hierarchical_winding && shape isa Nothing
-            throw(ArgumentError("`shape` must be of type `Shapes` when using hierarchical winding"))
+        if hierarchical_winding && geometry isa Nothing
+            throw(ArgumentError("`geometry` must be of type `Polygon` (2D) or `TriangleMesh` (3D) when using hierarchical winding"))
         end
 
-        winding = hierarchical_winding ? HierarchicalWinding(shape) : NaiveWinding()
+        winding = hierarchical_winding ? HierarchicalWinding(geometry) : NaiveWinding()
 
         return new{typeof(winding_number_factor)}(winding_number_factor, winding)
     end
@@ -67,9 +70,7 @@ function (point_in_poly::WindingNumberJacobson)(mesh::TriangleMesh{3}, points;
     @threaded points for query_point in axes(points, 2)
         p = point_position(points, mesh, query_point)
 
[…]
