+++ examples/preprocessing/complex_shape_2d.jl
@@ -6,9 +6,9 @@ file = "hexagon"
 filename = joinpath("examples", "preprocessing", file * ".asc")
 
 # Returns `Shape`
-shape = load_shape(filename)
+geomtery = load_geomtery(filename)
 
 # Returns `InitialCondition`.
-shape_sampled = ComplexShape(shape; particle_spacing, density=1.0)
+shape_sampled = ComplexShape(geometry; particle_spacing, density=1.0)
 
 trixi2vtk(shape_sampled)
