+++ src/preprocessing/geometries/geometries.jl
@@ -1,7 +1,7 @@
-include("polygon_shape.jl")
+include("polygon.jl")
 include("triangle_mesh.jl")
 include("io.jl")
 
 @inline eachface(mesh) = Base.OneTo(nfaces(mesh))
 
-@inline point_position(A, shape, i) = extract_svector(A, Val(ndims(shape)), i)
+@inline point_position(A, geometry, i) = extract_svector(A, Val(ndims(geometry)), i)
