### 2024-07-23 608a699
implement suggestions
