+++ test/count_allocations.jl
@@ -17,20 +17,19 @@ function copy_semi_with_no_update_nhs(semi)
 end
 
 # Forward `foreach_neighbor` to wrapped neighborhood search
-@inline function TrixiParticles.PointNeighbors.foreach_neighbor(f, system_coords,
-                                                                neighbor_coords,
-                                                                neighborhood_search::NoUpdateNeighborhoodSearch,
-                                                                particle;
-                                                                search_radius=PointNeighbors.search_radius(neighborhood_search.nhs))
-    TrixiParticles.PointNeighbors.foreach_neighbor(f, system_coords, neighbor_coords,
-                                                   neighborhood_search.nhs, particle,
-                                                   search_radius=search_radius)
+@inline function PointNeighbors.foreach_neighbor(f, system_coords,
+                                                 neighbor_coords,
+                                                 neighborhood_search::NoUpdateNeighborhoodSearch,
+                                                 particle;
+                                                 search_radius=PointNeighbors.search_radius(neighborhood_search.nhs))
+    PointNeighbors.foreach_neighbor(f, system_coords, neighbor_coords,
+                                    neighborhood_search.nhs, particle,
+                                    search_radius=search_radius)
 end
 
 # No update
-@inline function TrixiParticles.PointNeighbors.update!(search::NoUpdateNeighborhoodSearch,
-                                                       x, y;
-                                                       points_moving=(true, true))
+@inline function PointNeighbors.update!(search::NoUpdateNeighborhoodSearch, x, y;
+                                        points_moving=(true, true))
     return search
 end
 
