+++ src/schemes/boundary/open_boundary/mirroring.jl
@@ -11,14 +11,14 @@
 struct BoundaryModelTafuni end
 
 function update_quantities!(system, ::BoundaryModelTafuni, v, u, v_ode, u_ode, semi, t)
-    @trixi_timeit timer() "interpolate and correct values" begin
-        interpolate_values!(system, v_ode, u_ode, semi, t; system.cache...)
+    @trixi_timeit timer() "extrapolate and correct values" begin
+        extrapolate_values!(system, v_ode, u_ode, semi, t; system.cache...)
     end
 end
 
 update_final!(system, ::BoundaryModelTafuni, v, u, v_ode, u_ode, semi, t) = system
 
-function interpolate_values!(system, v_ode, u_ode, semi, t; prescribed_density=false,
+function extrapolate_values!(system, v_ode, u_ode, semi, t; prescribed_density=false,
                              prescribed_pressure=false, prescribed_velocity=false)
     (; pressure, density, fluid_system, boundary_zone, reference_density,
     reference_velocity, reference_pressure) = system
@@ -27,31 +27,27 @@ function interpolate_values!(system, v_ode, u_ode, semi, t; prescribed_density=f
 
     svector_(f, system) = SVector(ntuple(@inline(dim->f[dim + 1]), ndims(system)))
 
-    zero_svector(system) = SVector(ntuple(@inline(dim->zero(eltype(system))),
-                                          ndims(system) + 1))
-    zero_smatrix(system) = SMatrix{ndims(system) + 1,
-                                   ndims(system) + 1}(ntuple(@inline(i->zero(eltype(system))),
-                                                             Val((ndims(system) + 1)^2)))
-
     v_open_boundary = wrap_v(v_ode, system, semi)
     v_fluid = wrap_v(v_ode, fluid_system, semi)
     u_open_boundary = wrap_u(u_ode, system, semi)
     u_fluid = wrap_u(u_ode, fluid_system, semi)
 
-    neighborhood_search = get_neighborhood_search(system, fluid_system, semi)
+    # Use the fluid-fluid nhs, since the boundary particles are mirrored into the fluid domain
+    neighborhood_search = get_neighborhood_search(fluid_system, fluid_system, semi)
 
-    interpolated_velocity = [zero_svector(system) for _ in 1:ndims(system)]
+    # interpolated_velocity = [zero_svector(system) for _ in 1:ndims(system)]
 
     for particle in each_moving_particle(system)
         particle_coords = current_coords(u_open_boundary, system, particle)
         ghost_node_position = mirror_position(particle_coords, boundary_zone)
 
         # Set zero
-        correction_matrix = zero_smatrix(system)
-        interpolated_pressure = zero_svector(system)
-        @inbounds for dim in 1:ndims(system)
-            interpolated_velocity[dim] = zero_svector(system)
-        end
+        correction_matrix = zero(SMatrix{ndims(system) + 1, ndims(system) + 1,
+                                         eltype(system)})
+        interpolated_pressure = zero(SVector{ndims(system) + 1, eltype(system)})
+
+        interpolated_velocity = zero(SMatrix{ndims(system), ndims(system) + 1,
+                                             eltype(system)})
 
         for neighbor in PointNeighbors.eachneighbor(ghost_node_position,
                                                     neighborhood_search)
@@ -67,6 +63,10 @@ function interpolate_values!(system, v_ode, u_ode, semi, t; prescribed_density=f
                 pressure_b = particle_pressure(v_fluid, fluid_system, neighbor)
                 v_b = current_velocity(v_fluid, fluid_system, neighbor)
 
+                # Project `v_b` to the normal direction of the boundary zone
+                # See https://doi.org/10.1016/j.jcp.2020.110029
[…]
