### 2024-07-24 7c1db9b
implement suggestions
