+++ src/preprocessing/point_in_poly/winding_number_horman.jl
@@ -22,14 +22,11 @@ function (point_in_poly::WindingNumberHorman)(geometry, points; store_winding_nu
     winding_numbers = Float64[]
     store_winding_number && (winding_numbers = resize!(winding_numbers, length(inpoly)))
 
-    h_unit = SVector(1.0, 0.0)
-    v_unit = SVector(0.0, 1.0)
-
-    function quadrant(dot_v, dot_h)
-        (dot_v >= 0.0 && dot_h > 0.0) && return 0
-        (dot_v > 0.0 && dot_h <= 0.0) && return 1
-        (dot_v <= 0.0 && dot_h < 0.0) && return 2
-        (dot_v < 0.0 && dot_h >= 0.0) && return 3
+    function quadrant(v, p)
+        (v[1] > p[1] && v[2] >= p[2]) && return 0
+        (v[1] <= p[1] && v[2] > p[2]) && return 1
+        (v[1] < p[1] && v[2] <= p[2]) && return 2
+        (v[1] >= p[1] && v[2] < p[2]) && return 3
     end
 
     @threaded points for query_point in axes(points, 2)
@@ -40,44 +37,31 @@ function (point_in_poly::WindingNumberHorman)(geometry, points; store_winding_nu
             v1 = edge_vertices[edge][1]
             v2 = edge_vertices[edge][2]
 
-            direction1 = v1 - v_query
-            direction2 = v2 - v_query
+            # because 0 <= `quadrant_numbers` <= 3 we know that -3 <= `quarter_angle` <= 3
+            quarter_angle = quadrant(v2, v_query) - quadrant(v1, v_query)
 
-            dot_v1 = dot(direction1, v_unit)
-            dot_h1 = dot(direction1, h_unit)
-            dot_v2 = dot(direction2, v_unit)
-            dot_h2 = dot(direction2, h_unit)
+            direction_1 = v1 - v_query
+            direction_2 = v2 - v_query
 
-            # because 0 <= `quadrant_numbers` <= 3 we know that -3 <= `quarter_angel` <= 3
-            quarter_angel = quadrant(dot_v2, dot_h2) - quadrant(dot_v1, dot_h1)
-
-            if quarter_angel == -3
+            # Check the orientation of the triangle Δ(`v_query`, `v1`, `v2`) by finding the
+            # sign of the determinant.
+            # `+`: counter clockwise
+            # `-`: clockwise
+            if quarter_angle == -3
                 winding_number += 1
-            elseif quarter_angel == 3
+            elseif quarter_angle == 3
                 winding_number -= 1
-            elseif quarter_angel == -2 && positive_determinant(v1, v2, v_query)
+            elseif quarter_angle == -2 && det([direction_1 direction_2]) > 0
                 winding_number += 1
-            elseif quarter_angel == 2 && !positive_determinant(v1, v2, v_query)
+            elseif quarter_angle == 2 && det([direction_1 direction_2]) < 0
                 winding_number -= 1
             end
         end
 
         store_winding_number && (winding_numbers[query_point] = winding_number)
 
-        !(winding_number == 0) && (inpoly[query_point] = true)
+        winding_number != 0 && (inpoly[query_point] = true)
     end
 
     return inpoly, winding_numbers
-end
-
-function positive_determinant(v1, v2, v_query)
-
-    # Check the orientation of the triangle Δ(`v_query`, `v1`, `v2`) by finding the
-    # sign of the determinant.
-    # `+`: counter clockwise
-    # `-`: clockwise
-    positive_sign = (v1[1] - v_query[1]) * (v2[2] - v_query[2]) -
-                    (v2[1] - v_query[1]) * (v1[2] - v_query[2]) > 0.0
-
-    return positive_sign
 end
