+++ examples/preprocessing/complex_shape_3d.jl
@@ -9,7 +9,6 @@ filename = joinpath("examples", "preprocessing", file * ".stl")
 # For more robustness, use `winding_number_factor=0.4`.
 # filename = joinpath("examples", "preprocessing", "drive_gear.stl")
 
-# Returns `Shape`
 geometry = load_geometry(filename)
 
 point_in_geometry_algorithm = WindingNumberJacobson(; geometry,
