+++ examples/preprocessing/complex_shape_2d.jl
@@ -5,8 +5,9 @@ particle_spacing = 0.05
 file = "hexagon"
 filename = joinpath("examples", "preprocessing", file * ".asc")
 
-# Returns `Shape`
 geometry = load_geometry(filename)
+
+trixi2vtk(geometry)
 
 # Returns `InitialCondition`.
 shape_sampled = ComplexShape(geometry; particle_spacing, density=1.0)
