### 2024-07-24 c01d9ad
Make `SolutionSavingCallback` work on GPUs (#579)
* Make `SolutionSavingCallback` work on GPUs

* Avoid overhead on CPUs

---------

Co-authored-by: Sven Berger <berger.sven@gmail.com>
Co-authored-by: Niklas Neher <73897120+LasNikas@users.noreply.github.com>
