### 2024-07-24 f97e9d9
sum-init with `zero(eltype))`
