### 2024-07-24 6735f4d
add static indices slicing
