### 2024-07-24 5729c48
Make Adami loop flipping optimization optional
