### 2024-07-24 ab18cfb
Make Adami loop flipping optimization optional
