### 2024-07-25 2f0daf1
remove example file
