+++ src/TrixiParticles.jl
@@ -65,7 +65,8 @@ export ArtificialViscosityMonaghan, ViscosityAdami, ViscosityMorris
 export DensityDiffusion, DensityDiffusionMolteniColagrossi, DensityDiffusionFerrari,
        DensityDiffusionAntuono
 export BoundaryModelMonaghanKajtar, BoundaryModelDummyParticles, AdamiPressureExtrapolation,
-       PressureMirroring, PressureZeroing, BoundaryModelLastiwka, BernoulliPressureExtrapolation
+       PressureMirroring, PressureZeroing, BoundaryModelLastiwka,
+       BernoulliPressureExtrapolation
 
 export BoundaryMovement
 export examples_dir, validation_dir, trixi_include
