+++ src/schemes/boundary/dummy_particles/dummy_particles.jl
@@ -332,8 +332,9 @@ end
 end
 
 function compute_pressure!(boundary_model,
-                           bnd_density_calc::Union{AdamiPressureExtrapolation,
-                                                   BernoulliPressureExtrapolation}, system,
+                           boundary_density_calc::Union{AdamiPressureExtrapolation,
+                                                        BernoulliPressureExtrapolation},
+                           system,
                            v, u,
                            v_ode, u_ode, semi)
     (; pressure, cache, viscosity) = boundary_model
@@ -362,18 +363,18 @@ function compute_pressure!(boundary_model,
             nhs = get_neighborhood_search(neighbor_system, system, semi)
 
             # Loop over fluid particles and then the neighboring boundary particles to extrapolate fluid pressure to the boundaries
-            bnd_pressure_extrapolation_neighbor!(boundary_model, system,
-                                                 neighbor_system,
-                                                 system_coords, neighbor_coords, v,
-                                                 v_neighbor_system, nhs)
+            boundary_pressure_extrapolation_neighbor!(boundary_model, system,
+                                                      neighbor_system,
+                                                      system_coords, neighbor_coords, v,
+                                                      v_neighbor_system, nhs)
         else
             nhs = get_neighborhood_search(system, neighbor_system, semi)
 
             # Loop over boundary particles and then the neighboring fluid particles to extrapolate fluid pressure to the boundaries
-            bnd_pressure_extrapolation!(boundary_model, system,
-                                        neighbor_system,
-                                        system_coords, neighbor_coords, v,
-                                        v_neighbor_system, nhs)
+            boundary_pressure_extrapolation!(boundary_model, system,
+                                             neighbor_system,
+                                             system_coords, neighbor_coords, v,
+                                             v_neighbor_system, nhs)
         end
 
         @threaded system for particle in eachparticle(system)
@@ -415,35 +416,20 @@ function compute_pressure!(boundary_model, ::Union{PressureMirroring, PressureZe
     return boundary_model
 end
 
-@inline function dynamic_pressure(density_neighbor, v, v_neighbor_system, particle,
-                                  neighbor, system::BoundarySystem)
-    if system.ismoving[]
-        return 0.5 * density_neighbor *
-               norm(current_velocity(v, system, particle) -
-                    v_neighbor_system[1:ndims(system), neighbor])^2
-    end
-    return 0.0
-end
-
-@inline function dynamic_pressure(density_neighbor, v, v_neighbor_system, particle,
-                                  neighbor, system::SolidSystem)
-    return 0.5 * density_neighbor *
-           norm(current_velocity(v, system, particle) -
-                v_neighbor_system[1:ndims(system), neighbor])^2
-end
-
-@inline function bnd_pressure_extrapolation_neighbor!(boundary_model, system,
-                                                      neighbor_system, system_coords,
-                                                      neighbor_coords, v, v_neighbor_system,
-                                                      neighborhood_search)
+@inline function boundary_pressure_extrapolation_neighbor!(boundary_model, system,
+                                                           neighbor_system, system_coords,
+                                                           neighbor_coords, v,
+                                                           v_neighbor_system,
+                                                           neighborhood_search)
     return boundary_model
 end
 
-@inline function bnd_pressure_extrapolation_neighbor!(boundary_model, system,
-                                                      neighbor_system::FluidSystem,
-                                                      system_coords, neighbor_coords, v,
-                                                      v_neighbor_system,
-                                                      neighborhood_search)
+@inline function boundary_pressure_extrapolation_neighbor!(boundary_model, system,
[…]
