+++ test/schemes/boundary/dummy_particles/dummy_particles.jl
@@ -55,12 +55,13 @@
             TrixiParticles.reset_cache!(boundary_system.boundary_model.cache,
                                         boundary_system.boundary_model.viscosity)
             TrixiParticles.boundary_pressure_extrapolation!(boundary_model,
-                                                       boundary_system,
-                                                       fluid_system, boundary.coordinates,
-                                                       fluid.coordinates, v_fluid,
-                                                       v_fluid .*
-                                                       ones(size(fluid.coordinates)),
-                                                       neighborhood_search)
+                                                            boundary_system,
+                                                            fluid_system,
+                                                            boundary.coordinates,
+                                                            fluid.coordinates, v_fluid,
+                                                            v_fluid .*
+                                                            ones(size(fluid.coordinates)),
+                                                            neighborhood_search)
 
             for particle in TrixiParticles.eachparticle(boundary_system)
                 if volume[particle] > eps()
@@ -94,10 +95,11 @@
             TrixiParticles.reset_cache!(boundary_system.boundary_model.cache,
                                         boundary_system.boundary_model.viscosity)
             TrixiParticles.boundary_pressure_extrapolation!(boundary_model, boundary_system,
-                                                       fluid_system, boundary.coordinates,
-                                                       fluid.coordinates, v_fluid,
-                                                       v_fluid,
-                                                       neighborhood_search)
+                                                            fluid_system,
+                                                            boundary.coordinates,
+                                                            fluid.coordinates, v_fluid,
+                                                            v_fluid,
+                                                            neighborhood_search)
 
             for particle in TrixiParticles.eachparticle(boundary_system)
                 if volume[particle] > eps()
