+++ src/general/system.jl
@@ -148,14 +148,16 @@ end
 # Only for systems requiring a mandatory callback
 reset_callback_flag!(system) = system
 
-function neighbor_number(::Val{D}, initial_particle_spacing, smoothing_length) where {D}
+# Assuming a constant particle spacing one can calculate the number of neighbors within the
+# compact support for an undisturbed particle distribution.
+function neighbor_number(::Val{D}, particle_spacing, compact_support) where {D}
     throw(ArgumentError("Unsupported dimension: $D"))
 end
 
-@inline function neighbor_number(::Val{2}, initial_particle_spacing, smoothing_length)
-    return pi * smoothing_length^2 / initial_particle_spacing^2
+@inline function neighbor_number(::Val{2}, particle_spacing, compact_support)
+    return pi * compact_support^2 / particle_spacing^2
 end
 
-@inline function neighbor_number(::Val{3}, initial_particle_spacing, smoothing_length)
-    return 4.0 / 3.0 * pi * smoothing_length^3 / initial_particle_spacing^3
+@inline @fastpow function neighbor_number(::Val{3}, particle_spacing, compact_support)
+    return 4.0 / 3.0 * pi * compact_support^3 / particle_spacing^3
 end
