+++ src/schemes/fluid/entropically_damped_sph/system.jl
@@ -90,7 +90,7 @@ struct EntropicallyDampedSPHSystem{NDIMS, ELTYPE <: Real, IC, M, DC, K, V,
         end
 
         if surface_tension !== nothing && surface_normal_method === nothing
-            surface_normal_method = AkinciSurfaceNormal(smoothing_kernel, smoothing_length)
+            surface_normal_method = ColorfieldSurfaceNormal(smoothing_kernel, smoothing_length)
         end
 
         pressure_acceleration = choose_pressure_acceleration_formulation(pressure_acceleration,
