### 2024-07-26 #6961 @HaoZeke
Sorry for being slow on the uptake @LasNikas.

Thanks for the transparency about the potential conflicts, @williamfgc and @giordano. I think they don't warrant an exclusion from the review in this case. I think the timelines would also be reasonable (within our 6 week guideline) so I'll assign y'all (@williamfgc, @giordano, @luraess) as reviewers.
