### 2024-07-26 #7044 @editorialbot
[REVIEW]: TrixiParticles.jl: Particle-based multiphysics simulation in Julia
**Submitting author:** @LasNikas (<a href="http://orcid.org/0009-0004-2472-0923">Niklas Neher</a>)
**Repository:** https://github.com/trixi-framework/TrixiParticles.jl
**Branch with paper.md** (empty if default branch): paper-2024-joss
**Version:** v0.2.5
