### 2024-07-27 50b8592
implement suggestions
