### 2024-07-27 71eb74e
construct hierarchy in constructor
