+++ src/preprocessing/point_in_poly/hierarchical_winding.jl
@@ -5,16 +5,47 @@ struct BoundingBoxTree{MC}
     faces         :: Vector{Int}
     min_corner    :: MC
     max_corner    :: MC
-    is_leaf       :: Ref{Bool}
+    is_leaf       :: Bool
     closing_faces :: Vector{NTuple{3, Int}}
-    children      :: Vector{BoundingBoxTree}
+    child_left    :: BoundingBoxTree
+    child_right   :: BoundingBoxTree
 
-    function BoundingBoxTree(faces, min_corner, max_corner)
+    function BoundingBoxTree(geometry, faces, directed_edges, min_corner, max_corner)
         closing_faces = Vector{NTuple{3, Int}}()
-        children = Vector{BoundingBoxTree}()
 
-        return new{typeof(min_corner)}(faces, min_corner, max_corner, Ref(false),
-                                       closing_faces, children)
+        if length(faces) < 100
+            return new{typeof(min_corner)}(faces, min_corner, max_corner, true,
+                                           closing_faces)
+        end
+
+        determine_closure!(closing_faces, min_corner, max_corner, geometry, faces,
+                           directed_edges)
+
+        if length(closing_faces) >= length(faces)
+            return new{typeof(min_corner)}(faces, min_corner, max_corner, true,
+                                           closing_faces)
+        end
+
+        # Bisect the box splitting its longest side
+        box_edges = max_corner - min_corner
+
+        split_direction = argmax(box_edges)
+
+        uvec = (1:3) .== split_direction
+
+        max_corner_left = max_corner - 0.5box_edges[split_direction] * uvec
+        min_corner_right = min_corner + 0.5box_edges[split_direction] * uvec
+
+        faces_left = in_bounding_box(geometry, faces, min_corner, max_corner_left)
+        faces_right = in_bounding_box(geometry, faces, min_corner_right, max_corner)
+
+        child_left = BoundingBoxTree(geometry, faces_left, directed_edges,
+                                     min_corner, max_corner_left)
+        child_right = BoundingBoxTree(geometry, faces_right, directed_edges,
+                                      min_corner_right, max_corner)
+
+        return new{typeof(min_corner)}(faces, min_corner, max_corner, false, closing_faces,
+                                       child_left, child_right)
     end
 end
 
@@ -22,12 +53,13 @@ struct HierarchicalWinding{BB}
     bounding_box::BB
 
     function HierarchicalWinding(geometry)
-        bounding_box = BoundingBoxTree(eachface(geometry), geometry.min_corner,
-                                       geometry.max_corner)
+        min_corner = geometry.min_corner .- sqrt(eps())
+        max_corner = geometry.max_corner .+ sqrt(eps())
 
         directed_edges = zeros(Int, length(geometry.edge_normals))
 
-        construct_hierarchy!(bounding_box, geometry, directed_edges)
+        bounding_box = BoundingBoxTree(geometry, eachface(geometry), directed_edges,
+                                       min_corner, max_corner)
 
         return new{typeof(bounding_box)}(bounding_box)
     end
@@ -42,7 +74,7 @@ end
 function hierarchical_winding(bounding_box, mesh, query_point)
     (; min_corner, max_corner) = bounding_box
 
-    if bounding_box.is_leaf[]
+    if bounding_box.is_leaf
         return naive_winding(mesh, bounding_box.faces, query_point)
 
     elseif !in_bounding_box(query_point, min_corner, max_corner)
@@ -50,60 +82,18 @@ function hierarchical_winding(bounding_box, mesh, query_point)
[…]
