### 2024-07-27 93b8444
use ASCII characters
