+++ src/preprocessing/point_in_poly/winding_number_jacobson.jl
@@ -1,12 +1,29 @@
 struct NaiveWinding end
 
-@inline function (winding::NaiveWinding)(mesh, query_point)
+@inline function (winding::NaiveWinding)(polygon::Polygon{2}, query_point)
+    (; edge_vertices) = polygon
+
+    return naive_winding(polygon, edge_vertices, query_point)
+end
+
+@inline function (winding::NaiveWinding)(mesh::TriangleMesh{3}, query_point)
     (; face_vertices) = mesh
 
     return naive_winding(mesh, face_vertices, query_point)
 end
 
-@inline function naive_winding(mesh, faces, query_point)
+@inline function naive_winding(polygon::Polygon{2}, edges, query_point)
+    winding_number = sum(edges, init=zero(eltype(polygon))) do edge
+        a = edge[1] - query_point
+        b = edge[2] - query_point
+
+        return atan(det([a b]), (dot(a, b)))
+    end
+
+    return winding_number
+end
+
+@inline function naive_winding(mesh::TriangleMesh{3}, faces, query_point)
     winding_number = sum(faces, init=zero(eltype(mesh))) do face
 
         # Eq. 6 of Jacobsen et al. Based on A. Van Oosterom (1983),
@@ -59,8 +76,7 @@ struct WindingNumberJacobson{ELTYPE, W}
     end
 end
 
-function (point_in_poly::WindingNumberJacobson)(mesh::TriangleMesh{3}, points;
-                                                store_winding_number=false)
+function (point_in_poly::WindingNumberJacobson)(geometry, points; store_winding_number=false)
     (; winding_number_factor, winding) = point_in_poly
 
     inpoly = falses(size(points, 2))
@@ -68,43 +84,12 @@ function (point_in_poly::WindingNumberJacobson)(mesh::TriangleMesh{3}, points;
     winding_numbers = Float64[]
     store_winding_number && (winding_numbers = resize!(winding_numbers, length(inpoly)))
 
-    @threaded points for query_point in axes(points, 2)
-        p = point_position(points, mesh, query_point)
-
-        winding_number = winding(mesh, p) / 4pi
-
-        store_winding_number && (winding_numbers[query_point] = winding_number)
-
-        # Relaxed restriction of `(winding_number != 0.0)`
-        if !(-winding_number_factor < winding_number < winding_number_factor)
-            inpoly[query_point] = true
-        end
-    end
-
-    return inpoly, winding_numbers
-end
-
-function (point_in_poly::WindingNumberJacobson)(mesh::Polygon{2}, points;
-                                                store_winding_number=false)
-    (; winding_number_factor) = point_in_poly
-    (; edge_vertices) = mesh
-
-    inpoly = falses(size(points, 2))
-
-    winding_numbers = Float64[]
-    store_winding_number && (winding_numbers = resize!(winding_numbers, length(inpoly)))
+    divisor = ndims(geometry) == 2 ? 2pi : 4pi
 
     @threaded points for query_point in axes(points, 2)
-        p = point_position(points, mesh, query_point)
+        p = point_position(points, geometry, query_point)
 
-        winding_number = sum(edge_vertices, init=zero(eltype(mesh))) do edge
-            a = edge[1] - p
-            b = edge[2] - p
-
-            return atan(det([a b]), (dot(a, b)))
-        end
-
-        winding_number /= 2pi
+        winding_number = winding(geometry, p) / divisor
 
         store_winding_number && (winding_numbers[query_point] = winding_number)
 
