+++ src/preprocessing/geometries/polygon.jl
@@ -16,11 +16,6 @@ struct Polygon{NDIMS, ELTYPE}
 
     # Function barrier to make `NDIMS` static and therefore `SVector`s type-stable
     function Polygon{NDIMS}(vertices_) where {NDIMS}
-        # Close the polygon if it's open
-        if !isapprox(vertices_[:, end], vertices_[:, 1])
-            vertices_ = hcat(vertices_, vertices_[:, 1])
-        end
-
         n_vertices = size(vertices_, 2)
         ELTYPE = eltype(vertices_)
 
@@ -64,7 +59,7 @@ struct Polygon{NDIMS, ELTYPE}
             edge_normal = SVector{NDIMS}(normalize([-edge[2], edge[1]]))
 
             push!(edge_vertices, (v1, v2))
-            push!(edge_vertices_ids, (i, mod(i, n_vertices - 1) + 1))
+            push!(edge_vertices_ids, (i, i+1))
             push!(edge_normals, edge_normal)
         end
 
@@ -96,8 +91,6 @@ struct Polygon{NDIMS, ELTYPE}
 
             push!(vertex_normals, (vortex_normal_1, vortex_normal_2))
         end
-
-        vertices = vertices[1:(end - 1)]
 
         return new{NDIMS, ELTYPE}(vertices, edge_vertices, vertex_normals, edge_normals,
                                   edge_vertices_ids, min_corner, max_corner)
