+++ examples/preprocessing/complex_shape_2d.jl
@@ -2,10 +2,10 @@ using TrixiParticles
 
 particle_spacing = 0.05
 
-file = "hexagon"
-filename = joinpath("examples", "preprocessing", file * ".asc")
+filename = "hexagon"
+file = joinpath("examples", "preprocessing", "data", filename * ".asc")
 
-geometry = load_geometry(filename)
+geometry = load_geometry(file)
 
 trixi2vtk(geometry)
 
@@ -13,7 +13,7 @@ point_in_geometry_algorithm = WindingNumberJacobson(; geometry,
                                                     #winding_number_factor=0.4,
                                                     hierarchical_winding=true)
 
-# Returns `InitialCondition`.
+# Returns `InitialCondition`
 shape_sampled = ComplexShape(geometry; particle_spacing, density=1.0,
                              store_winding_number=true,
                              point_in_geometry_algorithm)
