+++ examples/preprocessing/complex_shape_3d.jl
@@ -2,20 +2,16 @@ using TrixiParticles
 
 particle_spacing = 0.05
 
-file = "sphere"
-filename = joinpath("examples", "preprocessing", file * ".stl")
+filename = "sphere"
+file = joinpath("examples", "preprocessing", "data", filename * ".stl")
 
-# The following triangle mesh is corrupt.
-# For more robustness, use `winding_number_factor=0.4`.
-# filename = joinpath("examples", "preprocessing", "drive_gear.stl")
-
-geometry = load_geometry(filename)
+geometry = load_geometry(file)
 
 point_in_geometry_algorithm = WindingNumberJacobson(; geometry,
                                                     # winding_number_factor=0.4,
                                                     hierarchical_winding=true)
 
-# Returns `InitialCondition`.
+# Returns `InitialCondition`
 shape_sampled = ComplexShape(geometry; particle_spacing, density=1.0,
                              point_in_geometry_algorithm)
 

[…]
