+++ src/preprocessing/geometries/polygon.jl
@@ -59,7 +59,7 @@ struct Polygon{NDIMS, ELTYPE}
             edge_normal = SVector{NDIMS}(normalize([-edge[2], edge[1]]))
 
             push!(edge_vertices, (v1, v2))
-            push!(edge_vertices_ids, (i, i+1))
+            push!(edge_vertices_ids, (i, i + 1))
             push!(edge_normals, edge_normal)
         end
 
