+++ src/visualization/recipes_plots.jl
@@ -67,3 +67,53 @@ RecipesBase.@recipe function f(v_ode, u_ode, semi::Semidiscretization;
         end
     end
 end
+
+RecipesBase.@recipe function f(ic::InitialCondition; zcolor=nothing,
+                               size=(600, 400), label=false, colorbar_title="")
+    (; coordinates, particle_spacing) = ic
+
+    x = collect(coordinates[1, :])
+    y = collect(coordinates[2, :])
+
+    if particle_spacing < 0
+        particle_spacing = 0.0
+    end
+
+    x_min, y_min = minimum(coordinates, dims=2) .- 0.5particle_spacing
+    x_max, y_max = maximum(coordinates, dims=2) .+ 0.5particle_spacing
+
+    # Note that this assumes the plot area to be ~10% smaller than `size`,
+    # which is the case when showing a single plot with the legend inside.
+    # With the legend outside, this is no longer the case, so the `markersize` has to be
+    # set manually.
+    pixel_size = max((x_max - x_min) / size[1], (y_max - y_min) / size[2])
+
+    xlims --> (x_min, x_max)
+    ylims --> (y_min, y_max)
+    aspect_ratio --> :equal
+
+    seriestype --> :scatter
+    markerstrokewidth --> 0
+    grid --> false
+
+    @series begin
+        if particle_spacing < eps()
+            # Fall back to 1px marker radius
+            markersize --> 1
+        else
+            pixels_per_particle = particle_spacing / pixel_size
+
+            # Marker radius in pixels
+            markersize --> 0.5 * pixels_per_particle
+        end
+
+        label --> label
+
+        colorbar_title --> colorbar_title
+
+        zcolor --> zcolor
+
+        # Return data for plotting
+        x, y
+    end
+end
