+++ src/preprocessing/point_in_poly/winding_number_horman.jl
@@ -4,8 +4,6 @@
 Algorithm for inside-outside segmentation of a complex geometry proposed by [Horman et al. (2001)](@ref references_complex_shape).
 It is only supported for 2D geometries.
 [`WindingNumberHorman`](@ref) might handle edge cases a bit better, since the winding number is an integer value.
-Also, it is faster than [`WindingNumberJacobson`](@ref) for 2D geometries with more than about 100 edges.
-
 
 !!! warning "Experimental Implementation"
     This is an experimental feature and may change in any future releases.
