+++ docs/src/preprocessing/preprocessing.md
@@ -1,15 +1,147 @@
-```@autodocs
-Modules = [TrixiParticles]
-Pages = [joinpath("preprocessing", "point_in_poly", "point_in_poly.jl")]
-```
-```@autodocs
-Modules = [TrixiParticles]
-Pages = [joinpath("preprocessing", "geometries", "geometries.jl")]
-```
-# Hierarchical Winding
+# Sample Geometry
+Generating the initial configuration of a simulation requires filling volumes (3D) or surfaces (2D) of complex geometries with particles.
+The algorithm to sample a complex geometry should be robust and fast,
+since for large problems (high number of particles) or complex geometries (many geometry-faces),
+generating the initial configuration is not trivial and can be very expensive in terms of computational costs.
+We therefore use a winding number approach for an inside-outside segmentation of an object.
+
+We provide the following methods to calculate the [winding number](https://en.wikipedia.org/wiki/Winding_number):
+- Horman et al. (2001) evaluated the winding number combined with an even-odd rule and is only for 2D polygons (see [WindingNumberHorman](@ref)).
+- Naive winding: Jacobson et al. (2013) generalized the winding number so that the algorithm can be applied for 2D and also for 3D geometries (see [WindingNumberJacobson](@ref)).
+- Hierarchical winding: Jacobson et al. (2013) also introduced a fast hierarchical evaluation of the winding number. For further information see description below.
 
 
-## Bounding Volume Hierarchy
+## Hierarchical Winding
+According to Jacobson et al. (2013) the winding number is the sum of harmonic functions defined for each edge (2D) or face (3D).
+We can show this with the following example in which we determine the winding number for each edge of a triangle separately and sum them up (TODO: insert code):
+
+$w = w_1 + w_2 + w_3$
+
+![alt text](image.png)
+
+That means, if we want to calculate the winding number only for a specific area of interest, e.g. the top right square of a circle, we can use the closure of this area as we can see in the following image.
+
+![alt text](image-2.png)
+
+For a continuous problem, we then have an open surface $\mathcal{S}$ and an arbitrary closing surface $\bar{\mathcal{S}}$, where $\partial \bar{\mathcal{S}} = \partial \mathcal{S}$,
+such that $\mathcal{B} = \bar{\mathcal{S}} \cup \mathcal{S}$ is some closed oriented surface.
+If an arbitrary query point $\mathbf{p}$ is outside the convex hull of $\mathcal{B}$, we know that
+
+$w_{\mathcal{S}}(\mathbf{p}) + w_{\bar{\mathcal{S}}}(\mathbf{p}) = w_{\mathcal{B}}(\mathbf{p}) = 0.$
+
+This means
+
+$w_{\mathcal{S}}(\mathbf{p}) = - w_{\bar{\mathcal{S}}}(\mathbf{p}) ,$
+
+regardless of how $\bar{\mathcal{S}}$ is constructed.
+
+If we construct a bounding volume hierarchy $T$ and evaluate $w$ hierarchical,
+we can perform asymptotically better than the naive winding computing $w$ for the whole geometry as is shown by Jacobson et al. (2013).
+
+The algorithm behind (Jacobson et al., Algorithm 2, p. 5) calls itself recursively, where the recursion stops with the following criteria.
+
+- if $T$ is a leaf then evaluate the winding number naively $w(\mathbf{p}, T.\mathcal{S})$ where $T.\mathcal{S}$ is the open surface in $T$.
[…]
