+++ src/schemes/fluid/surface_normal_sph.jl
@@ -7,11 +7,26 @@ function ColorfieldSurfaceNormal(; smoothing_kernel, smoothing_length)
     return ColorfieldSurfaceNormal(smoothing_kernel, smoothing_length)
 end
 
+function create_cache_surface_normal(surface_normal_method, ELTYPE, NDIMS, nparticles)
+    return (;)
+end
+
+function create_cache_surface_normal(::ColorfieldSurfaceNormal, ELTYPE, NDIMS, nparticles)
+    surface_normal = Array{ELTYPE, 2}(undef, NDIMS, nparticles)
+    neighbor_count = Array{ELTYPE, 1}(undef, nparticles)
+    return (; surface_normal, neighbor_count)
+end
+
+function calc_normal!(system, neighbor_system, u_system, v, v_neighbor_system,
+                      u_neighbor_system, semi, surfn)
+    # Normal not needed
+    return system
+end
+
 # Section 2.2 in Akinci et al. 2013 "Versatile Surface Tension and Adhesion for SPH Fluids"
-# Note: This is the simplest form of normal approximation commonly used in SPH and comes
-# with serious deficits in accuracy especially at corners, small neighborhoods and boundaries
-function calc_normal_akinci!(system, neighbor_system::FluidSystem, u_system, v,
-                             v_neighbor_system, u_neighbor_system, semi, surfn)
+# and Section 5 in Morris 2000 "Simulating surface tension with smoothed particle hydrodynamics"
+function calc_normal!(system::FluidSystem, neighbor_system::FluidSystem, u_system, v,
+                      v_neighbor_system, u_neighbor_system, semi, surfn)
     (; cache) = system
     (; smoothing_kernel, smoothing_length) = surfn
 
@@ -37,9 +52,8 @@ function calc_normal_akinci!(system, neighbor_system::FluidSystem, u_system, v,
         grad_kernel = kernel_grad(smoothing_kernel, pos_diff, distance, smoothing_length)
         for i in 1:ndims(system)
             # The `smoothing_length` here is used for scaling
-            # TODO move this to the surface tension model since this is not a general thing
             cache.surface_normal[i, particle] += m_b / density_neighbor *
-                                                 grad_kernel[i] * smoothing_length
+                                                 grad_kernel[i]
         end
 
         cache.neighbor_count[particle] += 1
@@ -49,10 +63,9 @@ function calc_normal_akinci!(system, neighbor_system::FluidSystem, u_system, v,
 end
 
 # Section 2.2 in Akinci et al. 2013 "Versatile Surface Tension and Adhesion for SPH Fluids"
-# Note: This is the simplest form of normal approximation commonly used in SPH and comes
-# with serious deficits in accuracy especially at corners, small neighborhoods and boundaries
-function calc_normal_akinci!(system, neighbor_system::BoundarySystem, u_system, v,
-                             v_neighbor_system, u_neighbor_system, semi, surfn)
+# and Section 5 in Morris 2000 "Simulating surface tension with smoothed particle hydrodynamics"
+function calc_normal!(system::FluidSystem, neighbor_system::BoundarySystem, u_system, v,
+                      v_neighbor_system, u_neighbor_system, semi, surfn)
     (; cache) = system
     (; colorfield, colorfield_bnd) = neighbor_system.boundary_model.cache
     (; smoothing_kernel, smoothing_length) = surfn
@@ -101,9 +114,8 @@ function calc_normal_akinci!(system, neighbor_system::BoundarySystem, u_system, 
                                       smoothing_length)
             for i in 1:ndims(system)
                 # The `smoothing_length` here is used for scaling
-                # TODO move this to the surface tension model since this is not a general thing
                 cache.surface_normal[i, particle] += m_b / density_neighbor *
[…]
