+++ docs/src/preprocessing/preprocessing.md
@@ -3,25 +3,61 @@ Generating the initial configuration of a simulation requires filling volumes (3
 The algorithm to sample a complex geometry should be robust and fast,
 since for large problems (high number of particles) or complex geometries (many geometry-faces),
 generating the initial configuration is not trivial and can be very expensive in terms of computational costs.
-We therefore use a winding number approach for an inside-outside segmentation of an object.
+We therefore use a [winding number](https://en.wikipedia.org/wiki/Winding_number) approach for an inside-outside segmentation of an object.
+The winding number $w(\mathbf{p})$ is a signed integer-valued property of a point $\mathbf{p}$ and is defined as
 
-We provide the following methods to calculate the [winding number](https://en.wikipedia.org/wiki/Winding_number):
+$w(\mathbf{p}) = \frac{1}{2 \pi} \sum^n_{i=1} \Theta_i,$
+
+where  $\Theta_i$ is the **signed** angle between vectors from two consecutive vertices to $\mathbf{p}$.
+In 3D, we refer to the solid angle of an *oriented* triangle with respect to $\mathbf{p}$
+
+We provide the following methods to calculate $w(\mathbf{p})$:
 - Horman et al. (2001) evaluated the winding number combined with an even-odd rule and is only for 2D polygons (see [WindingNumberHorman](@ref)).
 - Naive winding: Jacobson et al. (2013) generalized the winding number so that the algorithm can be applied for 2D and also for 3D geometries (see [WindingNumberJacobson](@ref)).
 - Hierarchical winding: Jacobson et al. (2013) also introduced a fast hierarchical evaluation of the winding number. For further information see description below.
 
 
-## Hierarchical Winding
+## [Hierarchical Winding](@id hierarchical_winding)
 According to Jacobson et al. (2013) the winding number is the sum of harmonic functions defined for each edge (2D) or face (3D).
-We can show this with the following example in which we determine the winding number for each edge of a triangle separately and sum them up (TODO: insert code):
+We can show this with the following example in which we determine the winding number for each edge of a triangle separately and sum them up:
 
-$w = w_1 + w_2 + w_3$
+```julia
+using TrixiParticles
 
-![alt text](image.png)
+particle_spacing = 0.025
+
+triangle = stack([[0.0, 0.0], [1.0, 0.0], [0.5, 0.7], [0.0, 0.0]])
+
+edge1 = deleteat!(TrixiParticles.Polygon(triangle), [2, 3])
+edge2 = deleteat!(TrixiParticles.Polygon(triangle), [1, 3])
+edge3 = deleteat!(TrixiParticles.Polygon(triangle), [1, 2])
+
+algorithm = WindingNumberJacobson()
+
+grid = RectangularShape(particle_spacing, (80, 80), (-0.5, -0.7); tlsph=true, density=1.0)
+
+inpoly, w1 = algorithm(edge1, grid.coordinates; store_winding_number=true)
+inpoly, w2 = algorithm(edge2, grid.coordinates; store_winding_number=true)
+inpoly, w3 = algorithm(edge3, grid.coordinates; store_winding_number=true)
+
+w = w1 + w2 + w3
+
+trixi2vtk(grid, w=w, w1=w1, w2=w2, w3=w3)
+```
+
+```@raw html
+<figure>
+  <img src="https://github.com/user-attachments/assets/89f0b7e3-5d47-4eba-9d29-34cf54f1b246" alt="triangle"/>
+</figure>
+```
 
[…]
