### 2024-08-01 19b2c6c
implement suggestions
