+++ examples/preprocessing/complex_shape_2d.jl
@@ -1,8 +1,9 @@
 using TrixiParticles
+using Plots
 
 particle_spacing = 0.05
 
-filename = "hexagon"
+filename = "inverted_open_curve"
 file = joinpath("examples", "preprocessing", "data", filename * ".asc")
 
 geometry = load_geometry(file)
@@ -10,7 +11,7 @@ geometry = load_geometry(file)
 trixi2vtk(geometry)
 
 point_in_geometry_algorithm = WindingNumberJacobson(; geometry,
-                                                    #winding_number_factor=0.4,
+                                                    winding_number_factor=0.4,
                                                     hierarchical_winding=true)
 
 # Returns `InitialCondition`
@@ -19,4 +20,9 @@ shape_sampled = ComplexShape(geometry; particle_spacing, density=1.0,
                              point_in_geometry_algorithm)
 
 trixi2vtk(shape_sampled.initial_condition)
-trixi2vtk(shape_sampled.grid, w=shape_sampled.winding_numbers)
+
+# Plot the winding number field
+plot(InitialCondition(; coordinates=shape_sampled.grid, density=1.0, particle_spacing),
+     zcolor=shape_sampled.winding_numbers)
+
+# trixi2vtk(shape_sampled.grid, w=shape_sampled.winding_numbers)
