+++ test/preprocessing/geometries/geometries.jl
@@ -17,7 +17,7 @@
                 rot([-1.0, 0.0], rot_angle),
                 rot([0.0, 1.0], rot_angle),
                 rot([1.0, 0.0], rot_angle),
-                rot([0.0, -1.0], rot_angle)
+                rot([0.0, -1.0], rot_angle),
             ]
 
             sqrt2 = sqrt(2)
@@ -25,7 +25,7 @@
                 rot([[-sqrt2 / 2, -sqrt2 / 2], [-sqrt2 / 2, sqrt2 / 2]], rot_angle), # edge 1
                 rot([[-sqrt2 / 2, sqrt2 / 2], [sqrt2 / 2, sqrt2 / 2]], rot_angle),   # edge 2
                 rot([[sqrt2 / 2, sqrt2 / 2], [sqrt2 / 2, -sqrt2 / 2]], rot_angle),   # edge 3
-                rot([[sqrt2 / 2, -sqrt2 / 2], [-sqrt2 / 2, -sqrt2 / 2]], rot_angle) # edge 4
+                rot([[sqrt2 / 2, -sqrt2 / 2], [-sqrt2 / 2, -sqrt2 / 2]], rot_angle), # edge 4
             ]
 
             geometry_clockwise = TrixiParticles.Polygon(points_rectangular_clockwise)
@@ -55,7 +55,7 @@
     end
 
     @testset verbose=true "Real World Data" begin
-        data_dir = pkgdir(TrixiParticles, "test", "preprocessing", "data")
+        data_dir = pkgdir(TrixiParticles, "examples", "preprocessing", "data")
 
         @testset verbose=true "2D" begin
             files = ["hexagon", "circle", "inverted_open_curve"]
@@ -63,7 +63,7 @@
 
             @testset "Test File `$(files[i])`" for i in eachindex(files)
                 # Checked in ParaView with `trixi2vtk(geometry)`
-                data = TrixiParticles.CSV.read(joinpath(data_dir,
+                data = TrixiParticles.CSV.read(joinpath(validation_dir(), "preprocessing",
                                                         "normals_" * files[i] * ".csv"),
                                                TrixiParticles.DataFrame)
 
@@ -92,7 +92,7 @@
 
             @testset "Test File `$(files[i])`" for i in eachindex(files)
                 # Checked in ParaView with `trixi2vtk(geometry)`
-                data = TrixiParticles.CSV.read(joinpath(data_dir,
+                data = TrixiParticles.CSV.read(joinpath(validation_dir(), "preprocessing",
                                                         "normals_" * files[i] * ".csv"),
                                                TrixiParticles.DataFrame)
 
@@ -143,44 +143,44 @@ end
 # ==========================================================================================
 # ==== The following functions are only used for debugging yet
 
-function save(filename, mesh; faces=TrixiParticles.eachface(mesh))
-    save(TrixiParticles.File{TrixiParticles.format"STL_BINARY"}(filename), mesh; faces)
-end
+# function save(filename, mesh; faces=TrixiParticles.eachface(mesh))
+#     save(TrixiParticles.File{TrixiParticles.format"STL_BINARY"}(filename), mesh; faces)
+# end
 
-function save(fn::TrixiParticles.File{TrixiParticles.format"STL_BINARY"},
-              mesh::TrixiParticles.TriangleMesh; faces=TrixiParticles.eachface(mesh))
-    open(fn, "w") do s
-        save(s, mesh; faces)
-    end
-end
+# function save(fn::TrixiParticles.File{TrixiParticles.format"STL_BINARY"},
+#               mesh::TrixiParticles.TriangleMesh; faces=TrixiParticles.eachface(mesh))
[…]
