+++ src/setups/complex_shape.jl
@@ -50,6 +50,10 @@ function ComplexShape(geometry::Union{TriangleMesh, Polygon}; particle_spacing, 
         throw(ArgumentError("`WindingNumberHorman` only supports 2D geometries"))
     end
 
+    if grid_offset < 0.0
+        thorw(ArgumentError("only a positive `grid_offset` is supported"))
+    end
+
     return sample(geometry; particle_spacing, density, pressure, mass, velocity,
                   point_in_geometry_algorithm, store_winding_number, grid_offset,
                   max_nparticles,
