+++ examples/fluid/sphere_surface_tension_2d.jl
@@ -5,7 +5,7 @@ using OrdinaryDiffEq
 
 fluid_density = 1000.0
 
-particle_spacing = 0.1
+particle_spacing = 0.05
 
 # Note: Only square shapes will result in a sphere.
 # Furthermore, changes of the coefficients might be necessary for higher resolutions or larger squares.
@@ -20,7 +20,7 @@ state_equation = StateEquationCole(; sound_speed, reference_density=fluid_densit
 # smoothing_kernel = WendlandC2Kernel{2}()
 # nu = 0.01
 
-smoothing_length = 1.0 * particle_spacing
+smoothing_length = 1.5 * particle_spacing
 fluid_smoothing_kernel = SchoenbergCubicSplineKernel{2}()
 nu = 0.025
 
@@ -29,20 +29,27 @@ fluid = RectangularShape(particle_spacing, round.(Int, fluid_size ./ particle_sp
 
 alpha = 8 * nu / (smoothing_length * sound_speed)
 source_terms = SourceTermDamping(; damping_coefficient=0.5)
+# fluid_system = WeaklyCompressibleSPHSystem(fluid, SummationDensity(),
+#                                            state_equation, fluid_smoothing_kernel,
+#                                            smoothing_length,
+#                                            viscosity=ArtificialViscosityMonaghan(alpha=alpha,
+#                                                                                  beta=0.0),
+#                                            surface_tension=SurfaceTensionAkinci(surface_tension_coefficient=0.02),
+#                                            correction=AkinciFreeSurfaceCorrection(fluid_density),
+#                                            source_terms=source_terms)
 fluid_system = WeaklyCompressibleSPHSystem(fluid, SummationDensity(),
                                            state_equation, fluid_smoothing_kernel,
                                            smoothing_length,
                                            viscosity=ArtificialViscosityMonaghan(alpha=alpha,
                                                                                  beta=0.0),
-                                           surface_tension=SurfaceTensionAkinci(surface_tension_coefficient=0.02),
-                                           correction=AkinciFreeSurfaceCorrection(fluid_density),
+                                           surface_tension=SurfaceTensionMorris(surface_tension_coefficient=0.1),
                                            source_terms=source_terms)
 
 # ==========================================================================================
 # ==== Simulation
 semi = Semidiscretization(fluid_system)
 
-tspan = (0.0, 3.0)
+tspan = (0.0, 20.0)
 ode = semidiscretize(semi, tspan)
 
 info_callback = InfoCallback(interval=100)
