+++ src/schemes/fluid/surface_tension.jl
@@ -159,7 +159,7 @@ end
 
     return cohesion_force_akinci(surface_tension_a, support_radius, m_b,
                                  pos_diff, distance) .-
-           (surface_tension_coefficient * (n_a - n_b) * smoothing_length)
+           (surface_tension_coefficient * (n_a - n_b) * support_radius)
 end
 
 @inline function surface_tension_force(surface_tension_a::SurfaceTensionMorris,
