+++ examples/fluid/falling_water_spheres_morris_2d.jl
@@ -5,7 +5,7 @@ using OrdinaryDiffEq
 
 # ==========================================================================================
 # ==== Resolution
-fluid_particle_spacing = 0.005
+fluid_particle_spacing = 0.003
 
 boundary_layers = 3
 spacing_ratio = 1
@@ -17,7 +17,7 @@ tspan = (0.0, 0.3)
 
 # Boundary geometry and initial fluid particle positions
 initial_fluid_size = (0.0, 0.0)
-tank_size = (2.0, 0.5)
+tank_size = (2.5, 0.5)
 
 fluid_density = 1000.0
 sound_speed = 100
@@ -32,7 +32,7 @@ tank = RectangularTank(fluid_particle_spacing, initial_fluid_size, tank_size, fl
 sphere_radius = 0.05
 
 sphere1_center = (0.5, 0.2)
-sphere2_center = (1.5, 0.2)
+sphere2_center = (1.75, 0.2)
 sphere1 = SphereShape(fluid_particle_spacing, sphere_radius, sphere1_center,
                       fluid_density, sphere_type=VoxelSphere(), velocity=(0.0, -3.0))
 sphere2 = SphereShape(fluid_particle_spacing, sphere_radius, sphere2_center,
@@ -40,28 +40,29 @@ sphere2 = SphereShape(fluid_particle_spacing, sphere_radius, sphere2_center,
 
 # ==========================================================================================
 # ==== Fluid
-fluid_smoothing_length = 1.0 * fluid_particle_spacing - eps()
-fluid_smoothing_kernel = SchoenbergCubicSplineKernel{2}()
+fluid_smoothing_length = 3.5 * fluid_particle_spacing
+fluid_smoothing_kernel = WendlandC2Kernel{2}()
 
 fluid_density_calculator = ContinuityDensity()
 
 nu = 0.005
 alpha = 8 * nu / (fluid_smoothing_length * sound_speed)
 viscosity = ArtificialViscosityMonaghan(alpha=alpha, beta=0.0)
-density_diffusion = DensityDiffusionAntuono(sphere2, delta=0.1)
+# density_diffusion = DensityDiffusionAntuono(sphere2, delta=0.1)
 
 sphere_surface_tension = EntropicallyDampedSPHSystem(sphere1, fluid_smoothing_kernel,
                                                      fluid_smoothing_length,
                                                      sound_speed, viscosity=viscosity,
                                                      density_calculator=ContinuityDensity(),
                                                      acceleration=(0.0, -gravity),
-                                                     surface_tension=SurfaceTensionMorris(surface_tension_coefficient=0.05))
+                                                     surface_tension=SurfaceTensionMorris(surface_tension_coefficient=200*0.0728))
 
-sphere = WeaklyCompressibleSPHSystem(sphere2, fluid_density_calculator,
-                                     state_equation, fluid_smoothing_kernel,
-                                     fluid_smoothing_length, viscosity=viscosity,
-                                     density_diffusion=density_diffusion,
-                                     acceleration=(0.0, -gravity))
+sphere = EntropicallyDampedSPHSystem(sphere2, fluid_smoothing_kernel,
+                                                     fluid_smoothing_length,
+                                                     sound_speed, viscosity=viscosity,
+                                                     density_calculator=ContinuityDensity(),
+                                                     acceleration=(0.0, -gravity),
+                                                     surface_normal_method=ColorfieldSurfaceNormal(fluid_smoothing_kernel, fluid_smoothing_length))
 
 # ==========================================================================================
 # ==== Boundary
@@ -73,8 +74,7 @@ boundary_model = BoundaryModelDummyParticles(tank.boundary.density, tank.boundar
[…]
