+++ src/schemes/fluid/weakly_compressible_sph/system.jl
@@ -364,11 +364,6 @@ end
     extract_smatrix(system.cache.correction_matrix, system, particle)
 end
 
-@inline function surface_normal(particle_system::FluidSystem, particle)
-    (; cache) = particle_system
-    return extract_svector(cache.surface_normal, particle_system, particle)
-end
-
 @inline function surface_tension_model(system::FluidSystem)
     return system.surface_tension
 end
@@ -379,5 +374,5 @@ end
 
 @inline function curvature(particle_system::FluidSystem, particle)
     (; cache) = particle_system
-    return extract_svector(cache.curvature, particle_system, particle)
+    return cache.curvature[particle]
 end
