### 2024-08-02 eb4edbd
implement suggestions
