### 2024-08-02 80aff81
modify NEWS.md
