+++ examples/preprocessing/complex_shape_3d.jl
@@ -14,6 +14,8 @@ point_in_geometry_algorithm = WindingNumberJacobson(; geometry,
 # Returns `InitialCondition`
 shape_sampled = ComplexShape(geometry; particle_spacing, density=1.0,
                              boundary_thickness=5particle_spacing,
+                             create_signed_distance_field=true,
                              sample_boundary=false, point_in_geometry_algorithm)
 
 trixi2vtk(shape_sampled.initial_condition, filename="initial_condition_fluid")
+trixi2vtk(shape_sampled.signed_distance_field)
