+++ src/preprocessing/particle_packing/system.jl
@@ -34,31 +34,30 @@ For more information about the methods, see desription below.
 - `smoothing_length`: Smoothing length to be used for this system.
                       See [Smoothing Kernels](@ref smoothing_kernel).
 """
-struct ParticlePackingSystem{NDIMS, ELTYPE <: Real, IC, B, K, S,
-                             C, N} <: FluidSystem{NDIMS, IC}
-    initial_condition              :: IC
-    boundary                       :: B
-    smoothing_kernel               :: K
-    smoothing_length               :: ELTYPE
-    background_pressure            :: ELTYPE
-    tlsph                          :: Bool
-    signed_distance_field          :: S
-    is_boundary                    :: Bool
-    shift_condition                :: ELTYPE
-    constrain_particles_on_surface :: C
-    neighborhood_search            :: N
-    signed_distances               :: Vector{ELTYPE} # Only for visualization
-    buffer                         :: Nothing
-    update_callback_used           :: Ref{Bool}
+struct ParticlePackingSystem{NDIMS, ELTYPE <: Real, IC, B, K,
+                             S, N} <: FluidSystem{NDIMS, IC}
+    initial_condition     :: IC
+    boundary              :: B
+    smoothing_kernel      :: K
+    smoothing_length      :: ELTYPE
+    background_pressure   :: ELTYPE
+    tlsph                 :: Bool
+    signed_distance_field :: S
+    is_boundary           :: Bool
+    shift_condition       :: ELTYPE
+    neighborhood_search   :: N
+    signed_distances      :: Vector{ELTYPE} # Only for visualization
+    buffer                :: Nothing
+    update_callback_used  :: Ref{Bool}
 
-    function ParticlePackingSystem(shape;
+    function ParticlePackingSystem(shape::Union{InitialCondition, ComplexShape};
                                    smoothing_kernel=SchoenbergCubicSplineKernel{ndims(shape)}(),
-                                   smoothing_length=1.2complex_shape.particle_spacing,
-                                   signed_distance_field=shape.signed_distance_field,
-                                   is_boundary=false, boundary=shape.shape,
-                                   neighborhood_search=true,
-                                   background_pressure, tlsph=false)
-        (; particle_spacing) = shape
+                                   smoothing_length=1.2shape.particle_spacing,
+                                   signed_distance_field::SignedDistanceField=shape.signed_distance_field,
+                                   is_boundary=false,
+                                   boundary::Union{Polygon, TriangleMesh}=shape.geometry,
+                                   neighborhood_search=GridNeighborhoodSearch{ndims(shape)}(),
+                                   background_pressure, tlsph=true)
         NDIMS = ndims(shape)
         ELTYPE = eltype(shape)
 
@@ -66,51 +65,31 @@ struct ParticlePackingSystem{NDIMS, ELTYPE <: Real, IC, B, K, S,
             throw(ArgumentError("smoothing kernel dimensionality must be $NDIMS for a $(NDIMS)D problem"))
         end
 
-        search_radius = compact_support(smoothing_kernel, smoothing_length)
-
         # Create neighborhood search
-        if neighborhood_search && isnothing(signed_distance_field)
-            # Calculate signed distance on the fly using a face nhs
+        nhs_ = isnothing(neighborhood_search) ? TrivialNeighborhoodSearch{NDIMS}() :
+               neighborhood_search
+        nhs = copy_neighborhood_search(nhs_,
+                                       compact_support(smoothing_kernel, smoothing_length),
+                                       length(signed_distance_field.positions))
 
-            nhs = FaceNeighborhoodSearch{NDIMS}(search_radius)
-            initialize!(nhs, boundary)
-
-        elseif signed_distance_field isa SignedDistanceField
[…]
