### 2024-08-05 3ebddfe
implement suggestions
