+++ test/preprocessing/geometries/geometries.jl
@@ -1,6 +1,6 @@
-@testset verbose=true "Shapes" begin
+@testset verbose=true "Geometries" begin
     @testset verbose=true "Rectangular Analytical" begin
-        rot(vec, α) = SVector{2}([cos(α) -sin(α); sin(α) cos(α)] * vec)
+        rot(vec, alpha) = SVector{2}([cos(alpha) -sin(alpha); sin(alpha) cos(alpha)] * vec)
 
         @testset "Rotation Angle $(rot_angle)" for rot_angle in 0.0:(pi / 8):(2pi)
             v1 = rot([0.0, 0.0], rot_angle)
@@ -22,10 +22,10 @@
 
             sqrt2 = sqrt(2)
             vertex_normals = [
-                rot([[-sqrt2 / 2, -sqrt2 / 2], [-sqrt2 / 2, sqrt2 / 2]], rot_angle), # edge 1
-                rot([[-sqrt2 / 2, sqrt2 / 2], [sqrt2 / 2, sqrt2 / 2]], rot_angle),   # edge 2
-                rot([[sqrt2 / 2, sqrt2 / 2], [sqrt2 / 2, -sqrt2 / 2]], rot_angle),   # edge 3
-                rot([[sqrt2 / 2, -sqrt2 / 2], [-sqrt2 / 2, -sqrt2 / 2]], rot_angle), # edge 4
+                rot.([[-sqrt2 / 2, -sqrt2 / 2], [-sqrt2 / 2, sqrt2 / 2]], rot_angle), # edge 1
+                rot.([[-sqrt2 / 2, sqrt2 / 2], [sqrt2 / 2, sqrt2 / 2]], rot_angle),   # edge 2
+                rot.([[sqrt2 / 2, sqrt2 / 2], [sqrt2 / 2, -sqrt2 / 2]], rot_angle),   # edge 3
+                rot.([[sqrt2 / 2, -sqrt2 / 2], [-sqrt2 / 2, -sqrt2 / 2]], rot_angle), # edge 4
             ]
 
             geometry_clockwise = TrixiParticles.Polygon(points_rectangular_clockwise)
@@ -56,6 +56,7 @@
 
     @testset verbose=true "Real World Data" begin
         data_dir = pkgdir(TrixiParticles, "examples", "preprocessing", "data")
+        validation_dir = pkgdir(TrixiParticles, "test", "preprocessing", "data")
 
         @testset verbose=true "2D" begin
             files = ["hexagon", "circle", "inverted_open_curve"]
@@ -63,7 +64,7 @@
 
             @testset "Test File `$(files[i])`" for i in eachindex(files)
                 # Checked in ParaView with `trixi2vtk(geometry)`
-                data = TrixiParticles.CSV.read(joinpath(validation_dir(), "preprocessing",
+                data = TrixiParticles.CSV.read(joinpath(validation_dir,
                                                         "normals_" * files[i] * ".csv"),
                                                TrixiParticles.DataFrame)
 
@@ -92,7 +93,7 @@
 
             @testset "Test File `$(files[i])`" for i in eachindex(files)
                 # Checked in ParaView with `trixi2vtk(geometry)`
-                data = TrixiParticles.CSV.read(joinpath(validation_dir(), "preprocessing",
+                data = TrixiParticles.CSV.read(joinpath(validation_dir,
                                                         "normals_" * files[i] * ".csv"),
                                                TrixiParticles.DataFrame)
 
@@ -139,48 +140,3 @@
         @test TrixiParticles.unique_sorted(copy(x)) == sort(unique(x))
     end
 end
-
-# ==========================================================================================
-# ==== The following functions are only used for debugging yet
-
-# function save(filename, mesh; faces=TrixiParticles.eachface(mesh))
[…]
