+++ .typos.toml
@@ -4,3 +4,4 @@ Shepard = "Shepard"
 shepard = "shepard"
 Strack = "Strack"
 Lok = "Lok"
+SPlisHSPlasH = "SPlisHSPlasH"
