### 2024-08-06 5ab14ee
adapt examples
