+++ examples/fluid/lid_driven_cavity_2d.jl
@@ -7,13 +7,13 @@
 
 using TrixiParticles
 using OrdinaryDiffEq
+
 # ==========================================================================================
 # ==== Resolution
 particle_spacing = 0.02
 
-# Change spacing ratio to 3 and boundary layers to 1 when using Monaghan-Kajtar boundary model
+# Make sure that the kernel support of fluid particles at a boundary is always fully sampled
 boundary_layers = 3
-spacing_ratio = 1
 
 # ==========================================================================================
 # ==== Experiment Setup
@@ -24,15 +24,15 @@ cavity_size = (1.0, 1.0)
 
 fluid_density = 1.0
 
-const velocity_lid = 1.0
-sound_speed = 10 * velocity_lid
+const VELOCITY_LID = 1.0
+sound_speed = 10 * VELOCITY_LID
 
 pressure = sound_speed^2 * fluid_density
 
-viscosity = ViscosityAdami(; nu=velocity_lid / reynolds_number)
+viscosity = ViscosityAdami(; nu=VELOCITY_LID / reynolds_number)
 
 cavity = RectangularTank(particle_spacing, cavity_size, cavity_size, fluid_density,
-                         n_layers=boundary_layers, spacing_ratio=spacing_ratio,
+                         n_layers=boundary_layers,
                          faces=(true, true, true, false), pressure=pressure)
 
 lid_position = 0.0 - particle_spacing * boundary_layers
@@ -48,13 +48,14 @@ smoothing_length = 1.0 * particle_spacing
 smoothing_kernel = SchoenbergQuinticSplineKernel{2}()
 
 fluid_system = EntropicallyDampedSPHSystem(cavity.fluid, smoothing_kernel, smoothing_length,
+                                           density_calculator=ContinuityDensity(),
                                            sound_speed, viscosity=viscosity,
                                            transport_velocity=TransportVelocityAdami(pressure))
 
 # ==========================================================================================
 # ==== Boundary
 
-movement_function(t) = SVector(velocity_lid * t, 0.0)
+movement_function(t) = SVector(VELOCITY_LID * t, 0.0)
 
 is_moving(t) = true
 
@@ -78,11 +79,11 @@ boundary_system_lid = BoundarySPHSystem(lid, boundary_model_lid, movement=moveme
 # ==========================================================================================
 # ==== Simulation
 bnd_thickness = boundary_layers * particle_spacing
+periodic_box = PeriodicBox(min_corner=[-bnd_thickness, -bnd_thickness],
+                           max_corner=cavity_size .+ [bnd_thickness, bnd_thickness])
+
 semi = Semidiscretization(fluid_system, boundary_system_cavity, boundary_system_lid,
-                          neighborhood_search=GridNeighborhoodSearch,
-                          periodic_box_min_corner=[-bnd_thickness, -bnd_thickness],
-                          periodic_box_max_corner=cavity_size .+
-                                                  [bnd_thickness, bnd_thickness])
+                          neighborhood_search=GridNeighborhoodSearch{2}(; periodic_box))
 
 ode = semidiscretize(semi, tspan)
 
@@ -91,14 +92,7 @@ info_callback = InfoCallback(interval=100)
 saving_callback = SolutionSavingCallback(dt=0.02)
 callbacks = CallbackSet(info_callback, saving_callback, UpdateCallback())
 
-# Use a Runge-Kutta method with automatic (error based) time step size control.
-# Enable threading of the RK method for better performance on multiple threads.
-# Limiting of the maximum stepsize is necessary to prevent crashing.
-# When particles are approaching a wall in a uniform way, they can be advanced
-# with large time steps. Close to the wall, the stepsize has to be reduced drastically.
[…]
