### 2024-08-06 6828fcf
Reopen #351: Complex shapes (#439)
* winding number proof of concept 2D

* faces and edges with `NTuples`

* change directory

* remove `@unpack`

* modify

* return `InitialCondition`

* add examples

* first working API

* add complex tank example

* fix typos

* thread parallel

* optional winding number range

* add `UpdateCallback`

* fix typo

* apply formatter

* simplify triangle mesh

* rework data types

* restructure

* SVector min-max-box

* make `TriangleMesh` faster

* fix vertex normal

* add IO

* add hierarchical winding

* store winding number

* clean up TriangleMesh

* modi
[…]
