+++ .typos.toml
@@ -1,7 +1,9 @@
+[default.extend-identifiers]
+SPlisHSPlasH = "SPlisHSPlasH"
+
 [default.extend-words]
 ba = "ba"
 Shepard = "Shepard"
 shepard = "shepard"
 Strack = "Strack"
 Lok = "Lok"
-SPlisHSPlasH = "SPlisHSPlasH"
