+++ Project.toml
@@ -41,7 +41,7 @@ GPUArraysCore = "0.1"
 JSON = "0.21"
 KernelAbstractions = "0.9"
 MuladdMacro = "0.2"
-PointNeighbors = "0.4.2"
+PointNeighbors = "0.4.5"
 Polyester = "0.7.5"
 RecipesBase = "1"
 Reexport = "1"
