+++ Project.toml
@@ -42,7 +42,7 @@ JSON = "0.21"
 KernelAbstractions = "0.9"
 MuladdMacro = "0.2"
 PointNeighbors = "0.4.5"
-Polyester = "0.7.5"
+Polyester = "0.7.15"
 RecipesBase = "1"
 Reexport = "1"
 SciMLBase = "1, 2"
