+++ Project.toml
@@ -41,8 +41,8 @@ GPUArraysCore = "0.1"
 JSON = "0.21"
 KernelAbstractions = "0.9"
 MuladdMacro = "0.2"
-PointNeighbors = "0.4.5"
-Polyester = "0.7.15"
+PointNeighbors = "0.4.2"
+Polyester = "0.7.8"
 RecipesBase = "1"
 Reexport = "1"
 SciMLBase = "1, 2"
