+++ Project.toml
@@ -42,7 +42,7 @@ JSON = "0.21"
 KernelAbstractions = "0.9"
 MuladdMacro = "0.2"
 PointNeighbors = "0.4.2"
-Polyester = "0.7.8"
+Polyester = "0.7.10"
 RecipesBase = "1"
 Reexport = "1"
 SciMLBase = "1, 2"
