### 2024-08-06 613a0d4
fix computation for a single color
