+++ examples/fluid/sphere_surface_tension_2d.jl
@@ -20,8 +20,8 @@ state_equation = StateEquationCole(; sound_speed, reference_density=fluid_densit
 # smoothing_kernel = WendlandC2Kernel{2}()
 # nu = 0.01
 
-smoothing_length = 1.5 * particle_spacing
-fluid_smoothing_kernel = SchoenbergCubicSplineKernel{2}()
+smoothing_length = 3.5 * particle_spacing
+fluid_smoothing_kernel = WendlandC2Kernel{2}()
 nu = 0.025
 
 fluid = RectangularShape(particle_spacing, round.(Int, fluid_size ./ particle_spacing),
@@ -37,13 +37,18 @@ source_terms = SourceTermDamping(; damping_coefficient=0.5)
 #                                            surface_tension=SurfaceTensionAkinci(surface_tension_coefficient=0.02),
 #                                            correction=AkinciFreeSurfaceCorrection(fluid_density),
 #                                            source_terms=source_terms)
-fluid_system = WeaklyCompressibleSPHSystem(fluid, SummationDensity(),
-                                           state_equation, fluid_smoothing_kernel,
+
+fluid_system = EntropicallyDampedSPHSystem(fluid, fluid_smoothing_kernel,
                                            smoothing_length,
+                                           sound_speed,
                                            viscosity=ArtificialViscosityMonaghan(alpha=alpha,
                                                                                  beta=0.0),
-                                           surface_tension=SurfaceTensionMorris(surface_tension_coefficient=0.1),
-                                           source_terms=source_terms)
+                                           density_calculator=ContinuityDensity(),
+                                           reference_particle_spacing=particle_spacing,
+                                           acceleration=(0.0, 0.0),
+                                           surface_normal_method=ColorfieldSurfaceNormal(fluid_smoothing_kernel,
+                                                                                         smoothing_length),
+                                           surface_tension=SurfaceTensionMorris(surface_tension_coefficient=100 * 0.0728))
 
 # ==========================================================================================
 # ==== Simulation
