### 2024-08-06 95d6a01
extend hydrostatic_water_column tests
