### 2024-08-07 9912935
CompatHelper: add new compat entry for DelimitedFiles at version 1, (keep existing compat)
