+++ Project.toml
@@ -34,6 +34,7 @@ WriteVTK = "64499a7a-5c06-52f2-abe2-ccb03c286192"
 Adapt = "3, 4"
 CSV = "0.10"
 DataFrames = "1.6"
+DelimitedFiles = "1"
 DiffEqCallbacks = "2.25, 3"
 FastPow = "0.1"
 FileIO = "1"
