### 2024-08-07 de94084
Fix threaded winding number calculation (#594)
* Fix threaded winding number calculation

* Update version

* Update NEWS.md
