### 2024-08-07 7bb3bc8
add bib and cite references
