+++ src/schemes/fluid/weakly_compressible_sph/density_diffusion.jl
@@ -23,7 +23,7 @@ end
 @doc raw"""
     DensityDiffusionMolteniColagrossi(; delta)
 
-The commonly used density diffusion term by Molteni & Colagrossi (2009).
+The commonly used density diffusion term by [Molteni2009](@cite).
 
 The term ``\psi_{ab}`` in the continuity equation in [`DensityDiffusion`](@ref) is defined
 by
@@ -35,12 +35,6 @@ and ``r_{ab} = r_a - r_b`` is the difference of the coordinates of particles ``a
 
 See [`DensityDiffusion`](@ref) for an overview and comparison of implemented density
 diffusion terms.
-
-## References
-- Diego Molteni, Andrea Colagrossi.
-  "A Simple Procedure to Improve the Pressure Evaluation in Hydrodynamic Context Using the SPH."
-  In: Computer Physics Communications 180.6 (2009), pages 861--872.
-  [doi: 10.1016/j.cpc.2008.12.004](https://doi.org/10.1016/j.cpc.2008.12.004)
 """
 struct DensityDiffusionMolteniColagrossi{ELTYPE} <: DensityDiffusion
     delta::ELTYPE
@@ -59,7 +53,7 @@ end
 @doc raw"""
     DensityDiffusionFerrari()
 
-A density diffusion term by Ferrari et al. (2009).
+A density diffusion term by [Ferrari2009](@cite).
 
 The term ``\psi_{ab}`` in the continuity equation in [`DensityDiffusion`](@ref) is defined
 by
@@ -72,12 +66,6 @@ where ``\rho_a`` and ``\rho_b`` denote the densities of particles ``a`` and ``b`
 
 See [`DensityDiffusion`](@ref) for an overview and comparison of implemented density
 diffusion terms.
-
-## References
-- Angela Ferrari, Michael Dumbser, Eleuterio F. Toro, Aronne Armanini.
-  "A New 3D Parallel SPH Scheme for Free Surface Flows."
-  In: Computers & Fluids 38.6 (2009), pages 1203--1217.
-  [doi: 10.1016/j.compfluid.2008.11.012](https://doi.org/10.1016/j.compfluid.2008.11.012).
 """
 struct DensityDiffusionFerrari <: DensityDiffusion
     delta::Int
@@ -97,9 +85,9 @@ end
 @doc raw"""
     DensityDiffusionAntuono(initial_condition; delta)
 
-The commonly used density diffusion terms by Antuono et al. (2010), also referred to as
-δ-SPH. The density diffusion term by Molteni & Colagrossi (2009) is extended by a second
-term, which is nicely written down by Antuono et al. (2012).
+The commonly used density diffusion terms by [Antuono2010](@cite), also referred to as
+δ-SPH. The density diffusion term by [Molteni2009](@cite) is extended by a second
+term, which is nicely written down by [Antuono2012](@cite).
 
 The term ``\psi_{ab}`` in the continuity equation in [`DensityDiffusion`](@ref) is defined
 by
@@ -121,20 +109,6 @@ where ``d`` is the number of dimensions.
 
 See [`DensityDiffusion`](@ref) for an overview and comparison of implemented density
 diffusion terms.
-
-## References
-- M. Antuono, A. Colagrossi, S. Marrone, D. Molteni.
-  "Free-Surface Flows Solved by Means of SPH Schemes with Numerical Diffusive Terms."
-  In: Computer Physics Communications 181.3 (2010), pages 532--549.
-  [doi: 10.1016/j.cpc.2009.11.002](https://doi.org/10.1016/j.cpc.2009.11.002)
[…]
