+++ src/TrixiParticles.jl
@@ -1,11 +1,11 @@
 module TrixiParticles
 
 # Use the README as the module docs
-@doc let
-    path = joinpath(dirname(@__DIR__), "README.md")
-    include_dependency(path)
-    read(path, String)
-end TrixiParticles
+# @doc let
+#     path = joinpath(dirname(@__DIR__), "README.md")
+#     include_dependency(path)
+#     read(path, String)
+# end TrixiParticles
 
 using Reexport: @reexport
 
