+++ src/TrixiParticles.jl
@@ -1,12 +1,5 @@
 module TrixiParticles
 
-# Use the README as the module docs
-# @doc let
-#     path = joinpath(dirname(@__DIR__), "README.md")
-#     include_dependency(path)
-#     read(path, String)
-# end TrixiParticles
-
 using Reexport: @reexport
 
 using Adapt: Adapt
