### 2024-08-07 20d222b
remove surface tension example with dam break
