### 2024-08-07 91aa702
simplify accelerated_tank example
