### 2024-08-07 317e045
add oriented bounding box
