### 2024-08-07 3778b05
fix accelerated_tank
