+++ examples/fluid/accelerated_tank_2d.jl
@@ -16,4 +16,5 @@ boundary_movement = BoundaryMovement(movement_function, is_moving)
 
 trixi_include(@__MODULE__,
               joinpath(examples_dir(), "fluid", "hydrostatic_water_column_2d.jl"),
-              fluid_particle_spacing=fluid_particle_spacing, movement=boundary_movement)
+              fluid_particle_spacing=fluid_particle_spacing, movement=boundary_movement,
+              tspan=(0.0, 1.0))
