### 2024-08-08 a756ef5
run more tests for EDAC
