### 2024-08-08 3d1050d
fix edac tests
