+++ src/schemes/fluid/weakly_compressible_sph/density_diffusion.jl
@@ -239,8 +239,7 @@ end
     distance < sqrt(eps()) && return
 
     (; delta) = density_diffusion
-    (; smoothing_length, state_equation) = particle_system
-    (; sound_speed) = state_equation
+    (; smoothing_length) = particle_system
 
     volume_b = m_b / rho_b
 
@@ -248,7 +247,7 @@ end
                                 particle_system, neighbor_system, particle, neighbor)
     density_diffusion_term = dot(psi, grad_kernel) * volume_b
 
-    dv[end, particle] += delta * smoothing_length * sound_speed * density_diffusion_term
+    dv[end, particle] += delta * smoothing_length * system_sound_speed(particle_system) * density_diffusion_term
 end
 
 # Density diffusion `nothing` or interaction other than fluid-fluid
