+++ docs/src/preprocessing/preprocessing.md
@@ -1,93 +1,106 @@
-# Sample Geometry
+# Sampling of Geometries
 
 Generating the initial configuration of a simulation requires filling volumes (3D) or surfaces (2D) of complex geometries with particles.
 The algorithm to sample a complex geometry should be robust and fast,
 since for large problems (high numbers of particles) or complex geometries (many geometry faces),
 generating the initial configuration is not trivial and can be very expensive in terms of computational cost.
 We therefore use a [winding number](https://en.wikipedia.org/wiki/Winding_number) approach for an inside-outside segmentation of an object.
-The winding number $w(\mathbf{p})$ is a signed integer-valued property of a point $\mathbf{p}$ and is defined as
+The winding number $w(\mathbf{p})$ is a signed integer-valued function of a point $\mathbf{p}$ and is defined as
 
 $w(\mathbf{p}) = \frac{1}{2 \pi} \sum^n_{i=1} \Theta_i,$
 
-where  $\Theta_i$ is the **signed** angle between vectors from two consecutive vertices to $\mathbf{p}$.
-In 3D, we refer to the solid angle of an *oriented* triangle with respect to $\mathbf{p}$
+where  $\Theta_i$ is the **signed** angle between vectors from two consecutive vertices $\mathbf{c}_i$ and $\mathbf{c}_{i+1}$ on a curve to $\mathbf{p}$.
+In 3D, we refer to the solid angle of an *oriented* triangle with respect to $\mathbf{p}$.
 
 We provide the following methods to calculate $w(\mathbf{p})$:
-- Horman et al. (2001) evaluated the winding number combined with an even-odd rule and is only for 2D polygons (see [WindingNumberHorman](@ref)).
-- Naive winding: Jacobson et al. (2013) generalized the winding number so that the algorithm can be applied for 2D and also for 3D geometries (see [WindingNumberJacobson](@ref)).
+- Horman et al. (2001) evaluate the winding number combined with an even-odd rule, but only for 2D polygons (see [WindingNumberHorman](@ref)).
+- Naive winding: Jacobson et al. (2013) generalized the winding number so that the algorithm can be applied for both 2D and 3D geometries (see [WindingNumberJacobson](@ref)).
 - Hierarchical winding: Jacobson et al. (2013) also introduced a fast hierarchical evaluation of the winding number. For further information see the description below.
 
 ## [Hierarchical Winding](@id hierarchical_winding)
-According to Jacobson et al. (2013) the winding number is the sum of harmonic functions defined for each edge (2D) or face (3D).
+According to Jacobson et al. (2013) the winding number with respect to a polygon (2D) or triangle mesh (3D) is the sum of the winding numbers with respect to each edge (2D) or face (3D).
 We can show this with the following example in which we determine the winding number for each edge of a triangle separately and sum them up:
 
 ```julia
 using TrixiParticles
+using Plots
 
-particle_spacing = 0.025
+triangle = [125.0 375.0 250.0 125.0;
+            175.0 175.0 350.0 175.0]
 
-triangle = stack([[0.0, 0.0], [1.0, 0.0], [0.5, 0.7], [0.0, 0.0]])
-
+# Delete all edges but one
 edge1 = deleteat!(TrixiParticles.Polygon(triangle), [2, 3])
 edge2 = deleteat!(TrixiParticles.Polygon(triangle), [1, 3])
[…]
