+++ test/examples/examples.jl
@@ -3,77 +3,77 @@
 @testset verbose=true "Examples" begin
     @testset verbose=true "Fluid" begin
         @trixi_testset "fluid/hydrostatic_water_column_2d.jl" begin
-            # import variables into scope
+            # Import variables into scope
             trixi_include(@__MODULE__,
                           joinpath(examples_dir(), "fluid",
                                    "hydrostatic_water_column_2d.jl"),
                           fluid_system=nothing, sol=nothing, semi=nothing, ode=nothing)
 
             hydrostatic_water_column_tests = Dict(
-                "(WCSPH) default" => (),
-                "(WCSPH) with source term damping" => (source_terms=SourceTermDamping(damping_coefficient=1e-4),),
-                "(WCSPH) with SummationDensity" => (fluid_density_calculator=SummationDensity(),
-                                                    clip_negative_pressure=true),
-                "(WCSPH) with ViscosityAdami" => (
-                                                  # from 0.02*10.0*1.2*0.05/8
-                                                  viscosity=ViscosityAdami(nu=0.0015),),
-                "(WCSPH) with ViscosityMorris" => (
-                                                   # from 0.02*10.0*1.2*0.05/8
-                                                   viscosity=ViscosityMorris(nu=0.0015),),
-                "(WCSPH) with ViscosityAdami and SummationDensity" => (
-                                                                       # from 0.02*10.0*1.2*0.05/8
-                                                                       viscosity=ViscosityAdami(nu=0.0015),
-                                                                       fluid_density_calculator=SummationDensity(),
-                                                                       clip_negative_pressure=true),
-                "(WCSPH) with ViscosityMorris and SummationDensity" => (
-                                                                        # from 0.02*10.0*1.2*0.05/8
-                                                                        viscosity=ViscosityMorris(nu=0.0015),
-                                                                        fluid_density_calculator=SummationDensity(),
-                                                                        clip_negative_pressure=true),
-                "(WCSPH) with smoothing_length=1.3" => (smoothing_length=1.3,),
-                "(WCSPH) with SchoenbergQuarticSplineKernel" => (smoothing_length=1.1,
-                                                                 smoothing_kernel=SchoenbergQuarticSplineKernel{2}()),
-                "(WCSPH) with SchoenbergQuinticSplineKernel" => (smoothing_length=1.1,
-                                                                 smoothing_kernel=SchoenbergQuinticSplineKernel{2}()),
-                "(WCSPH) with WendlandC2Kernel" => (smoothing_length=3.0,
-                                                    smoothing_kernel=WendlandC2Kernel{2}()),
-                "(WCSPH) with WendlandC4Kernel" => (smoothing_length=3.5,
-                                                    smoothing_kernel=WendlandC4Kernel{2}()),
-                "(WCSPH) with WendlandC6Kernel" => (smoothing_length=4.0,
-                                                    smoothing_kernel=WendlandC6Kernel{2}()),
-                "(EDAC) with source term damping" => (source_terms=SourceTermDamping(damping_coefficient=1e-4),
-                                                      fluid_system=EntropicallyDampedSPHSystem(tank.fluid,
-                                                                                               smoothing_kernel,
-                                                                                               smoothing_length,
-                                                                                               sound_speed,
-                                                                                               viscosity=viscosity,
-                                                                                               density_calculator=ContinuityDensity(),
-                                                                                               acceleration=(0.0,
-                                                                                                             -gravity))),
-                "(EDAC) with SummationDensity" => (fluid_system=EntropicallyDampedSPHSystem(tank.fluid,
-                                                                                            smoothing_kernel,
-                                                                                            smoothing_length,
-                                                                                            sound_speed,
-                                                                                            viscosity=viscosity,
-                                                                                            density_calculator=SummationDensity(),
[…]
