### 2024-08-09 eb07671
implement suggestions
