+++ docs/src/preprocessing/preprocessing.md
@@ -2,17 +2,19 @@
 
 Generating the initial configuration of a simulation requires filling volumes (3D) or surfaces (2D) of complex geometries with particles.
 The algorithm to sample a complex geometry should be robust and fast,
-since for large problems (high numbers of particles) or complex geometries (many geometry faces),
+since for large problems (large numbers of particles) or complex geometries (many geometry faces),
 generating the initial configuration is not trivial and can be very expensive in terms of computational cost.
 We therefore use a [winding number](https://en.wikipedia.org/wiki/Winding_number) approach for an inside-outside segmentation of an object.
-The winding number $w(\mathbf{p})$ is a signed integer-valued function of a point $\mathbf{p}$ and is defined as
+The winding number ``w(\mathbf{p})`` is a signed integer-valued function of a point ``\mathbf{p}`` and is defined as
 
-$w(\mathbf{p}) = \frac{1}{2 \pi} \sum^n_{i=1} \Theta_i,$
+```math
+w(\mathbf{p}) = \frac{1}{2 \pi} \sum^n_{i=1} \Theta_i,
+```
 
-where  $\Theta_i$ is the **signed** angle between vectors from two consecutive vertices $\mathbf{c}_i$ and $\mathbf{c}_{i+1}$ on a curve to $\mathbf{p}$.
-In 3D, we refer to the solid angle of an *oriented* triangle with respect to $\mathbf{p}$.
+where  ``\Theta_i`` is the **signed** angle between the directions of ``\mathbf{p}`` to two consecutive vertices ``\mathbf{c}_i`` and ``\mathbf{c}_{i+1}`` on a curve.
+In 3D, we refer to the solid angle of an *oriented* triangle with respect to ``\mathbf{p}``.
 
-We provide the following methods to calculate $w(\mathbf{p})$:
+We provide the following methods to calculate ``w(\mathbf{p})``:
 - Horman et al. (2001) evaluate the winding number combined with an even-odd rule, but only for 2D polygons (see [WindingNumberHorman](@ref)).
 - Naive winding: Jacobson et al. (2013) generalized the winding number so that the algorithm can be applied for both 2D and 3D geometries (see [WindingNumberJacobson](@ref)).
 - Hierarchical winding: Jacobson et al. (2013) also introduced a fast hierarchical evaluation of the winding number. For further information see the description below.
@@ -61,22 +63,29 @@ heatmap(1:500, 1:500, reshape(w, 500, 500)', color=:coolwarm, showaxis=false,
 ```
 
 This summation property has some interesting consequences that we can utilize for an efficient computation of the winding number.
-Let $\mathcal{S}$ be an open surface and $\overline{\mathcal{S}}$ an arbitrary closing surface, such that
+Let ``\mathcal{S}`` be an open surface and ``\bar{\mathcal{S}}`` an arbitrary closing surface, such that
 
-$\partial \overline{\mathcal{S}} = \partial \mathcal{S}$
+```math
+\partial \bar{\mathcal{S}} = \partial \mathcal{S}
+```
 
-and $\mathcal{B} = \overline{\mathcal{S}} \cup \mathcal{S}$ is some closed oriented surface.
-For any query point $\mathbf{p}$ outside of $\mathcal{B}$, we know that
+and ``\mathcal{B} = \bar{\mathcal{S}} \cup \mathcal{S}`` is some closed oriented surface.
+For any query point ``\mathbf{p}`` outside of ``\mathcal{B}``, we know that
 
[…]
