+++ src/schemes/boundary/open_boundary/system.jl
@@ -294,9 +294,6 @@ end
         v_new[dim, particle_new] = v_old[dim, particle_old]
     end
 
-    # Only when using TVF
-    set_transport_velocity!(system_new, particle_new, particle_old, v_new, v_old)
-
     return system_new
 end
 
