+++ docs/src/systems/entropically_damped_sph.md
@@ -57,49 +57,58 @@ Pages = [joinpath("schemes", "fluid", "entropically_damped_sph", "system.jl")]
 
 ## [Transport Velocity Formulation (TVF)](@id transport_velocity_formulation)
 Standard SPH suffers from problems like tensile instability or the creation of void regions in the flow.
-Adami et al (2013) modified the advection velocity and added an extra term in the momentum equation to avoid this problems.
-The authors introduced the so called Transport Velocity Formulation (TVF) for WCSPH. Ramachandran (2019) et al applied the TVF
+To address these problems, Adami et al. (2013) modified the advection velocity and added an extra term to the momentum equation.
+The authors introduced the so-called Transport Velocity Formulation (TVF) for WCSPH. Ramachandran et al. (2019) applied the TVF
 also for the [EDAC](@ref edac) scheme.
 
 The transport velocity ``\tilde{v}_a`` of particle ``a`` is used to evolve the position of the particle ``r_a`` from one time step to the next by
+
 ```math
 \frac{\mathrm{d} r_a}{\mathrm{d}t} = \tilde{v}_a
 ```
+
 and is obtained at every time-step ``\Delta t`` from
+
 ```math
-\tilde{v}_a (t + \Delta t) = v_a (t) + \Delta t \left(\frac{\tilde{\mathrm{d}} v_a}{\mathrm{d}t} - \frac{1}{\rho_a} \nabla p_{\text{background}} \right)
+\tilde{v}_a (t + \Delta t) = v_a (t) + \Delta t \left(\frac{\tilde{\mathrm{d}} v_a}{\mathrm{d}t} - \frac{1}{\rho_a} \nabla p_{\text{background}} \right),
 ```
 
 where ``\rho_a`` is the density of particle ``a`` and ``p_{\text{background}}`` is a constant background pressure field.
 
 The discretized form of the last term is
+
 ```math
- -\frac{1}{\rho_a} \nabla p_{\text{background}} \approx  -\frac{p_{\text{background}}}{m_a} \sum_b \left(V_a^2 + V_b^2 \right) \nabla_a W_{ab}
+ -\frac{1}{\rho_a} \nabla p_{\text{background}} \approx  -\frac{p_{\text{background}}}{m_a} \sum_b \left(V_a^2 + V_b^2 \right) \nabla_a W_{ab},
 ```
+
+where ``V_a``, ``V_b`` denote the volume of particles ``a`` and ``b`` respectively.
 Note that although ``\nabla p_{\text{background}} = 0``, the discretization is not 0th-order consistent for **non**-uniform particle distribution,
 which means that there is a non-vanishing contribution only when particles are disordered.
 That also means that ``p_{\text{background}}`` occurs as prefactor to correct the trajectory of a particle resulting in uniform pressure distributions.
 Suggested is a background pressure which is on the order of the reference pressure but can be chosen arbitrarily large when time-step criterion is adjusted.
 
 The inviscid momentum equation with an additional convection term for a particle moving with ``\tilde{v}`` is
+
 ```math
-\frac{\tilde{\mathrm{d}} \left( \rho v \right)}{\mathrm{d}t} = -\nabla p +  \nabla \cdot \bm{A}
+\frac{\tilde{\mathrm{d}} \left( \rho v \right)}{\mathrm{d}t} = -\nabla p +  \nabla \cdot \bm{A},
 ```
[…]
