+++ examples/fluid/taylor_green_vortex_2d.jl
@@ -80,75 +80,17 @@ semi = Semidiscretization(fluid_system,
 
 ode = semidiscretize(semi, tspan)
 
+info_callback = InfoCallback(interval=100)
+
+saving_callback = SolutionSavingCallback(dt=0.02)
+
+pp_callback = nothing
+
+callbacks = CallbackSet(info_callback, saving_callback, pp_callback, UpdateCallback())
+
 dt_max = min(smoothing_length / 4 * (sound_speed + U), smoothing_length^2 / (8 * nu))
 
-function compute_L1v_error(v, u, t, system)
-    v_analytical_avg = 0.0
-    L1v = 0.0
-
-    for particle in TrixiParticles.eachparticle(system)
-        position = TrixiParticles.current_coords(u, system, particle)
-
-        v_mag = TrixiParticles.norm(TrixiParticles.current_velocity(v, system, particle))
-        v_analytical = TrixiParticles.norm(velocity_function(position, t))
-
-        v_analytical_avg += abs(v_analytical)
-        L1v += abs(v_mag - v_analytical)
-    end
-    v_analytical_avg /= TrixiParticles.nparticles(system)
-
-    L1v /= TrixiParticles.nparticles(system)
-
-    return L1v /= v_analytical_avg
-end
-
-function compute_L1p_error(v, u, t, system)
-    p_max_exact = 0.0
-
-    L1p = 0.0
-
-    for particle in TrixiParticles.eachparticle(system)
-        position = TrixiParticles.current_coords(u, system, particle)
-
-        # compute pressure error
-        p_analytical = pressure_function(position, t)
-        p_max_exact = max(p_max_exact, abs(p_analytical))
-
-        # p_computed - p_average
-        p_computed = TrixiParticles.particle_pressure(v, system, particle) -
-                     TrixiParticles.average_pressure(system, particle)
-        L1p += abs(p_computed - p_analytical)
-    end
-
-    L1p /= TrixiParticles.nparticles(system)
-
-    return L1p /= p_max_exact
-end
-
-# The pressure plotted in the paper is the difference of the local pressure minus
-# the average of the pressure of all particles.
-function diff_p_loc_p_avg(v, u, t, system)
-    p_avg_tot = avg_pressure(v, u, t, system)
-
-    return v[end, :] .- p_avg_tot
-end
-
-info_callback = InfoCallback(interval=100)
-saving_callback = SolutionSavingCallback(dt=0.02,
-                                         L1v=compute_L1v_error,
-                                         L1p=compute_L1p_error,
-                                         diff_p_loc_p_avg=diff_p_loc_p_avg)
-
-callbacks = CallbackSet(info_callback, saving_callback, UpdateCallback())
-
-# Use a Runge-Kutta method with automatic (error based) time step size control.
-# Enable threading of the RK method for better performance on multiple threads.
-# Limiting of the maximum stepsize is necessary to prevent crashing.
-# When particles are approaching a wall in a uniform way, they can be advanced
-# with large time steps. Close to the wall, the stepsize has to be reduced drastically.
-# Sometimes, the method fails to do so with Monaghan-Kajtar BC because forces
-# become extremely large when fluid particles are very close to boundary particles,
-# and the time integration method interprets this as an instability.
+# Use a Runge-Kutta method with automatic (error based) time step size control
 sol = solve(ode, RDPK3SpFSAL49(),
             abstol=1e-8, # Default abstol is 1e-6 (may need to be tuned to prevent boundary penetration)
             reltol=1e-4, # Default reltol is 1e-3 (may need to be tuned to prevent boundary penetration)
