### 2024-08-12 2dd205e
add ldc validation
