+++ src/schemes/fluid/weakly_compressible_sph/density_diffusion.jl
@@ -247,7 +247,8 @@ end
                                 particle_system, neighbor_system, particle, neighbor)
     density_diffusion_term = dot(psi, grad_kernel) * volume_b
 
-    dv[end, particle] += delta * smoothing_length * system_sound_speed(particle_system) * density_diffusion_term
+    dv[end, particle] += delta * smoothing_length * system_sound_speed(particle_system) *
+                         density_diffusion_term
 end
 
 # Density diffusion `nothing` or interaction other than fluid-fluid
