### 2024-08-12 46a999e
increase `maxiters`
