+++ validation/lid_driven_cavity_2d/validation_lid_driven_cavity_2d.jl
@@ -63,7 +63,7 @@ for particle_spacing in particle_spacings, reynolds_number in reynolds_numbers
     global output_directory = joinpath("out_ldc",
                                        "validation_run_lid_driven_cavity_2d_nparticles_$(n_particles_xy)x$(n_particles_xy)_Re_$Re")
 
-    saving_callback = SolutionSavingCallback(dt=0.5, output_directory=output_directory)
+    saving_callback = SolutionSavingCallback(dt=0.1, output_directory=output_directory)
 
     pp_callback = PostprocessCallback(; dt=0.02,
                                       interpolated_velocity=interpolated_velocity,
