+++ examples/fluid/lid_driven_cavity_2d.jl
@@ -100,4 +100,5 @@ sol = solve(ode, RDPK3SpFSAL49(),
             abstol=1e-6, # Default abstol is 1e-6 (may needs to be tuned to prevent boundary penetration)
             reltol=1e-4, # Default reltol is 1e-3 (may needs to be tuned to prevent boundary penetration)
             dtmax=1e-2, # Limit stepsize to prevent crashing
+            maxiters=Int(1e7),
             save_everystep=false, callback=callbacks);
