+++ src/callbacks/steady_state.jl
@@ -1,21 +1,30 @@
-# Inspired by [Trixi.jl](https://github.com/trixi-framework/Trixi.jl).
 """
     SteadyStateCallback(; abstol=1.0e-8, reltol=1.0e-6)
 
-Terminates the integration when the [`residual_steady_state(du, equations)`](@ref)
-falls below the threshold specified by `abstol, reltol`.
+Terminates the integration when the residual falls below the threshold
+specified by `abstol, reltol`.
 """
-mutable struct SteadyStateCallback{RealT <: Real}
-    abstol::RealT
-    reltol::RealT
-    previous_ekin::Vector{Float64}
-    interval_size::Int
+mutable struct SteadyStateCallback{RealT <: Real, F}
+    abstol           :: RealT
+    reltol           :: RealT
+    previous_ekin    :: Vector{Float64}
+    interval_size    :: Int
+    exclude_boundary :: Bool
+    data             :: Dict{String, Vector{Any}}
+    filename         :: String
+    output_directory :: String
+    func             :: F
 end
 
-function SteadyStateCallback(; abstol=1.0e-8, reltol=1.0e-6, interval_size::Integer=10)
+function SteadyStateCallback(; abstol=1.0e-8, reltol=1.0e-6, interval_size::Integer=10,
+                             exclude_boundary=true, output_directory="out",
+                             filename="values", funcs...)
     abstol, reltol = promote(abstol, reltol)
 
-    steady_state_callback = SteadyStateCallback(abstol, reltol, [Inf64], interval_size)
+    steady_state_callback = SteadyStateCallback(abstol, reltol, [Inf64], interval_size,
+                                                exclude_boundary,
+                                                Dict{String, Vector{Any}}(),
+                                                filename, output_directory, funcs)
 
     DiscreteCallback(steady_state_callback, steady_state_callback,
                      save_positions=(false, false))
@@ -44,36 +53,64 @@ function Base.show(io::IO, ::MIME"text/plain",
     end
 end
 
-# affect!
+# `affect!`
 (::SteadyStateCallback)(integrator) = terminate!(integrator)
 
-# the condition
+# `condition`
 function (steady_state_callback::SteadyStateCallback)(vu_ode, t, integrator)
-    (; abstol, reltol, previous_ekin, interval_size) = steady_state_callback
+    (; abstol, reltol, previous_ekin, interval_size,
+    output_directory, filename) = steady_state_callback
     v_ode, u_ode = vu_ode.x
     semi = integrator.p
 
-    # Calculate kinetic energy
+    filenames = system_names(semi.systems)
+
+    # Calculate kinetic energy and store custom data
     ekin = 0.0
     foreach_system(semi) do system
-        v = wrap_v(v_ode, system, semi)
+        if system isa BoundarySystem && steady_state_callback.exclude_boundary
 
+            # Exclude boundary systems
+            return
+        end
+
+        v = wrap_v(v_ode, system, semi)
+        u = wrap_u(u_ode, system, semi)
+
+        # Calculate kintetic energy
         for particle in each_moving_particle(system)
             velocity = current_velocity(v, system, particle)
             ekin += 0.5 * system.mass[particle] * dot(velocity, velocity)
         end
+
+        system_index = system_indices(system, semi)
+        # Store custom data
+        for (key, f) in steady_state_callback.func
+            result = f(v, u, t, system)
+            if result !== nothing
+                add_entry!(steady_state_callback, string(key), t, result,
+                           filenames[system_index])
+            end
+        end
     end
+
+    push!(previous_ekin, ekin)
 
     terminate = false
 
     if integrator.stats.naccept > interval_size
         popfirst!(previous_ekin)
 
[…]
