+++ src/callbacks/steady_state.jl
@@ -4,38 +4,54 @@
 Terminates the integration when the residual falls below the threshold
 specified by `abstol, reltol`.
 """
-mutable struct SteadyStateCallback{RealT <: Real, F}
-    abstol           :: RealT
-    reltol           :: RealT
-    previous_ekin    :: Vector{Float64}
-    interval_size    :: Int
-    exclude_boundary :: Bool
-    data             :: Dict{String, Vector{Any}}
-    filename         :: String
-    output_directory :: String
-    func             :: F
+mutable struct SteadyStateCallback{I, RealT <: Real}
+    interval      :: I
+    abstol        :: RealT
+    reltol        :: RealT
+    previous_ekin :: Vector{Float64}
+    interval_size :: Int
 end
 
-function SteadyStateCallback(; abstol=1.0e-8, reltol=1.0e-6, interval_size::Integer=10,
-                             exclude_boundary=true, output_directory="out",
-                             filename="values", funcs...)
+function SteadyStateCallback(; interval::Integer=0, dt=0.0, abstol=1.0e-8, reltol=1.0e-6,
+                             interval_size::Integer=10)
     abstol, reltol = promote(abstol, reltol)
 
-    steady_state_callback = SteadyStateCallback(abstol, reltol, [Inf64], interval_size,
-                                                exclude_boundary,
-                                                Dict{String, Vector{Any}}(),
-                                                filename, output_directory, funcs)
+    if dt > 0 && interval > 0
+        throw(ArgumentError("setting both `interval` and `dt` is not supported"))
+    end
 
-    DiscreteCallback(steady_state_callback, steady_state_callback,
-                     save_positions=(false, false))
+    if dt > 0
+        interval = Float64(dt)
+    end
+
+    steady_state_callback = SteadyStateCallback(interval, abstol, reltol, [Inf64],
+                                                interval_size)
+
+    if dt > 0
+        return PeriodicCallback(steady_state_callback, dt, save_positions=(false, false),
+                                final_affect=true)
+    else
+        return DiscreteCallback(steady_state_callback, steady_state_callback,
+                                save_positions=(false, false))
+    end
 end
 
 function Base.show(io::IO, cb::DiscreteCallback{<:Any, <:SteadyStateCallback})
     @nospecialize cb # reduce precompilation time
 
-    steady_state_callback = cb.affect!
-    print(io, "SteadyStateCallback(abstol=", steady_state_callback.abstol, ", ",
-          "reltol=", steady_state_callback.reltol, ")")
+    cb_ = cb.affect!
+
+    print(io, "SteadyStateCallback(abstol=", cb_.abstol, ", ", "reltol=", cb_.reltol, ")")
+end
+
+function Base.show(io::IO,
+                   cb::DiscreteCallback{<:Any,
+                                        <:PeriodicCallbackAffect{<:SteadyStateCallback}})
+    @nospecialize cb # reduce precompilation time
+
+    cb_ = cb.affect!.affect!
+
+    print(io, "SteadyStateCallback(abstol=", cb_.abstol, ", reltol=", cb_.reltol, ")")
 end
 
 function Base.show(io::IO, ::MIME"text/plain",
@@ -45,10 +61,30 @@ function Base.show(io::IO, ::MIME"text/plain",
     if get(io, :compact, false)
         show(io, cb)
     else
-        steady_state_callback = cb.affect!
+        cb_ = cb.affect!
 
-        setup = ["absolute tolerance" => steady_state_callback.abstol,
-            "relative tolerance" => steady_state_callback.reltol]
+        setup = ["absolute tolerance" => cb_.abstol,
+            "relative tolerance" => cb_.reltol,
+            "interval" => cb_.interval,
[…]
