### 2024-08-13 8543619
save size in pp_callback
