+++ examples/fluid/sphere_surface_tension_wall_2d.jl
@@ -9,6 +9,7 @@ fluid_particle_spacing = 0.0025
 
 # ==========================================================================================
 # ==== Experiment Setup
+gravity = 9.81
 tspan = (0.0, 0.5)
 
 # Boundary geometry and initial fluid particle positions
@@ -16,6 +17,9 @@ tank_size = (0.5, 0.1)
 
 fluid_density = 1000.0
 sound_speed = 120.0
+
+state_equation = StateEquationCole(; sound_speed, reference_density=fluid_density,
+                                   exponent=1)
 
 sphere_radius = 0.05
 
@@ -25,7 +29,8 @@ sphere1 = SphereShape(fluid_particle_spacing, sphere_radius, sphere1_center,
 
 # ==========================================================================================
 # ==== Fluid
-fluid_smoothing_length = 1.0 * fluid_particle_spacing
+fluid_smoothing_length = 1.0 * fluid_particle_spacing - eps()
+fluid_smoothing_kernel = SchoenbergCubicSplineKernel{2}()
 
 # For perfect wetting
 # nu = 0.0005
@@ -35,10 +40,21 @@ nu = 0.001
 alpha = 8 * nu / (fluid_smoothing_length * sound_speed)
 # `adhesion_coefficient = 1.0` and `surface_tension_coefficient = 0.01` for perfect wetting
 # `adhesion_coefficient = 0.001` and `surface_tension_coefficient = 2.0` for no wetting
+
+viscosity = ArtificialViscosityMonaghan(alpha=alpha, beta=0.0)
+sphere_surface_tension = WeaklyCompressibleSPHSystem(sphere1, ContinuityDensity(),
+                                                     state_equation, fluid_smoothing_kernel,
+                                                     fluid_smoothing_length,
+                                                     viscosity=viscosity,
+                                                     acceleration=(0.0, -gravity),
+                                                     surface_tension=SurfaceTensionAkinci(surface_tension_coefficient=2.0),
+                                                     correction=AkinciFreeSurfaceCorrection(fluid_density))
+
 trixi_include(@__MODULE__,
               joinpath(examples_dir(), "fluid", "falling_water_spheres_2d.jl"),
               sphere=nothing, sphere1=sphere1, adhesion_coefficient=0.001,
-              wall_viscosity=4.0 * nu, surface_tension_coefficient=2.0, alpha=alpha,
+              wall_viscosity=4.0 * nu, surface_tension_coefficient=0.9, alpha=alpha,
               sound_speed=sound_speed, fluid_density=fluid_density, nu=nu,
               fluid_particle_spacing=fluid_particle_spacing, tspan=tspan,
-              tank_size=tank_size)
+              tank_size=tank_size, fluid_smoothing_length=fluid_smoothing_length,
+              sphere_surface_tension=sphere_surface_tension)
