+++ src/callbacks/post_process.jl
@@ -156,7 +156,7 @@ function Base.show(io::IO, ::MIME"text/plain",
             "output directory" => callback.output_directory,
             "append timestamp" => callback.append_timestamp ? "yes" : "no",
             "write json file" => callback.write_csv ? "yes" : "no",
-            "write csv file" => callback.write_json ? "yes" : "no",
+            "write csv file" => callback.write_json ? "yes" : "no"
         ]
 
         for (i, key) in enumerate(keys(callback.func))
@@ -193,7 +193,7 @@ function Base.show(io::IO, ::MIME"text/plain",
             "output directory" => callback.output_directory,
             "append timestamp" => callback.append_timestamp ? "yes" : "no",
             "write json file" => callback.write_csv ? "yes" : "no",
-            "write csv file" => callback.write_json ? "yes" : "no",
+            "write csv file" => callback.write_json ? "yes" : "no"
         ]
 
         for (i, key) in enumerate(keys(callback.func))
@@ -269,12 +269,14 @@ function (pp::PostprocessCallback)(integrator)
     u_modified!(integrator, false)
 end
 
-@inline function backup_condition(cb::PostprocessCallback{Int}, integrator)
+@inline function backup_condition(cb::Union{PostprocessCallback{Int},
+                                            TableDataSavingCallback{Int}}, integrator)
     return integrator.stats.naccept > 0 &&
            round(integrator.stats.naccept / cb.interval) % cb.write_file_interval == 0
 end
 
-@inline function backup_condition(cb::PostprocessCallback, integrator)
+@inline function backup_condition(cb::Union{PostprocessCallback, TableDataSavingCallback},
+                                  integrator)
     return integrator.stats.naccept > 0 &&
            round(Int, integrator.t / cb.interval) % cb.write_file_interval == 0
 end
