+++ src/general/system.jl
@@ -150,14 +150,14 @@ reset_callback_flag!(system) = system
 
 # Assuming a constant particle spacing one can calculate the number of neighbors within the
 # compact support for an undisturbed particle distribution.
-function neighbor_number(::Val{D}, particle_spacing, compact_support) where {D}
+function number_density(::Val{D}, particle_spacing, compact_support) where {D}
     throw(ArgumentError("Unsupported dimension: $D"))
 end
 
-@inline function neighbor_number(::Val{2}, particle_spacing, compact_support)
+@inline function number_density(::Val{2}, particle_spacing, compact_support)
     return pi * compact_support^2 / particle_spacing^2
 end
 
-@inline @fastpow function neighbor_number(::Val{3}, particle_spacing, compact_support)
+@inline @fastpow function number_density(::Val{3}, particle_spacing, compact_support)
     return 4.0 / 3.0 * pi * compact_support^3 / particle_spacing^3
 end
