+++ src/schemes/fluid/surface_tension.jl
@@ -1,4 +1,5 @@
-abstract type AkinciTypeSurfaceTension end
+abstract type SurfaceTensionModel end
+abstract type AkinciTypeSurfaceTension <: SurfaceTensionModel end
 
 @doc raw"""
     CohesionForceAkinci(surface_tension_coefficient=1.0)
