+++ src/schemes/fluid/fluid.jl
@@ -75,6 +75,23 @@ function calculate_dt(v_ode, u_ode, cfl_number, system::FluidSystem)
     return min(dt_viscosity, dt_acceleration, dt_sound_speed)
 end
 
+@inline function surface_tension_model(system::FluidSystem)
+    return system.surface_tension
+end
+
+@inline function surface_tension_model(system)
+    return nothing
+end
+
+@inline function surface_normal_method(system::FluidSystem)
+    return system.surface_normal_method
+end
+
+@inline function surface_normal_method(system)
+    return nothing
+end
+
+
 include("pressure_acceleration.jl")
 include("viscosity.jl")
 include("surface_tension.jl")
