### 2024-08-15 1dfc4a3
remove double velocity initialization
