+++ src/schemes/fluid/transport_velocity.jl
@@ -50,7 +50,6 @@ function write_v0!(v0, system::FluidSystem, ::TransportVelocityAdami)
     for particle in eachparticle(system)
         # Write particle velocities
         for dim in 1:ndims(system)
-            v0[dim, particle] = initial_condition.velocity[dim, particle]
             v0[ndims(system) + dim, particle] = initial_condition.velocity[dim, particle]
         end
     end
