+++ validation/taylor_green_vortex_2d/validation_taylor_green_vortex_2d.jl
@@ -2,7 +2,7 @@ using TrixiParticles
 
 # ==========================================================================================
 # ==== Resolution
-particle_spacings = [0.02, 0.015, 0.01, 0.005]
+particle_spacings = [0.02, 0.01, 0.005]
 
 # ==========================================================================================
 # ==== Experiment Setup
