+++ src/callbacks/post_process.jl
@@ -156,7 +156,7 @@ function Base.show(io::IO, ::MIME"text/plain",
             "output directory" => callback.output_directory,
             "append timestamp" => callback.append_timestamp ? "yes" : "no",
             "write json file" => callback.write_csv ? "yes" : "no",
-            "write csv file" => callback.write_json ? "yes" : "no"
+            "write csv file" => callback.write_json ? "yes" : "no",
         ]
 
         for (i, key) in enumerate(keys(callback.func))
@@ -193,7 +193,7 @@ function Base.show(io::IO, ::MIME"text/plain",
             "output directory" => callback.output_directory,
             "append timestamp" => callback.append_timestamp ? "yes" : "no",
             "write json file" => callback.write_csv ? "yes" : "no",
-            "write csv file" => callback.write_json ? "yes" : "no"
+            "write csv file" => callback.write_json ? "yes" : "no",
         ]
 
         for (i, key) in enumerate(keys(callback.func))
