+++ src/schemes/fluid/weakly_compressible_sph/state_equations.jl
@@ -49,3 +49,5 @@ function inverse_state_equation(state_equation::StateEquationCole, pressure)
 
     return reference_density * tmp^(1 / exponent)
 end
+
+@inline sound_speed(eos::StateEquationCole) = eos.sound_speed
