+++ src/schemes/fluid/fluid.jl
@@ -91,7 +91,6 @@ end
     return nothing
 end
 
-
 include("pressure_acceleration.jl")
 include("viscosity.jl")
 include("surface_tension.jl")
