+++ src/schemes/boundary/system.jl
@@ -156,9 +156,10 @@ end
 create_cache_boundary(::Nothing, initial_condition) = (;)
 
 function create_cache_boundary(::BoundaryMovement, initial_condition)
+    initial_coordinates = copy(initial_condition.coordinates)
     velocity = zero(initial_condition.velocity)
     acceleration = zero(initial_condition.velocity)
-    return (; velocity, acceleration)
+    return (; velocity, acceleration, initial_coordinates)
 end
 
 function Base.show(io::IO, system::BoundarySPHSystem)
@@ -194,11 +195,18 @@ timer_name(::Union{BoundarySPHSystem, BoundaryDEMSystem}) = "boundary"
     eltype(system.coordinates)
 end
 
-# This does not account for moving boundaries, but it's only used to initialize the
-# neighborhood search, anyway.
-@inline function initial_coordinates(system::Union{BoundarySPHSystem, BoundaryDEMSystem})
-    system.coordinates
+@inline function initial_coordinates(system::BoundarySPHSystem)
+    initial_coordinates(system::BoundarySPHSystem, system.movement)
 end
+
+@inline initial_coordinates(system::BoundarySPHSystem, nothing) = sytem.coordinates
+
+@inline function initial_coordinates(system::BoundarySPHSystem, movement)
+
+    return system.cache.initial_coordinates
+end
+
+@inline initial_coordinates(system::BoundaryDEMSystem) = system.coordinates
 
 function (movement::BoundaryMovement)(system, t)
     (; coordinates, cache) = system
