+++ examples/fluid/drops_on_box.jl
@@ -1,0 +1,140 @@
+# In this example two circles of water drop to the floor demonstrating the difference
+# between the behavior with and without surface tension modelling.
+using TrixiParticles
+using OrdinaryDiffEq
+
+# ==========================================================================================
+# ==== Resolution
+fluid_particle_spacing = 0.0005
+
+boundary_layers = 3
+spacing_ratio = 1
+
+# ==========================================================================================
+# ==== Experiment Setup
+gravity = 9.81
+tspan = (0.0, 1.0)
+
+# Boundary geometry and initial fluid particle positions
+initial_fluid_size = (0.0, 0.0)
+tank_size = (1.0, 0.5)
+
+fluid_density = 1000.0
+sound_speed = 100
+state_equation = StateEquationCole(; sound_speed, reference_density=fluid_density,
+                                   exponent=1)
+
+tank = RectangularTank(fluid_particle_spacing, initial_fluid_size, tank_size, fluid_density,
+                       n_layers=boundary_layers, spacing_ratio=spacing_ratio,
+                       faces=(true, true, true, false),
+                       acceleration=(0.0, -gravity), state_equation=state_equation)
+
+
+box = RectangularShape(fluid_particle_spacing, (300, 125), (0.485, 0.0), density=fluid_density)
+
+# box = SphereShape(fluid_particle_spacing, 1.0, (0.5, 0.0),
+#                       fluid_density, sphere_type=VoxelSphere(), velocity=(0.0, 0.0))
+
+sphere_radius = 0.0025
+
+sphere1_center = (0.5, 0.05)
+sphere2_center = (0.5, 0.1)
+sphere3_center = (0.5, 0.15)
+sphere4_center = (0.5, 0.2)
+sphere1 = SphereShape(fluid_particle_spacing, sphere_radius, sphere1_center,
+                      fluid_density, sphere_type=VoxelSphere(), velocity=(0.0, -3.0))
+sphere2 = SphereShape(fluid_particle_spacing, sphere_radius, sphere2_center,
+                      fluid_density, sphere_type=VoxelSphere(), velocity=(0.0, -3.0))
+sphere3 = SphereShape(fluid_particle_spacing, sphere_radius, sphere3_center,
+                      fluid_density, sphere_type=VoxelSphere(), velocity=(0.0, -3.0))
+sphere4 = SphereShape(fluid_particle_spacing, sphere_radius, sphere4_center,
+                      fluid_density, sphere_type=VoxelSphere(), velocity=(0.0, -3.0))
+
+water = union(sphere1, sphere2, sphere3, sphere4)
+
+# ==========================================================================================
+# ==== Fluid
+fluid_smoothing_length = 1.0 * fluid_particle_spacing
+fluid_smoothing_kernel = SchoenbergCubicSplineKernel{2}()
+
+fluid_density_calculator = ContinuityDensity()
+
+nu = 0.00089
+alpha = 0.75*8 * nu / (fluid_smoothing_length * sound_speed)
+# viscosity = ViscosityAdami(nu=nu)
+viscosity = ArtificialViscosityMonaghan(alpha=alpha, beta=0.0)
+
+# density_diffusion = DensityDiffusionAntuono(sphere2, delta=0.1)
+
+sphere_surface_tension = WeaklyCompressibleSPHSystem(water, fluid_density_calculator,
+                                                     state_equation, fluid_smoothing_kernel,
+                                                     fluid_smoothing_length,
+                                                     viscosity=viscosity,
+                                                     acceleration=(0.0, -gravity),
+                                                     surface_tension=SurfaceTensionAkinci(surface_tension_coefficient=0.1*0.011),
+                                                     correction=AkinciFreeSurfaceCorrection(fluid_density),
[…]
