+++ examples/fluid/callibration_water_drop_shape.jl
@@ -1,0 +1,255 @@
+# In this example we try to approach the static shape of a water droplet on a horizontal plane.
+# The shape of a static droplet can be calculated from the Young-Laplace equation.
+using TrixiParticles
+using OrdinaryDiffEq
+
+# ==========================================================================================
+# ==== Resolution
+fluid_particle_spacing = 0.0001
+
+boundary_layers = 4
+spacing_ratio = 1
+
+# ==========================================================================================
+# ==== Experiment Setup
+gravity = 9.81
+tspan = (0.0, 5.0)
+
+# Boundary geometry and initial fluid particle positions
+initial_fluid_size = (0.0, 0.0)
+tank_size = (0.05, 0.01)
+
+fluid_density = 1000.0
+sound_speed = 50
+
+state_equation = StateEquationCole(; sound_speed, reference_density=fluid_density,
+                                   exponent=1)
+
+tank = RectangularTank(fluid_particle_spacing, initial_fluid_size, tank_size, fluid_density,
+                       n_layers=boundary_layers, spacing_ratio=spacing_ratio,
+                       faces=(true, true, true, false),
+                       acceleration=(0.0, -gravity), state_equation=state_equation)
+
+sphere_radius = 0.0015
+
+
+sphere1_center = (0.5, sphere_radius)
+sphere1 = SphereShape(fluid_particle_spacing, sphere_radius, sphere1_center,
+                      fluid_density, sphere_type=VoxelSphere())
+
+# ==========================================================================================
+# ==== Fluid
+# fluid_smoothing_length = 1.2 * fluid_particle_spacing
+# fluid_smoothing_kernel = SchoenbergCubicSplineKernel{2}()
+
+fluid_smoothing_length = 2.5 * fluid_particle_spacing
+fluid_smoothing_kernel = WendlandC2Kernel{2}()
+
+fluid_smoothing_length_2 = 2.5 * fluid_particle_spacing
+fluid_smoothing_kernel_2 = WendlandC2Kernel{2}()
+
+fluid_density_calculator = ContinuityDensity()
+
+# water at 20C
+#nu=0.00089
+nu = 0.00089
+# too much 0.00089 -> 0.00045 -> 0.0002 -> 0.0001
+# no impact 0.00005
+
+# the following term only holds for 2d sims
+# alpha = 8 * nu / (fluid_smoothing_length * sound_speed)
+viscosity = ViscosityAdami(nu=nu)
+# density_diffusion = DensityDiffusionAntuono(sphere2, delta=0.1)
+
+# with increased smoothing length surface_tension is too small
+sphere_surface_tension = WeaklyCompressibleSPHSystem(sphere1, fluid_density_calculator,
+                                                     state_equation, fluid_smoothing_kernel,
+                                                     fluid_smoothing_length,
[…]
