+++ src/schemes/boundary/open_boundary/boundary_zones.jl
@@ -77,6 +77,7 @@ struct InFlow{NDIMS, IC, S, ZO, ZW, FD}
     zone_origin       :: ZO
     zone_width        :: ZW
     flow_direction    :: FD
+    nparticles        :: Int64
 
     function InFlow(; plane, flow_direction, density, particle_spacing,
                     initial_condition=nothing, extrude_geometry=nothing,
@@ -127,10 +128,12 @@ struct InFlow{NDIMS, IC, S, ZO, ZW, FD}
         # This check is only necessary when `initial_condition` or `extrude_geometry` are passed.
         ic = remove_outside_particles(initial_condition, spanning_set_, zone_origin)
 
+        nparticles = length(initial_condition.mass)
+
         return new{NDIMS, typeof(ic), typeof(spanning_set_), typeof(zone_origin),
                    typeof(zone_width),
                    typeof(flow_direction_)}(ic, spanning_set_, zone_origin, zone_width,
-                                            flow_direction_)
+                                            flow_direction_, nparticles)
     end
 end
 
@@ -213,37 +216,41 @@ struct OutFlow{NDIMS, IC, S, ZO, ZW, FD}
     zone_origin       :: ZO
     zone_width        :: ZW
     flow_direction    :: FD
+    nparticles        :: Int64
 
     function OutFlow(; plane, flow_direction, density, particle_spacing,
                      initial_condition=nothing, extrude_geometry=nothing,
                      open_boundary_layers::Integer)
-        if open_boundary_layers <= 0
-            throw(ArgumentError("`open_boundary_layers` must be positive and greater than zero"))
-        end
+        # if open_boundary_layers <= 0
+        #     throw(ArgumentError("`open_boundary_layers` must be positive and greater than zero"))
+        # end
+        NDIMS = length(flow_direction)
+        ELTYPE = eltype(flow_direction)
 
         # Unit vector pointing in downstream direction
         flow_direction_ = normalize(SVector(flow_direction...))
 
         # Sample particles in boundary zone
-        if isnothing(initial_condition) && isnothing(extrude_geometry)
+        if isnothing(initial_condition) && isnothing(extrude_geometry) && open_boundary_layers > 0
             initial_condition = TrixiParticles.extrude_geometry(plane; particle_spacing,
                                                                 density,
                                                                 direction=flow_direction_,
                                                                 n_extrude=open_boundary_layers)
-        elseif !isnothing(extrude_geometry)
+        elseif !isnothing(extrude_geometry) && open_boundary_layers > 0
             initial_condition = TrixiParticles.extrude_geometry(extrude_geometry;
                                                                 particle_spacing, density,
                                                                 direction=-flow_direction_,
                                                                 n_extrude=open_boundary_layers)
+        else
+            initial_condition = InitialCondition(; coordinates=zeros(ELTYPE, NDIMS, 0), density=zero(ELTYPE), particle_spacing=zero(ELTYPE))
         end
 
-        NDIMS = ndims(initial_condition)
-        ELTYPE = eltype(initial_condition)
-
-        zone_width = open_boundary_layers * initial_condition.particle_spacing
+        zone_width = max(open_boundary_layers * particle_spacing, particle_spacing)
 
         # Vectors spanning the boundary zone/box
         spanning_set, zone_origin = calculate_spanning_vectors(plane, zone_width)
+        println(spanning_set, typeof(spanning_set))
+        println(zone_origin, typeof(zone_origin))
 
         # First vector of `spanning_vectors` is normal to the outflow plane.
[…]
