+++ src/TrixiParticles.jl
@@ -68,7 +68,8 @@ export ArtificialViscosityMonaghan, ViscosityAdami, ViscosityMorris
 export DensityDiffusion, DensityDiffusionMolteniColagrossi, DensityDiffusionFerrari,
        DensityDiffusionAntuono
 export BoundaryModelMonaghanKajtar, BoundaryModelDummyParticles, AdamiPressureExtrapolation,
-       PressureMirroring, PressureZeroing, BoundaryModelLastiwka
+       PressureMirroring, PressureZeroing
+export BoundaryModelLastiwka, BasicOutlet
 export BoundaryMovement
 export examples_dir, validation_dir, trixi_include
 export trixi2vtk
