### 2024-08-20 bad7de8
add basic struct
