### 2024-08-20 7841461
make equation available
