+++ src/visualization/write2vtk.jl
@@ -256,10 +256,18 @@ function write2vtk!(vtk, v, u, t, system::FluidSystem; write_meta_data=true)
                 vtk["correction_rho0"] = system.correction.rho0
             end
             vtk["state_equation"] = type2string(system.state_equation)
-            vtk["state_equation_rho0"] = system.state_equation.reference_density
-            vtk["state_equation_pa"] = system.state_equation.background_pressure
-            vtk["state_equation_c"] = system.state_equation.sound_speed
-            vtk["state_equation_exponent"] = system.state_equation.exponent
+            if system.state_equation isa StateEquationCole
+                vtk["state_equation_rho0"] = system.state_equation.reference_density
+                vtk["state_equation_pa"] = system.state_equation.background_pressure
+                vtk["state_equation_c"] = system.state_equation.sound_speed
+                vtk["state_equation_exponent"] = system.state_equation.exponent
+            end
+
+            if system.state_equation isa StateEquationIdealGas
+                vtk["state_equation_gamma"] = system.state_equation.gamma
+                vtk["state_equation_gas_c"] = system.state_equation.gas_constant
+                vtk["state_equation_temperature"] = system.state_equation.temperature
+            end
 
             vtk["solver"] = "WCSPH"
         else
