+++ examples/fluid/dam_break_oil_film_2d.jl
@@ -48,7 +48,7 @@ for i in axes(oil.coordinates, 2)
 end
 
 oil_state_equation = StateEquationCole(; sound_speed, reference_density=oil_density,
-                  exponent=1, clip_negative_pressure=false)
+                                       exponent=1, clip_negative_pressure=false)
 oil_system = WeaklyCompressibleSPHSystem(oil, fluid_density_calculator,
                                          oil_state_equation, smoothing_kernel,
                                          smoothing_length, viscosity=oil_viscosity,
