+++ src/schemes/fluid/weakly_compressible_sph/density_diffusion.jl
@@ -240,7 +240,7 @@ end
 
     (; delta) = density_diffusion
     (; smoothing_length, state_equation) = particle_system
-    sound_speed = sound_speed(state_equation)
+    sound_speed_ = sound_speed(state_equation)
 
     volume_b = m_b / rho_b
 
@@ -248,7 +248,7 @@ end
                                 particle_system, neighbor_system, particle, neighbor)
     density_diffusion_term = dot(psi, grad_kernel) * volume_b
 
-    dv[end, particle] += delta * smoothing_length * sound_speed * density_diffusion_term
+    dv[end, particle] += delta * smoothing_length * sound_speed_ * density_diffusion_term
 end
 
 # Density diffusion `nothing` or interaction other than fluid-fluid
