### 2024-08-22 9c8beaf
tighter error bound
