+++ test/preprocessing/geometries/geometries.jl
@@ -140,6 +140,6 @@
         sorted_unique_x = sort(unique(x))
         sorted_unique_copy_x = TrixiParticles.unique_sorted(copy(x))
 
-        @test isapprox(sorted_unique_copy_x, sorted_unique_x; rtol=1e-9, atol=1e-9)
+        @test isapprox(sorted_unique_copy_x, sorted_unique_x; atol=eps())
     end
 end
