### 2024-08-22 eeedb3d
improve scripts
