### 2024-08-23 fbaf108
check dimensionality of reference functions
