### 2024-08-23 e720529
propagate characteristics
