+++ src/schemes/boundary/open_boundary/method_of_characteristics.jl
@@ -8,7 +8,8 @@ about the method see [description below](@ref method_of_characteristics).
 """
 struct BoundaryModelLastiwka end
 
-@inline function update_quantities!(system, boundary_model::BoundaryModelLastiwka,
+# Called from update callback
+@inline function update_boundary_quantities!(system, boundary_model::BoundaryModelLastiwka,
                                     v, u, v_ode, u_ode, semi, t)
     (; density, pressure, cache, flow_direction,
     reference_velocity, reference_pressure, reference_density) = system
@@ -45,19 +46,39 @@ struct BoundaryModelLastiwka end
     return system
 end
 
+# Called from semidiscretization
 function update_final!(system, ::BoundaryModelLastiwka, v, u, v_ode, u_ode, semi, t)
     @trixi_timeit timer() "evaluate characteristics" begin
         evaluate_characteristics!(system, v, u, v_ode, u_ode, semi, t)
     end
 end
 
+
+function evaluate_characteristics!(system, v, u, v_ode, u_ode, semi, t)
+    (; cache) = system
+    (; previous_characteristics_inited) = cache
+
+    # Propagate characteristics through the open boundary
+    if !previous_characteristics_inited[1]
+
+        for iteration in 1:4
+            calc_characteristics!(system, v, u, v_ode, u_ode, semi, t)
+        end
+
+        previous_characteristics_inited[1] = true
+    else
+        calc_characteristics!(system, v, u, v_ode, u_ode, semi, t)
+    end
+
+end
+
 # ==== Characteristics
 # J1: Associated with convection and entropy and propagates at flow velocity.
 # J2: Propagates downstream to the local flow
 # J3: Propagates upstream to the local flow
-function evaluate_characteristics!(system, v, u, v_ode, u_ode, semi, t)
+function calc_characteristics!(system, v, u, v_ode, u_ode, semi, t)
     (; volume, cache, boundary_zone) = system
-    (; characteristics, previous_characteristics) = cache
+    (; characteristics, previous_characteristics, previous_characteristics_inited) = cache
 
     for particle in eachparticle(system)
         previous_characteristics[1, particle] = characteristics[1, particle]
