### 2024-09-16 05e401e
Add YouTube video of JuliaCon talk to README
