### 2024-09-19 b6e7346
implement suggestions
