+++ paper/paper.md
@@ -39,7 +39,7 @@ bibliography: paper.bib
 
 # Summary
 
-TrixiParticles.jl is a Julia-based open-source package for particle-based multiphysics simulations and part of the Trixi Framework [@schlottkelakemper2020trixi]. It handles complex geometries and specialized applications, such as computational fluid dynamics (CFD) and structural dynamics, by providing a versatile platform for particle-based methods. TrixiParticles.jl allows for the straightforward addition of new particle systems and their interactions, facilitating the setup of coupled multiphysics simulations such as fluid-structure interaction (FSI). Furthermore, simulations are set up directly with Julia code, simplifying the integration of custom functionalities and promoting rapid prototyping.
+TrixiParticles.jl is a Julia-based open-source package for particle-based multiphysics simulations and part of the Trixi Framework @schlottkelakemper2020trixi. It handles complex geometries and specialized applications, such as computational fluid dynamics (CFD) and structural dynamics, by providing a versatile platform for particle-based methods. TrixiParticles.jl allows for the straightforward addition of new particle systems and their interactions, facilitating the setup of coupled multiphysics simulations such as fluid-structure interaction (FSI). Furthermore, simulations are set up directly with Julia code, simplifying the integration of custom functionalities and promoting rapid prototyping.
 
 Here, we give a brief overview of the software package TrixiParticles.jl, starting with the scientific background before going on to describe the functionality and benefit in more detail.
 Finally, exemplary results and implemented features are briefly presented.
@@ -51,9 +51,9 @@ For example, they involve complex geometries, free surfaces,
 deformable boundaries, and moving material interfaces, as well as the coupling of multiple systems with different mathematical models.
 
 One way to address these challenges is to use particle-based methods, in which the particles either represent physical particles or mathematical interpolation points.
-The former case refers to methods that model separate, discrete particles with rotational degrees of freedom such as the Discrete Element Method (DEM) proposed by [@Cundall:1979],
+The former case refers to methods that model separate, discrete particles with rotational degrees of freedom such as the Discrete Element Method (DEM) proposed by @Cundall:1979,
 whereas the latter case refers to methods such as Smoothed Particle Hydrodynamics (SPH), which is a numerical discretization method for solving problems in continuum mechanics.
-SPH was originally developed by [@Monaghan:1977] to simulate astrophysical applications and is now widely used to simulate CFD, structural mechanics, and even heat conduction problems.
+SPH was originally developed by @Monaghan:1977 to simulate astrophysical applications and is now widely used to simulate CFD, structural mechanics, and even heat conduction problems.
 
 The Lagrangian formalism in particle-based methods allows particles to move along a velocity field without any connection to neighboring particles,
 thus eliminating the need for a mesh to discretize the simulation domain.
@@ -62,9 +62,9 @@ By representing each material with its own set of particles,
 coupling multiple different physical systems into a single multiphysics setup is straightforward.
 In addition, particle-based methods are inherently suited to simulating free surfaces, material interfaces, and moving boundaries.
 
-There are several open-source software projects specialized for SPH methods, including DualSPHysics [@Dominguez:2021], SPlisHSPlasH [@Bender], and SPHinXsys [@Zhang:2021],
-written in C++, and  PySPH [@Ramachandran:2021], written in Python.
[…]
