### 2024-09-24 8587c75
Revert "remove square brackets"
This reverts commit 05d0af74e8c1d219325ae135c73ed4d8f96f5356.
