### 2024-09-24 5d98892
fix cite in text
