### 2024-09-24 8ecf5e6
[paper] Fix formatting of citations (#616)
* [paper] Fix formatting of citations

* update `upload-artifact` action

---------

Co-authored-by: LasNikas <niklas.neher@web.de>
