### 2024-09-25 ef33b2f
remove system soundspeed from obs
