### 2024-09-26 941a758
implement suggestions
