### 2024-09-26 bc9933f
terminate simulation with setting `maxiters` to `iter`
