### 2024-09-26 4816551
Fix moving boundaries (#603)
* fix

* typo

* fix again

* implement suggestions
