### 2024-09-27 19effe7
fix equation
