### 2024-09-27 f84cafa
move additonal changes from #584
