+++ src/schemes/boundary/open_boundary/boundary_zones.jl
@@ -1,3 +1,9 @@
+struct BidirectionalFlow end
+
+struct InFlow end
+
+struct OutFlow end
+
 @doc raw"""
     InFlow(; plane, flow_direction, density, particle_spacing,
            initial_condition=nothing, extrude_geometry=nothing,
@@ -71,205 +77,127 @@ inflow = InFlow(; plane=plane_points, particle_spacing=0.1, flow_direction, dens
 !!! warning "Experimental Implementation"
     This is an experimental feature and may change in any future releases.
 """
-struct InFlow{NDIMS, IC, S, ZO, ZW, FD}
+struct BoundaryZone{F, NDIMS, IC, S, ZO, ZW, FD, PN}
     initial_condition :: IC
     spanning_set      :: S
     zone_origin       :: ZO
     zone_width        :: ZW
     flow_direction    :: FD
+    plane_normal      :: PN
 
-    function InFlow(; plane, flow_direction, density, particle_spacing,
-                    initial_condition=nothing, extrude_geometry=nothing,
-                    open_boundary_layers::Integer)
+    function BoundaryZone(; plane, plane_normal, density, particle_spacing,
+                          initial_condition=nothing, extrude_geometry=nothing,
+                          open_boundary_layers::Integer,
+                          boundary_type=:bidirectional_flow)
         if open_boundary_layers <= 0
             throw(ArgumentError("`open_boundary_layers` must be positive and greater than zero"))
         end
 
-        # Unit vector pointing in downstream direction
-        flow_direction_ = normalize(SVector(flow_direction...))
+        # `plane_normal` always points in fluid domain
+        plane_normal_ = normalize(SVector(plane_normal...))
 
-        # Sample particles in boundary zone
-        if isnothing(initial_condition) && isnothing(extrude_geometry)
-            initial_condition = TrixiParticles.extrude_geometry(plane; particle_spacing,
-                                                                density,
-                                                                direction=-flow_direction_,
-                                                                n_extrude=open_boundary_layers)
-        elseif !isnothing(extrude_geometry)
-            initial_condition = TrixiParticles.extrude_geometry(extrude_geometry;
-                                                                particle_spacing,
-                                                                density,
-                                                                direction=-flow_direction_,
-                                                                n_extrude=open_boundary_layers)
+        if boundary_type === :bidirectional_flow
+            boundary_type_ = BidirectionalFlow()
+            flow_direction = nothing
+
+        elseif boundary_type === :inflow
+            # Unit vector pointing in downstream direction
+            flow_direction = plane_normal_
+
+            boundary_type_ = InFlow()
+
+        elseif boundary_type === :outflow
+            # Unit vector pointing in downstream direction
+            flow_direction = -plane_normal_
+
+            boundary_type_ = OutFlow()
         end
 
-        NDIMS = ndims(initial_condition)
-        ELTYPE = eltype(initial_condition)
+        ic, spanning_set_, zone_origin,
+        zone_width = set_up_boundary_zone(plane, plane_normal_, flow_direction, density,
+                                          particle_spacing, initial_condition,
+                                          extrude_geometry, open_boundary_layers;
+                                          boundary_type=boundary_type_)
 
-        zone_width = open_boundary_layers * initial_condition.particle_spacing
-
-        # Vectors spanning the boundary zone/box
-        spanning_set, zone_origin = calculate_spanning_vectors(plane, zone_width)
-
[…]
