+++ test/schemes/boundary/open_boundary/boundary_zone.jl
@@ -18,22 +18,22 @@
             ]
 
             @testset verbose=true "Flow Direction $j" for j in eachindex(flow_directions)
-                inflow = InFlow(; plane=(point_1, point_2), particle_spacing,
-                                flow_direction=flow_directions[j], density=1.0,
-                                open_boundary_layers)
-                outflow = OutFlow(; plane=(point_1, point_2), particle_spacing,
-                                  flow_direction=flow_directions[j], density=1.0,
-                                  open_boundary_layers)
+                inflow = BoundaryZone(; plane=(point_1, point_2), particle_spacing,
+                                      plane_normal=flow_directions[j], density=1.0,
+                                      open_boundary_layers, boundary_type=:inflow)
+                outflow = BoundaryZone(; plane=(point_1, point_2), particle_spacing,
+                                  plane_normal=-flow_directions[j], density=1.0,
+                                  open_boundary_layers, boundary_type=:outflow)
 
                 boundary_zones = [
                     inflow,
                     outflow,
                 ]
 
-                @testset verbose=true "$(nameof(typeof(boundary_zone)))" for boundary_zone in boundary_zones
+                @testset verbose=true "$(TrixiParticles.boundary_type_name(boundary_zone))" for boundary_zone in boundary_zones
                     zone_width = open_boundary_layers *
                                  boundary_zone.initial_condition.particle_spacing
-                    sign_ = (boundary_zone isa InFlow) ? -1 : 1
+                    sign_ = (first(typeof(boundary_zone).parameters) isa TrixiParticles.InFlow) ? -1 : 1
 
                     @test plane_points_1[i] == boundary_zone.zone_origin
                     @test plane_points_2[i] - boundary_zone.zone_origin ==
@@ -42,6 +42,11 @@
                                    normalize(boundary_zone.spanning_set[1]), atol=1e-14)
                     @test isapprox(zone_width, norm(boundary_zone.spanning_set[1]),
                                    atol=1e-14)
+
+                                   if  !(isapprox(sign_ * flow_directions[j],
+                                    normalize(boundary_zone.spanning_set[1]), atol=1e-14))
+                                    TrixiParticles.@autoinfiltrate
+                                   end
                 end
             end
         end
@@ -78,22 +83,23 @@
             ]
 
             @testset verbose=true "Flow Direction $j" for j in eachindex(flow_directions)
-                inflow = InFlow(; plane=(point_1, point_2, point_3), particle_spacing,
-                                flow_direction=flow_directions[j], density=1.0,
-                                open_boundary_layers)
-                outflow = OutFlow(; plane=(point_1, point_2, point_3), particle_spacing,
-                                  flow_direction=flow_directions[j], density=1.0,
-                                  open_boundary_layers)
+                inflow = BoundaryZone(; plane=(point_1, point_2, point_3), particle_spacing,
+                                      plane_normal=flow_directions[j], density=1.0,
+                                      open_boundary_layers, boundary_type=:inflow)
+                outflow = BoundaryZone(; plane=(point_1, point_2, point_3),
+                                       particle_spacing,
+                                       plane_normal=flow_directions[j], density=1.0,
+                                       open_boundary_layers, boundary_type=:outflow)
 
                 boundary_zones = [
                     inflow,
                     outflow,
                 ]
 
-                @testset verbose=true "$(nameof(typeof(boundary_zone)))" for boundary_zone in boundary_zones
+                @testset verbose=true "$(TrixiParticles.boundary_type_name(boundary_zone))" for boundary_zone in boundary_zones
                     zone_width = open_boundary_layers *
                                  boundary_zone.initial_condition.particle_spacing
-                    sign_ = (boundary_zone isa InFlow) ? -1 : 1
[…]
