### 2024-09-28 28589ca
CompatHelper: bump compat for DiffEqCallbacks to 4, (keep existing compat)
