+++ test/schemes/boundary/open_boundary/characteristic_variables.jl
@@ -47,7 +47,8 @@
             ]
 
             @testset "`$(TrixiParticles.boundary_type_name(boundary_zone))`" for boundary_zone in boundary_zones
-                sign_ = (first(typeof(boundary_zone).parameters) === TrixiParticles.InFlow) ? 1 : -1
+                sign_ = (first(typeof(boundary_zone).parameters) === TrixiParticles.InFlow) ?
+                        1 : -1
                 fluid = extrude_geometry(plane_points; particle_spacing, n_extrude=4,
                                          density, pressure,
                                          direction=(sign_ * flow_direction))
