+++ src/visualization/write2vtk.jl
@@ -329,7 +329,7 @@ function write2vtk!(vtk, v, u, t, system::OpenBoundarySPHSystem; write_meta_data
                        for particle in active_particles(system)]
 
     if write_meta_data
-        vtk["boundary_zone"] = first(typeof(boundary_zone).parameters)
+        vtk["boundary_zone"] = type2string(first(typeof(system.boundary_zone).parameters))
         vtk["width"] = round(system.boundary_zone.zone_width, digits=3)
         vtk["velocity_function"] = type2string(system.reference_velocity)
         vtk["pressure_function"] = type2string(system.reference_pressure)
