+++ src/schemes/fluid/fluid.jl
@@ -44,6 +44,8 @@ write_v0!(v0, system, density_calculator) = v0
 
 @inline system_state_equation(system::FluidSystem) = system.state_equation
 
+@inline system_smoothing_length(system::FluidSystem) = system.smoothing_length
+
 function compute_density!(system, u, u_ode, semi, ::ContinuityDensity)
     # No density update with `ContinuityDensity`
     return system
