+++ src/general/buffer.jl
@@ -20,12 +20,12 @@ function allocate_buffer(initial_condition, buffer::SystemBuffer)
     # Initialize particles far away from simulation domain
     coordinates = fill(1e16, ndims(initial_condition), buffer_size)
 
-    if all(rho -> isapprox(rho, first(initial_condition.density), atol=eps(), rtol=eps()),
-           initial_condition.density)
-        density = first(initial_condition.density)
-    else
-        throw(ArgumentError("`initial_condition.density` needs to be constant when using `SystemBuffer`"))
+    if !all(rho -> isapprox(rho, first(initial_condition.density), atol=eps(), rtol=eps()),
+            initial_condition.density)
+        @warn "allocated buffer is using constant values"
     end
+
+    density = first(initial_condition.density)
 
     particle_spacing = initial_condition.particle_spacing
 
