### 2024-10-02 6ed3a55
Add first fin simulation
