+++ examples/fsi/fin_2d.jl
@@ -1,0 +1,164 @@
+using TrixiParticles
+using OrdinaryDiffEq
+
+# ==========================================================================================
+# ==== Resolution
+n_particles_y = 3
+
+# ==========================================================================================
+# ==== Experiment Setup
+gravity = 2.0
+tspan = (0.0, 5.0)
+
+elastic_beam = (length=0.45, thickness=0.02)
+material = (density=1000.0, E=1.0e7, nu=0.4)
+clamp_radius = 0.05
+
+# The structure starts at the position of the first particle and ends
+# at the position of the last particle.
+particle_spacing = elastic_beam.thickness / (n_particles_y - 1)
+
+# Add particle_spacing/2 to the clamp_radius to ensure that particles are also placed on the radius
+fixed_particles = SphereShape(particle_spacing, clamp_radius + particle_spacing / 2,
+                              (0.0, elastic_beam.thickness / 2), material.density,
+                              cutout_min=(0.0, 0.0),
+                              cutout_max=(clamp_radius, elastic_beam.thickness),
+                              tlsph=true)
+
+n_particles_clamp_x = round(Int, clamp_radius / particle_spacing)
+
+# Beam and clamped particles
+n_particles_per_dimension = (round(Int, elastic_beam.length / particle_spacing) +
+                             n_particles_clamp_x + 1, n_particles_y)
+
+# Note that the `RectangularShape` puts the first particle half a particle spacing away
+# from the boundary, which is correct for fluids, but not for solids.
+# We therefore need to pass `tlsph=true`.
+beam = RectangularShape(particle_spacing, n_particles_per_dimension,
+                        (0.0, 0.0), density=material.density, tlsph=true)
+
+solid = union(beam, fixed_particles)
+solid.coordinates .+= 0.2 + 0.08
+
+# Movement function
+movement_function(t) = SVector(0.0, 0.04 * cos(5 * pi * t) - 0.04)
+
+is_moving(t) = true
+
+boundary_movement = BoundaryMovement(movement_function, is_moving)
+
+# ==========================================================================================
+# ==== Resolution
+fluid_particle_spacing = particle_spacing
+
+# Change spacing ratio to 3 and boundary layers to 1 when using Monaghan-Kajtar boundary model
+boundary_layers = 3
+spacing_ratio = 1
+
+# ==========================================================================================
+# ==== Experiment Setup
+tspan = (0.0, 2.0)
+
+# Boundary geometry and initial fluid particle positions
+tank_size = (1.0, 0.5)
+initial_fluid_size = tank_size
+initial_velocity = (0.0, 0.0)
+
+fluid_density = 1000.0
+sound_speed = 10.0
+state_equation = StateEquationCole(; sound_speed, reference_density=fluid_density,
+                                   exponent=7, background_pressure=10_000.0)
+
+tank = RectangularTank(fluid_particle_spacing, initial_fluid_size, tank_size, fluid_density,
+                       n_layers=boundary_layers, spacing_ratio=spacing_ratio,
+                       faces=(false, false, true, true), velocity=initial_velocity)
+fluid = setdiff(tank.fluid, solid)
+
+# ==========================================================================================
+# ==== Solid
+# The kernel in the reference uses a differently scaled smoothing length,
+# so this is equivalent to the smoothing length of `sqrt(2) * particle_spacing` used in the paper.
+smoothing_length = 2 * sqrt(2) * particle_spacing
+smoothing_kernel = WendlandC2Kernel{2}()
+
+boundary_density_calculator = AdamiPressureExtrapolation()
+viscosity_wall = nothing
+# Activate to switch to no-slip walls
[…]
