### 2024-10-02 044cf15
Add first fin simulation
