### 2024-10-02 8441ac4
Add first fin simulation
