### 2024-10-02 ccefcb0
Add first fin simulation
