### 2024-10-04 e22702d
add short tilde description in docs
