+++ docs/src/refs.bib
@@ -44,7 +44,7 @@
 
 @Article{Band2018,
   author    = {Band, Stefan and Gissler, Christoph and Peer, Andreas and Teschner, Matthias},
-  journal   = {Computers and; Graphics},
+  journal   = {Computers \& Graphics},
   title     = {MLS pressure boundaries for divergence-free and viscous SPH fluids},
   year      = {2018},
   issn      = {0097-8493},
@@ -115,15 +115,16 @@
 }
 
 
-@Misc{Bicanic2004,
-  author    = {Bićanić, Nenad},
-  month     = aug,
+@InCollection{Bicanic2004,
+  author    = {Bi{\'c}ani{\'c}, Nenad},
+  booktitle = {Discrete Element Methods},
+  publisher = {Wiley},
   title     = {Discrete Element Methods},
   year      = {2004},
-  doi       = {10.1002/0470091355.ecm006.pub2},
   isbn      = {9780470091357},
+  month     = aug,
+  doi       = {10.1002/0470091355.ecm006.pub2},
   journal   = {Encyclopedia of Computational Mechanics},
-  publisher = {Wiley},
 }
 
 
@@ -159,7 +160,7 @@
 
 @Article{Ramachandran2019,
   author    = {Ramachandran, Prabhu and Puri, Kunal},
-  journal   = {Computers and; Fluids},
+  journal   = {Computers \& Fluids},
   title     = {Entropically damped artificial compressibility for SPH},
   year      = {2019},
   issn      = {0045-7930},
@@ -200,7 +201,7 @@
 }
 
 @Article{Belytschko2000,
-  author   = {Belytschko, Ted and Guo, Yong and Kam Liu, Wing and Ping Xiao, Shao},
+  author   = {Belytschko, Ted and Guo, Yong and Liu, Wing Kam and Xiao, Shao Ping},
   journal  = {International Journal for Numerical Methods in Engineering},
   title    = {A unified stability analysis of meshless particle methods},
   year     = {2000},
@@ -444,15 +445,16 @@
   publisher = {Springer Science and Business Media LLC},
 }
 
-@Article{Mueller2003,
-  author  = {Müller, Matthias and Charypar, David and Gross, Markus},
-  title   = {Particle-Based Fluid Simulation for Interactive Applications},
-  year    = {2003},
-  month   = {07},
-  pages   = {154-159},
-  volume  = {2003},
-  isbn    = {1581136595},
-  journal = {Fluid Dynamics},
+@InProceedings{Mueller2003,
+  author    = {M{\"u}ller, Matthias and Charypar, David and Gross, Markus},
+  booktitle = {Proceedings of the 2003 ACM SIGGRAPH/Eurographics Symposium on Computer Animation},
+  title     = {Particle-Based Fluid Simulation for Interactive Applications},
+  year      = {2003},
+  month     = {07},
+  pages     = {154-159},
+  publisher = {Eurographics Association},
+  doi       = {10.2312/SCA/SCA03/154-159},
+  isbn      = {1581136595},
 }
 
 
@@ -502,7 +504,7 @@
 
 @Article{Fourtakas2019,
   author    = {Fourtakas, Georgios and Dominguez, Jose M. and Vacondio, Renato and Rogers, Benedict D.},
-  journal   = {Computers and; Fluids},
+  journal   = {Computers \& Fluids},
   title     = {Local uniform stencil (LUST) boundary condition for arbitrary 3-D boundaries in parallel smoothed particle hydrodynamics (SPH) models},
   year      = {2019},
   issn      = {0045-7930},
@@ -531,7 +533,7 @@
 
 @Article{Ferrari2009,
[…]
