+++ docs/src/refs.bib
@@ -13,20 +13,6 @@
 }
 
 
-@Article{Valizadeh2015,
-  author    = {Valizadeh, Alireza and Monaghan, Joseph J.},
-  journal   = {Journal of Computational Physics},
-  title     = {A study of solid wall models for weakly compressible SPH},
-  year      = {2015},
-  issn      = {0021-9991},
-  month     = nov,
-  pages     = {5--19},
-  volume    = {300},
-  doi       = {10.1016/j.jcp.2015.07.033},
-  publisher = {Elsevier BV},
-}
-
-
 @Article{Akinci2012,
   author    = {Akinci, Nadir and Ihmsen, Markus and Akinci, Gizem and Solenthaler, Barbara and Teschner, Matthias},
   journal   = {ACM Transactions on Graphics},
@@ -42,218 +28,32 @@
 }
 
 
-@Article{Band2018,
-  author    = {Band, Stefan and Gissler, Christoph and Peer, Andreas and Teschner, Matthias},
-  journal   = {Computers \& Graphics},
-  title     = {MLS pressure boundaries for divergence-free and viscous SPH fluids},
-  year      = {2018},
-  issn      = {0097-8493},
-  month     = nov,
-  pages     = {37--46},
-  volume    = {76},
-  doi       = {10.1016/j.cag.2018.08.001},
-  publisher = {Elsevier BV},
-}
-
-
-@Article{Monaghan2009,
-  author    = {Monaghan, J.J. and Kajtar, J.B.},
-  journal   = {Computer Physics Communications},
-  title     = {SPH particle boundary forces for arbitrary boundaries},
-  year      = {2009},
-  issn      = {0010-4655},
-  month     = oct,
-  number    = {10},
-  pages     = {1811--1820},
-  volume    = {180},
-  doi       = {10.1016/j.cpc.2009.05.008},
-  publisher = {Elsevier BV},
-}
-
-
-@Article{Giles1990,
-  author    = {Giles, Michael B.},
-  journal   = {AIAA Journal},
-  title     = {Nonreflecting boundary conditions for Euler equation calculations},
-  year      = {1990},
-  issn      = {1533-385X},
-  month     = dec,
-  number    = {12},
-  pages     = {2050--2058},
-  volume    = {28},
-  doi       = {10.2514/3.10521},
-  publisher = {American Institute of Aeronautics and Astronautics (AIAA)},
-}
-
-
-@Article{Lastiwka2008,
-  author    = {Lastiwka, Martin and Basa, Mihai and Quinlan, Nathan J.},
-  journal   = {International Journal for Numerical Methods in Fluids},
-  title     = {Permeable and non‐reflecting boundary conditions in SPH},
-  year      = {2008},
-  issn      = {1097-0363},
-  month     = dec,
-  number    = {7},
-  pages     = {709--724},
-  volume    = {61},
-  doi       = {10.1002/fld.1971},
-  publisher = {Wiley},
-}
-
-
-@Article{Negi2020,
-  author    = {Negi, Pawan and Ramachandran, Prabhu and Haftu, Asmelash},
-  journal   = {Computer Methods in Applied Mechanics and Engineering},
-  title     = {An improved non-reflecting outlet boundary condition for weakly-compressible SPH},
-  year      = {2020},
-  issn      = {0045-7825},
-  month     = aug,
-  pages     = {113119},
-  volume    = {367},
-  doi       = {10.1016/j.cma.2020.113119},
-  publisher = {Elsevier BV},
-}
-
-
-@InCollection{Bicanic2004,
-  author    = {Bi{\'c}ani{\'c}, Nenad},
-  booktitle = {Discrete Element Methods},
-  publisher = {Wiley},
[…]
