+++ docs/src/refs.bib
@@ -578,6 +578,21 @@
   publisher = {Elsevier BV},
 }
 
+
+@Article{Valizadeh2015,
+  author    = {Valizadeh, Alireza and Monaghan, Joseph J.},
+  journal   = {Journal of Computational Physics},
+  title     = {A study of solid wall models for weakly compressible SPH},
+  year      = {2015},
+  issn      = {0021-9991},
+  month     = nov,
+  pages     = {5--19},
+  volume    = {300},
+  doi       = {10.1016/j.jcp.2015.07.033},
+  publisher = {Elsevier BV},
+}
+
+
 @Article{Wendland1995,
   author    = {Wendland, Holger},
   journal   = {Advances in Computational Mathematics},
