### 2024-10-07 f1a1d48
implement suggestions
