+++ .github/workflows/ci.yml
@@ -77,20 +77,20 @@ jobs:
         uses: julia-actions/julia-runtest@v1
         with:
           annotate: true
-          # Only run coverage in the Ubuntu job with Julia 1.11
+          # Only run coverage in one Job (Ubuntu and latest Julia version)
           coverage: ${{ matrix.os == 'ubuntu-latest' && matrix.version == '1' }}
         env:
           TRIXIPARTICLES_TEST: unit
 
       - name: Process coverage results
-        # Only process coverage in the Ubuntu job with Julia 1.11
+        # Only run coverage in one Job (Ubuntu and latest Julia version)
         if: matrix.os == 'ubuntu-latest' && matrix.version == '1'
         uses: julia-actions/julia-processcoverage@v1
         with:
           directories: src,test
 
       - name: Upload coverage report to Codecov
-        # Only upload coverage in the Ubuntu job with Julia 1.11
+        # Only run coverage in one Job (Ubuntu and latest Julia version)
         if: matrix.os == 'ubuntu-latest' && matrix.version == '1'
         uses: codecov/codecov-action@v4
         with:
