+++ .github/workflows/ci.yml
@@ -35,29 +35,28 @@ concurrency:
 jobs:
   build:
     name: Run Tests (Julia ${{ matrix.version }} - ${{ matrix.os }} - ${{ matrix.arch }})
-    runs-on: ${{ matrix.os }}
+    runs-on: ubuntu-latest
     strategy:
       fail-fast: false
       matrix:
-        include:
-          # Ubuntu jobs with Julia 1.9, 1.10, and 1.11
-          - version: '1.9'
-            os: ubuntu-latest
-            arch: x64
-          - version: '1.10'
-            os: ubuntu-latest
-            arch: x64
-          - version: '1'
-            os: ubuntu-latest
-            arch: x64
-          # Windows job with Julia 1.10
-          - version: '1.10'
-            os: windows-latest
-            arch: x64
-          # macOS ARM job with Julia 1.10
-          - version: '1.10'
-            os: macos-14
-            arch: arm64
+          version:
+            - '1.9'
+            - '1.10'
+            - '1'
+          os:
+            - ubuntu-latest
+          arch:
+            - x64
+          include:
+            # Also run tests on Windows and macOS-ARM, but only with the latest Julia version
+            # TODO: change back to '1' after JuliaLang/julia#55009 has been backported    
+            - version: '1.10'
+              os: windows-latest
+              arch: x64
+            # macOS ARM job with Julia 1.10
+            - version: '1.10'
+              os: macos-14
+              arch: arm64
 
     steps:
       - name: Check out project
