+++ .github/workflows/ci.yml
@@ -48,13 +48,17 @@ jobs:
           arch:
             - x64
           include:
-            # Also run tests on Windows and macOS-ARM, but only with the latest Julia version
-            # TODO: change back to '1' after JuliaLang/julia#55009 has been backported    
+            # Also run tests on Windows and macOS-ARM, but only with the latest Julia version and 1.10
             - version: '1.10'
               os: windows-latest
               arch: x64
-            # macOS ARM job with Julia 1.10
             - version: '1.10'
+              os: macos-14
+              arch: arm64
+            - version: '1'
+              os: windows-latest
+              arch: x64
+            - version: '1'
               os: macos-14
               arch: arm64
 
