+++ test/validation/validation.jl
@@ -54,7 +54,7 @@
             @test isapprox(error_wcsph_P2, 0, atol=eps())
         else
             # 1.9 causes a large difference in the solution
-            # 1.11 requires a performance hotfix which will likely change these results again
+            # TODO 1.11 requires a performance hotfix which will likely change these results again
             @test isapprox(error_edac_P1, 0, atol=4e-9)
             @test isapprox(error_edac_P2, 0, atol=3e-11)
             @test isapprox(error_wcsph_P1, 0, atol=26.3)
