+++ src/schemes/boundary/open_boundary/method_of_characteristics.jl
@@ -53,29 +53,13 @@ function update_final!(system, ::BoundaryModelLastiwka, v, u, v_ode, u_ode, semi
     end
 end
 
-function evaluate_characteristics!(system, v, u, v_ode, u_ode, semi, t)
-    (; cache) = system
-    (; previous_characteristics_inited) = cache
-
-    # Propagate characteristics through the open boundary
-    if !previous_characteristics_inited[1]
-        for iteration in 1:4
-            calc_characteristics!(system, v, u, v_ode, u_ode, semi, t)
-        end
-
-        previous_characteristics_inited[1] = true
-    else
-        calc_characteristics!(system, v, u, v_ode, u_ode, semi, t)
-    end
-end
-
 # ==== Characteristics
 # J1: Associated with convection and entropy and propagates at flow velocity.
 # J2: Propagates downstream to the local flow
 # J3: Propagates upstream to the local flow
-function calc_characteristics!(system, v, u, v_ode, u_ode, semi, t)
+function evaluate_characteristics!(system, v, u, v_ode, u_ode, semi, t)
     (; volume, cache, boundary_zone) = system
-    (; characteristics, previous_characteristics, previous_characteristics_inited) = cache
+    (; characteristics, previous_characteristics) = cache
 
     for particle in eachparticle(system)
         previous_characteristics[1, particle] = characteristics[1, particle]
