+++ src/schemes/boundary/open_boundary/system.jl
@@ -209,9 +209,11 @@ function update_open_boundary_eachstep!(system::OpenBoundarySPHSystem, v_ode, u_
     # Update density, pressure and velocity based on the characteristic variables.
     # See eq. 13-15 in Lastiwka (2009) https://doi.org/10.1002/fld.1971
     @trixi_timeit timer() "update boundary quantities" update_boundary_quantities!(system,
-                                                                          system.boundary_model,
-                                                                          v, u, v_ode,
-                                                                          u_ode, semi, t)
+                                                                                   system.boundary_model,
+                                                                                   v, u,
+                                                                                   v_ode,
+                                                                                   u_ode,
+                                                                                   semi, t)
 
     @trixi_timeit timer() "check domain" check_domain!(system, v, u, v_ode, u_ode, semi)
 
