+++ Project.toml
@@ -35,7 +35,7 @@ Adapt = "3, 4"
 CSV = "0.10"
 DataFrames = "1.6"
 DelimitedFiles = "1"
-DiffEqCallbacks = "2.25, 3, 4"
+DiffEqCallbacks = "2.28, 3, 4"
 FastPow = "0.1"
 FileIO = "1"
 ForwardDiff = "0.10"
