+++ src/schemes/boundary/open_boundary/boundary_zones.jl
@@ -290,7 +290,7 @@ function spanning_vectors(plane_points::NTuple{3}, zone_width)
     edge2 = plane_points[3] - plane_points[1]
 
     # Check if the edges are linearly dependent (to avoid degenerate planes)
-    if isapprox(dot(normalize(edge1), normalize(edge2)), 1.0; atol=1e-7)
+    if isapprox(norm(cross(edge1, edge2)), 0.0; atol=eps())
         throw(ArgumentError("The vectors `AB` and `AC` must not be collinear"))
     end
 
