### 2024-10-11 c3683e9
fix division by zero
