### 2024-10-15 335f209
First working MPI simulation without communication
