### 2024-10-15 9a18f7a
Simulation works on one rank
On two ranks, the ghost particles are outside of the nhs grid.
We need to query the ghost cells from p4est to create a separate grid for the ghost system nhs.
