+++ src/schemes/fluid/transport_velocity.jl
@@ -13,7 +13,7 @@ See [TVF](@ref transport_velocity_formulation) for more details of the method.
     Thus, it is highly recommended to use [`ViscosityAdami`](@ref) as viscosity model,
     since [`ArtificialViscosityMonaghan`](@ref) leads to bad results.
 """
-struct TransportVelocityAdami{T<:Real}
+struct TransportVelocityAdami{T <: Real}
     background_pressure::T
 end
 
