### 2024-10-16 8d58435
implement CFL condition
