### 2024-10-16 248b780
Running for some steps on two ranks
Then,  rank 0 exits the integration with NaNs, while rank 1 continues and blocks.
