+++ examples/fluid/sphere_surface_tension_3d.jl
@@ -5,7 +5,7 @@ using OrdinaryDiffEq
 
 fluid_density = 1000.0
 
-particle_spacing = 0.1
+particle_spacing = 0.15
 fluid_size = (0.9, 0.9, 0.9)
 
 sound_speed = 20.0
@@ -13,13 +13,12 @@ sound_speed = 20.0
 # For all surface tension simulations, we need a compact support of `2 * particle_spacing`
 smoothing_length = 1.0 * particle_spacing
 
-nu = 0.01
+nu = 0.3
 
 trixi_include(@__MODULE__,
               joinpath(examples_dir(), "fluid", "sphere_surface_tension_2d.jl"),
-              dt=0.1, cfl=1.2, surface_tension_coefficient=0.1,
-              tspan=(0.0, 20.0), nu=nu,
-              alpha=10 * nu / (smoothing_length * sound_speed),
-              fluid_smoothing_kernel=SchoenbergCubicSplineKernel{3}(),
+              surface_tension_coefficient=0.15,
+              tspan=(0.0, 100.0), nu=nu, smoothing_length = 3.0 * particle_spacing,
+              fluid_smoothing_kernel=WendlandC2Kernel{3}(),
               particle_spacing=particle_spacing, sound_speed=sound_speed,
               fluid_density=fluid_density, fluid_size=fluid_size)
