+++ examples/fluid/sphere_surface_tension_3d.jl
@@ -13,11 +13,11 @@ sound_speed = 20.0
 # For all surface tension simulations, we need a compact support of `2 * particle_spacing`
 smoothing_length = 1.0 * particle_spacing
 
-nu = 0.3
+nu = 0.04
 
 trixi_include(@__MODULE__,
               joinpath(examples_dir(), "fluid", "sphere_surface_tension_2d.jl"),
-              surface_tension_coefficient=0.15,
+              surface_tension_coefficient=0.5, dt=0.25,
               tspan=(0.0, 100.0), nu=nu, smoothing_length = 3.0 * particle_spacing,
               fluid_smoothing_kernel=WendlandC2Kernel{3}(),
               particle_spacing=particle_spacing, sound_speed=sound_speed,
