### 2024-10-18 7b8d123
CompatHelper: bump compat for GPUArraysCore to 0.2, (keep existing compat)
