+++ Project.toml
@@ -39,7 +39,7 @@ DiffEqCallbacks = "2.25, 3"
 FastPow = "0.1"
 FileIO = "1"
 ForwardDiff = "0.10"
-GPUArraysCore = "0.1"
+GPUArraysCore = "0.1, 0.2"
 JSON = "0.21"
 KernelAbstractions = "0.9"
 MuladdMacro = "0.2"
