### 2024-10-18 afcdb85
add boundary compress factor
