### 2024-10-22 6728f74
Fix plotting of `ODESolution`
