+++ Project.toml
@@ -35,7 +35,7 @@ Adapt = "3, 4"
 CSV = "0.10"
 DataFrames = "1.6"
 DelimitedFiles = "1"
-DiffEqCallbacks = "2.25, 3"
+DiffEqCallbacks = "3"
 FastPow = "0.1"
 FileIO = "1"
 ForwardDiff = "0.10"
@@ -47,7 +47,7 @@ PointNeighbors = "0.4.2"
 Polyester = "0.7.10"
 RecipesBase = "1"
 Reexport = "1"
-SciMLBase = "1, 2"
+SciMLBase = "2.45"
 StaticArrays = "1"
 StrideArrays = "0.1"
 TimerOutputs = "0.5.25"
