### 2024-10-22 2f5d84f
Allow version 2 of DiffEqCallbacks.jl to work with Julia 1.9
