### 2024-10-22 4002272
fix signed distance field
