+++ examples/preprocessing/complex_shape_2d.jl
@@ -1,12 +1,19 @@
 using TrixiParticles
 using Plots
 
-particle_spacing = 0.05
+particle_spacing = 0.08
 
-filename = "inverted_open_curve"
+filename = "box"
 file = joinpath("examples", "preprocessing", "data", filename * ".asc")
 
 geometry = load_geometry(file)
+
+filename_diff = "circle"
+file_diff = joinpath("examples", "preprocessing", "data", filename_diff * ".asc")
+
+geometry_diff = load_geometry(file_diff)
+
+setdiff!(geometry, geometry_diff)
 
 trixi2vtk(geometry)
 
@@ -16,8 +23,8 @@ point_in_geometry_algorithm = WindingNumberJacobson(; geometry,
 
 # Returns `InitialCondition`
 shape_sampled = ComplexShape(geometry; particle_spacing, density=1.0,
-                             store_winding_number=true,
-                             point_in_geometry_algorithm)
+                             grid_offset=0.1particle_spacing,
+                             store_winding_number=true, point_in_geometry_algorithm)
 
 trixi2vtk(shape_sampled.initial_condition)
 
