+++ src/schemes/fluid/weakly_compressible_sph/system.jl
@@ -266,9 +266,17 @@ function update_pressure!(system::WeaklyCompressibleSPHSystem, v, u, v_ode, u_od
                             density_calculator)
     compute_pressure!(system, v)
     compute_surface_normal!(system, surface_normal_method, v, u, v_ode, u_ode, semi, t)
-    compute_curvature!(system, surface_tension, v, u, v_ode, u_ode, semi, t)
-
+    compute_surface_delta_function!(system, surface_tension)
     return system
+end
+
+function update_final!(system::WeaklyCompressibleSPHSystem, v, u, v_ode, u_ode, semi, t;
+                       update_from_callback=false)
+    (; surface_tension) = system
+
+    # Surface normal of neighbor and boundary needs to have been calculated already
+    compute_curvature!(system, surface_tension, v, u, v_ode, u_ode, semi, t)
+    compute_stress_tensors!(system, surface_tension, v, u, v_ode, u_ode, semi, t)
 end
 
 function kernel_correct_density!(system::WeaklyCompressibleSPHSystem, v, u, v_ode, u_ode,
