+++ src/schemes/fluid/entropically_damped_sph/rhs.jl
@@ -41,7 +41,7 @@ function interact!(dv, v_particle_system, u_particle_system,
         dv_surface_tension = surface_tension_force(surface_tension_a, surface_tension_b,
                                                    particle_system, neighbor_system,
                                                    particle, neighbor, pos_diff, distance,
-                                                   rho_a, rho_b)
+                                                   rho_a, rho_b, grad_kernel)
 
         dv_adhesion = adhesion_force(surface_tension_a, particle_system, neighbor_system,
                                      particle, neighbor, pos_diff, distance)
