### 2024-10-23 0589a04
implement suggestions
