+++ test/systems/edac_system.jl
@@ -227,7 +227,7 @@
 
         @test all(i -> system.cache.neighbor_counter[i] == nparticles(system),
                   nparticles(system))
-        @test all(i -> system.cache.pressure_average[i] ≈ -50.968532955185964,
+        @test all(isapprox(i -> system.cache.pressure_average[i], -50.968532955185964),
                   nparticles(system))
     end
 end
