### 2024-10-23 f20e19e
add set operations for geometries
