### 2024-10-31 550b95b
Fix the URL's in the docs (#637)
* Add template directory and copy the replaced markdown files to the tutorials directory.

* Format code

* Replace strings of file paths with joinpath()

---------

Co-authored-by: Erik Faulhaber <44124897+efaulhaber@users.noreply.github.com>
