### 2024-11-08 trixi-framework/TrixiParticles.jl#654 @luraess
Images do not render in deployed docs
Issue related to revision https://github.com/openjournals/joss-reviews/issues/7044

The images linked in the visu part of the docs:

https://github.com/trixi-framework/TrixiParticles.jl/blob/550b95b5c088841b3bda2232f6e1b9310bae46bd/docs/src/visualization.md?plain=1#L19

and

https://github.com/trixi-framework/TrixiParticles.jl/blob/550b95b5c088841b3bda2232f6e1b9310bae46bd/docs/src/visualization.md?plain=1#L23

does not render in the deployed docs.
