### 2024-11-08 #7044 @luraess
I checked to paper draft, and all looks mostly good to me.

The following minor points would require further attention before the draft could be published:
1. I spotted an issue with image or assets rendering in the deployed docs. I opened an issue so you could get this fixed. Please x-check if there would be other places where this issue occurs.
2. I do not see any performance related comparions, although you make some claims about performance of the software. If any, these would be very valuable to show in order to support your statements.
3. The paper mentions GPU computing, but I can't find anything related in the docs nor hints in how GPU support could be enabled. Here as well, clarification would be welcome and if any possibility to use GPUs, that would be fantastic as particle-based methods may perform well on those architectures.
4. I suspect you are using explicit time integration and thus require weakly compressible formulations. It would be valuable to state how to deal with incompressible flow, or how to achieve this limit with this software.
5. A minor point I am wondering about is how the method and your software would relate or compare to MPM (material point method), such as https://github.com/LandslideSIM/MaterialPointSolver.jl

That's most of it from my side - great job!
