### 2024-11-13 trixi-framework/TrixiParticles.jl#654 @sloede
@luraess Thanks a lot for reporting this!
