### 2024-11-18 #7044 @LasNikas
Thank you @luraess  for your review. We really appreciate your feedback. Here are our comments:

> 1. I spotted an issue with image or assets rendering in the deployed docs. I opened an issue so you could get this fixed. Please x-check if there would be other places where this issue occurs.

Done in [#658](https://github.com/trixi-framework/TrixiParticles.jl/pull/658)

> 2. I do not see any performance related comparions, although you make some claims about performance of the software. If any, these would be very valuable to show in order to support your statements.

We intend to publish a thorough performance review at a later date, but it is not yet ready to be published.

> 3. The paper mentions GPU computing, but I can't find anything related in the docs nor hints in how GPU support could be enabled. Here as well, clarification would be welcome and if any possibility to use GPUs, that would be fantastic as particle-based methods may perform well on those architectures.

Done in [#660](https://github.com/trixi-framework/TrixiParticles.jl/pull/660)

> 4. I suspect you are using explicit time integration and thus require weakly compressible formulations. It would be valuable to state how to deal with incompressible flow, or how to achieve this limit with this software.

The weakly compressible formulation is one way to model an incompressible fluid.
Incompressible SPH is also using an explicit time integrator, but can use larger time steps by running multiple preconditioner steps in each time step.
Both are different strategies to limit density variation to a certain threshold.
However, in the future we also plan to implement incompressible methods that can be solved implicitly.
We will probably need to restructure our code to support implicit time integration schemes.

> 5. A minor point I am wondering about is how the method and your software would relate or compare to MPM (material point method), such as https://github.com/LandslideSIM/MaterialPointSolver.jl

Our framework could be adapted in the future to also support MPM. So far, we haven't compared SPH to MPM.
MPM like SPH can be a variety of different methods with different applications and a direct comparison is not possible.
The main difference between MPM and our currently implemented particle methods is the use of an overset grid.
