### 2024-12-06 #7044 @luraess
> Thank you again @luraess and @giordano for the review. Can you give us a quick update?

Looks all good to go from my side. Happy to briefly skim through the latest proof if you generate it.
