### 2024-12-06 #7044 @danielskatz
`@editorialbot` commands need to be the first thing in a new comment
