### 2024-12-10 #7044 @giordano
I finally finished my review, thanks for the patience!  I [opened a few PRs and issues](https://github.com/trixi-framework/TrixiParticles.jl/issues?q=author%3Agiordano) along the way, most of the issues have been resolved swiftly (although I'd node https://github.com/trixi-framework/TrixiParticles.jl/pull/658 hasn't found its way in a stable release just yet).
