### 2024-12-10 #7044 @sloede
> I finally finished my review, thanks for the patience! I [opened a few PRs and issues](https://github.com/trixi-framework/TrixiParticles.jl/issues?q=author%3Agiordano) along the way, most of the issues have been resolved swiftly (although I'd node [trixi-framework/TrixiParticles.jl#658](https://github.com/trixi-framework/TrixiParticles.jl/pull/658) hasn't found its way in a stable release just yet).

Thanks a lot for your review and the helpful contributions! The referenced PR has already been merged and will become part of the release we will create once the review process here has been completed.
