### 2025-01-23 #7044 @HaoZeke
@LasNikas could you ping me after updating to the final version and also generating the DOI as mentioned here?

https://github.com/openjournals/joss-reviews/issues/7044#issuecomment-2609306845

Apologies again for the delay.
