### 2025-01-24 #7044 @LasNikas
Thank you @HaoZeke for taking the editing.
The final version is v0.2.5 and the DOI is [10.5281/zenodo.14680972](https://doi.org/10.5281/zenodo.14680972)
We had a fix in the affiliation, so I'm re-generating the pdf.
I have checked the authors, affiliations and title. Everything is fine now.
