### 2025-01-26 #7044 @danielskatz
👋 @LasNikas - as track editor, I've proofread the paper and suggested some changes in https://github.com/trixi-framework/TrixiParticles.jl/pull/708 - please merge this, or let me know what you disagree with, then we can continue to publication
