### 2025-01-27 #7044 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
